% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc1stBioDem.R
\name{calc1stBioDem}
\alias{calc1stBioDem}
\title{calc1stBioDem}
\usage{
calc1stBioDem(subtype = "all")
}
\arguments{
\item{subtype}{all per default. ethanol_oils for selecting 1st gen crop types relevant for REMIND input.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates projections of first generation biofuels demand,including biogas, bioethamol and biodiesel, from IEA database. 
The unit is Petajoule.
}
\examples{

\dontrun{ 
calcOutput("1stBioDem")
}
}
\seealso{
\code{\link{calc1stBioenergyPast}}
}
\author{
Xiaoxi Wang, David Klein
}
