% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc1stBioenergyPast.R
\name{calc1stBioenergyPast}
\alias{calc1stBioenergyPast}
\title{calc1stBioenergyPast}
\usage{
calc1stBioenergyPast()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates first generation biofuels production, imports, exports for biogas, bioethanol and biodiesel from IEA database. 
The unit is Petajoule.
}
\examples{

\dontrun{ 
calcOutput("1stBioenergyPast")
}
}
\seealso{
\code{\link{setConfig}}
}
\author{
Xiaoxi Wang, Isabelle Weindl
}
