% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAfforestCosts.R
\name{calcAfforestCosts}
\alias{calcAfforestCosts}
\title{Aggregation and calculation of the mean of each MAgPIE region for the source
SathayeForest}
\usage{
calcAfforestCosts()
}
\value{
MAgPIE object of the calculated means of each MAgPIE region
}
\description{
This function aggregates the data from source SathayeForest. A weight is
implemented as the mean for each MAgPIE region is calculated.
}
\examples{

\dontrun{ 

a <- calcOutput("AfforestCosts")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSathayeForest}},
\code{\link{convertSathayeForest}}
}
\author{
Nele Steinmetz
}
