% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgProductionValue.R
\name{calcAgProductionValue}
\alias{calcAgProductionValue}
\title{calcAgProductionValue}
\usage{
calcAgProductionValue(datasource = "FAO")
}
\arguments{
\item{datasource}{Currently available: \code{"FAO"}}
}
\value{
FAO Value Of Production as a list of MAgPIE objects
}
\description{
Calculate FAO Value Of Production
}
\examples{

\dontrun{ 
a <- calcOutput("AgProductionValue", datasource="FAO")
}
 
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{readSource}}
}
\author{
Roman Popov, Mishko Stevanovic
}
