% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAirPollBaseyearEmi.R
\name{calcAirPollBaseyearEmi}
\alias{calcAirPollBaseyearEmi}
\title{Calculate air pollutant emissions for a reference year
at different sectoral aggregation levels
based on CEDS2025 and GAINS2025 data.}
\usage{
calcAirPollBaseyearEmi(
  baseyear = 2020,
  CEDS.5yearmean = TRUE,
  source = "CEDS2025",
  outsectors = "GAINS"
)
}
\arguments{
\item{baseyear}{base year for which emissions are calculated}

\item{CEDS.5yearmean}{computes 5-year average around base year for CEDS2025
("TRUE"/"FALSE")}

\item{source}{"CEDS2025" or "GAINS2025"}

\item{outsectors}{total ("TOT"),
62 CEDS sectors ("CEDS"),
16 intermediary sectors used to link CEDS and GAINS ("INT"),
35 GAINS sectors ("GAINS"),
13 CMIP7 Harmonization sectors ("CMIP7")}
}
\value{
magclass object
}
\description{
Calculate air pollutant emissions for a reference year
at different sectoral aggregation levels
based on CEDS2025 and GAINS2025 data.
}
\author{
Gabriel Abrahao, Laurin Koehler-Schindler
}
