% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAirPollEmiRef.R
\name{calcAirPollEmiRef}
\alias{calcAirPollEmiRef}
\title{Calculate air pollutant emissions for a reference year, for use
in combination with GAINS data at different sectoral aggregations}
\usage{
calcAirPollEmiRef(
  baseyear = 2020,
  outunits = "Mt/yr",
  namesformat = "GAINS2025"
)
}
\arguments{
\item{baseyear}{year to take as a reference from CEDS, ignored for the EDGAR2005 LUC CO2 emissions}

\item{outunits}{Units of the output, either "Mt/yr" or "kt/yr". Default is "Mt/yr"}

\item{namesformat}{Format of the output pollutant names, either
"GAINS2025" (default, for original GAINS names), "REMIND" (for REMIND internal names)
or "REMINDexo" (for names as used in REMIND's calcEXOGAINS.R)}
}
\value{
magclass object
}
\description{
Calculate air pollutant emissions for a reference year, for use
in combination with GAINS data at different sectoral aggregations
}
\author{
Gabriel Abrahao
}
