% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAnimalStocks.R
\name{calcAnimalStocks}
\alias{calcAnimalStocks}
\title{calcAnimalStocks}
\usage{
calcAnimalStocks(grouping = "IPCC")
}
\arguments{
\item{grouping}{IPCC: Animal grouping of IPCC Guidelines}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates stocks of animals of different categories.
}
\examples{

\dontrun{ 
calcOutput("AnimalStocks")
}
}
\seealso{
\code{\link{calcExcretionIPCC}},
\code{\link{readIPCC}}
}
\author{
Benjamin Leon Bodirsky
}
