% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaEquippedForIrrigation.R
\name{calcAreaEquippedForIrrigation}
\alias{calcAreaEquippedForIrrigation}
\title{calcAreaEquippedForIrrigation}
\usage{
calcAreaEquippedForIrrigation(cellular = FALSE)
}
\arguments{
\item{cellular}{if true, dataset is returned on 0.5 degree resolution}
}
\value{
List of magpie objects with results on country/cellular level, weight on country level, unit and description.
}
\description{
Calculates the area equipped for irrigation based on LU2v1 dataset. It assumes, that all cropland irrigated in the last 20 years at least once is equipped for irrigation.
}
\examples{

\dontrun{ 
calcOutput("LUH2v2")
}
}
\seealso{
\code{\link{calcLanduseInitialisation}}
}
\author{
Benjamin Leon Bodirsky
}
