% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAtmosphericDeposition.R
\name{calcAtmosphericDeposition}
\alias{calcAtmosphericDeposition}
\title{calcAtmosphericDeposition}
\usage{
calcAtmosphericDeposition(datasource = "ACCMIP",
  glo_incl_oceans = FALSE, cellular = FALSE, emission = FALSE,
  scenario = NULL)
}
\arguments{
\item{datasource}{deposition inventory}

\item{glo_incl_oceans}{provides global values that include oceans, as ocenas are not part of the country mapping}

\item{cellular}{cellular or country level emissions}

\item{emission}{if TRUE, not the depositio but the cellular emissions are reported}

\item{scenario}{if dataset contains several scenarios (e.g. ACCMIP), one scenario can be selected.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Conputes Atmospheric (nitrogen) deposition on different land-use types. It distinguishes ammonia (Nh3) and Nitrogen oxides (NOx) as well
}
\examples{

\dontrun{ 
calcOutput("AtmosphericDeposition")
}

}
\seealso{
\code{\link{calcAtmosphericDepositionRates}},
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
