% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAttributes.R
\name{calcAttributes}
\alias{calcAttributes}
\title{calcAttributes}
\usage{
calcAttributes(subtype = "Products")
}
\arguments{
\item{subtype}{subtype of readProductAttributes function.}
}
\value{
List of magpie objects with results on global level, empty weight, unit and description.
}
\description{
provides attributes of different products
}
\examples{

\dontrun{ 
calcOutput("Attributes")
}

}
\seealso{
\code{\link{readProductAttributes}}
}
\author{
Benjamin Leon Bodirsky
}
