% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAWMSconfShr.R
\name{calcAWMSconfShr}
\alias{calcAWMSconfShr}
\title{calcAWMSconfShr}
\usage{
calcAWMSconfShr()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the share of manure managed in different animal waste management systems in confinements. Starting with IPCC 2005 values, turning into scenarios for the future.
}
\examples{

\dontrun{ 
calcOutput("AWMSconfShr")
}
}
\seealso{
\code{\link{calcAWMSconfShrPast}}
}
\author{
Benjamin Leon Bodirsky
}
