% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAWMSconfShrPast.R
\name{calcAWMSconfShrPast}
\alias{calcAWMSconfShrPast}
\title{calcAWMSconfShrPast}
\usage{
calcAWMSconfShrPast(products = "magpie")
}
\arguments{
\item{products}{IPCC: IPCC products. MAgPIE: Magpie products}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates the share of manure excreted in different types of animal waste management systems in  confinements in the year 2005 using the IPCC Guidelines excretion rates.
}
\examples{

\dontrun{ 
calcOutput("AWMSconfShrPast")
}
}
\seealso{
\code{\link{calcAWMSconfShr}},
\code{\link{calcExcretionIPCC}}
}
\author{
Benjamin Leon Bodirsky
}
