% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBioenergyCombined.R
\name{calcBioenergyCombined}
\alias{calcBioenergyCombined}
\title{calcBioenergyCombined}
\usage{
calcBioenergyCombined()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates projections of combined Bioenergy demand adding first and second generation demands based on data from 
Lotze Campen (2014) and from REMIND-MAgPIE-coupling. The Unit is Petajoule
}
\examples{

\dontrun{ 
calcOutput("BioenergyCombined")
}
}
\seealso{
\code{\link{calc1stBioDem}}
}
\author{
Ewerton Araujo
}
