% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiogasProdMSW.R
\name{calcBiogasProdMSW}
\alias{calcBiogasProdMSW}
\title{calcBiogasProdMSW}
\usage{
calcBiogasProdMSW()
}
\value{
Magpie object of CH4 captured
}
\description{
Mt CH4 captured from landfills and dumps
}
\examples{

\dontrun{ a <- calcOutput(type="BiogasProdMSW")
}
}
\author{
David Chen
}
