% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiomassPrices.R
\name{calcBiomassPrices}
\alias{calcBiomassPrices}
\title{read biomass supply curves from Magpie emulator}
\usage{
calcBiomassPrices()
}
\value{
Magpie object with two parameters determining linear biomass supply curve
}
\description{
read biomass supply curves from Magpie emulator
}
