% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBMI.R
\name{calcBMI}
\alias{calcBMI}
\title{calcBMIshr}
\usage{
calcBMI()
}
\value{
List with a magpie object
}
\description{
estimates average BMI of a BMI group for a population group
}
\examples{

\dontrun{ 
calcOutput("BMI",aggregate=FALSE)
}


}
\seealso{
\code{\link{readNCDrisc}},
\code{\link{calcIntake}},
\code{\link{readKuczmarski}}
}
\author{
Benjamin Leon Bodirsky
}
