% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBMIshr.R
\name{calcBMIshr}
\alias{calcBMIshr}
\title{calcBMIshr}
\usage{
calcBMIshr(convert = TRUE)
}
\arguments{
\item{convert}{Use raw data or interpolated data. Raw data should only be used for regressions.}
}
\value{
List with a magpie object
}
\description{
estimates population based on BMI share
}
\examples{

\dontrun{ 
calcOutput("BMIshr")
}


}
\seealso{
\code{\link{readNCDrisc}},
\code{\link{calcIntake}},
\code{\link{readKuczmarski}}
}
\author{
Benjamin Leon Bodirsky
}
