% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBodyHeight.R
\name{calcBodyHeight}
\alias{calcBodyHeight}
\title{calcBodyHeight}
\usage{
calcBodyHeight(convert = TRUE)
}
\arguments{
\item{convert}{if TRUE, the convertscript of Lutz et al is activated.
Also,the year 1965 is extrapolatedusing the worldbank population 
data and sex, age, and education structure of 1970.}
}
\description{
reads in the Lutz et al dataset. 
Aggregates the age structure. Population is divided 
by sex male (M) , female (F) and both (B)
and divided by 8 age classes: 
0-4, 5-9, 10-14, 15-19, AG1 (20-29), AG2 (30-59), AG3(60-79), AG4(80+)
}
