% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCalibrationPrices.R
\name{calcCalibrationPrices}
\alias{calcCalibrationPrices}
\title{Prices of CES variables for calibration runs in REMIND}
\usage{
calcCalibrationPrices()
}
\description{
Uses former REMIND runs
}
\author{
Antoine Levesque
}
