% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacity.R
\name{calcCapacity}
\alias{calcCapacity}
\title{calc Capacity}
\usage{
calcCapacity(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "capacityByTech" or "capacityByPE"}
}
\value{
magpie object of  capacity data
}
\description{
provides historical capacity values in TW
}
\examples{

\dontrun{
calcOutput("Capacity",subtype="capacityByTech")
}
}
\author{
Renato Rodrigues, Stephen Bi
}
