% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacityBounds.R
\name{calcCapacityBounds}
\alias{calcCapacityBounds}
\title{calc capacity bounds}
\usage{
calcCapacityBounds(subtype)
}
\arguments{
\item{subtype}{either \code{historical} for data until 2024 or \code{pipeline}
for projections in 2020, 2025 and 2030 (including some redistribution on EU/NEU level)}
}
\description{
This function gathers project pipeline data for different technologies to generate REMIND input data
for setting historical and near-term bounds. The output is technology capacity in three different categories:
"operational": installed capacity that is operational in the respective year
"construction": capacity that is currently under construction but expected to be completed in the respective year
"planned": capacity that is planned to be completed by this year but not yet under construction. The exact definition of this category
(e.g. whether only projects with FID are included)may vary by technology and is documented in calcProjectPipelines.
Moreover, the function provides project pipeline data for the historical mif file.
}
\author{
Felix Schreyer
}
