% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacityEV.R
\name{calcCapacityEV}
\alias{calcCapacityEV}
\title{CapacityEV
calculate capacity of EV}
\usage{
calcCapacityEV()
}
\value{
magpie object
}
\description{
CapacityEV
calculate capacity of EV
}
\examples{

\dontrun{ a <- calcOutput(type="CapacityEV")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Lavinia Baumstark
}
