% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacityFactor.R
\name{calcCapacityFactor}
\alias{calcCapacityFactor}
\title{calc Capacity Factor}
\usage{
calcCapacityFactor()
}
\value{
magpie object of the capacity factor data
}
\description{
provides capacity factor values
}
\examples{

\dontrun{ 
calcOutput("CapacityFactor")
}

}
\author{
Renato Rodrigues, Stephen Bi
}
