% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacityFactorHist.R
\name{calcCapacityFactorHist}
\alias{calcCapacityFactorHist}
\title{calc Capacity Factor}
\usage{
calcCapacityFactorHist()
}
\value{
magpie object of the capacity factor data
}
\description{
provides capacity factor values
}
\examples{

\dontrun{ 
calcOutput("CapacityFactor")
}

}
\author{
Renato Rodrigues
}
