% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacityNuclear.R
\name{calcCapacityNuclear}
\alias{calcCapacityNuclear}
\title{Historical nuclear capacities and near-term capacity addition bounds for REMIND}
\usage{
calcCapacityNuclear()
}
\description{
use historical nuclear electricity generation capacity and
calculate near-term estimates based on current nuclear power project status
per country.
}
\author{
Robert Pietzcker, Christoph Bertram, Aman Malik, Pascal Weigmann
}
