% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapital.R
\name{calcCapital}
\alias{calcCapital}
\title{Calculate macroeconomic capital stock}
\usage{
calcCapital()
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Compute macroeconomic capital stock based on capital intensities from PWT and GDP scenarios from [mrdrivers].
The PWT capital intensities are used up until 2010. After that, the capital intensities converge towards
that of Japan in 2010, at speeds that vary across scenarios. The final capital stocks are the product
of the capital intensities and the gdp scenarios from [mrdrivers].
}
\seealso{
\itemize{
  \item See the vignette \code{vignette("scenarios", "mrdrivers")} for information on the GDP scenarios.
  \item [readPWT()] for information on the PWT version used.
}
}
