% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapTarget.R
\name{calcCapTarget}
\alias{calcCapTarget}
\title{Calculate Capacity Targets}
\usage{
calcCapTarget(sources)
}
\arguments{
\item{sources}{must be "NewClimate"}
}
\description{
Capacity targets (GW) at regional level from the NewClimate NPI policy database.
}
\author{
Aman Malik, Oliver Richters, Rahel Mandaroux, Léa Hayez, Falk Benke
}
