% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonBudget.R
\name{calcCarbonBudget}
\alias{calcCarbonBudget}
\title{calcCarbonBudget}
\usage{
calcCarbonBudget(decomp_fix = FALSE, cellular = FALSE)
}
\arguments{
\item{decomp_fix}{default FALSE (if TRUE )}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Carbon Budgets for Cropland soils on country levels.
}
\examples{

\dontrun{ 
calcOutput("CarbonBudget")
}
}
\author{
Kristine Karstens
}
