% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonLitter.R
\name{calcCarbonLitter}
\alias{calcCarbonLitter}
\title{calcCarbonLitter}
\usage{
calcCarbonLitter(landuse_types = "carbon", cellular = TRUE)
}
\arguments{
\item{landuse_types}{Sets: carbon or ini_new (magpie initialisation)}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Carbon Input from litter for land use types.
}
\examples{

\dontrun{ 
calcOutput("CarbonLitter")
}
}
\author{
Kristine Karstens
}
