% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonManure.R
\name{calcCarbonManure}
\alias{calcCarbonManure}
\title{calcCarbonManure}
\usage{
calcCarbonManure(landuse_types = "carbon", cellular = TRUE)
}
\arguments{
\item{landuse_types}{Sets: carbon or ini_new (magpie initialisation)}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates carbon input from manure for cropland and pasture soils (other land use types set to 0).
}
\examples{

\dontrun{ 
calcOutput("CarbonManure")
}
}
\author{
Kristine Karstens
}
