% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonMineralization.R
\name{calcCarbonMineralization}
\alias{calcCarbonMineralization}
\title{calcCarbonMineralization}
\usage{
calcCarbonMineralization(landuse_types = "carbon", cellular = TRUE)
}
\arguments{
\item{landuse_types}{Sets: carbon or ini_new (magpie initialisation)}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results, weight, unit and description.
}
\description{
Calculates decompostion rates for land use types.
}
\examples{

\dontrun{ 
calcOutput("CarbonMineralization")
}
}
\author{
Kristine Karstens
}
