% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonResidues.R
\name{calcCarbonResidues}
\alias{calcCarbonResidues}
\title{calcCarbonResidues}
\usage{
calcCarbonResidues(landuse_types = "carbon", cellular = TRUE)
}
\arguments{
\item{landuse_types}{Sets: carbon or ini_new (magpie initialisation)}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates carbon input from residues for cropland soils (other land use types set to 0).
}
\examples{

\dontrun{ 
calcOutput("CarbonResidues")
}
}
\author{
Kristine Karstens
}
