% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonSoilIni.R
\name{calcCarbonSoilIni}
\alias{calcCarbonSoilIni}
\title{calcCarbonSoilIni}
\usage{
calcCarbonSoilIni(landuse_types = "carbon", cellular = TRUE)
}
\arguments{
\item{landuse_types}{Sets: carbon or ini_new (magpie initialisation)}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Initialize soil carbon pools for land use types.
}
\examples{

\dontrun{ 
calcOutput("CarbonSoilIni")
}
}
\author{
Kristine Karstens
}
