% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCCSbounds.R
\name{calcCCSbounds}
\alias{calcCCSbounds}
\title{Calculate CCS bound indicator for 2025 and 2030}
\usage{
calcCCSbounds()
}
\description{
Calculate CCS bound indicator for 2025 and 2030
}
\author{
Jessica Strefler, Lavinia Baumstark
}
