% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCCScapacity.R
\name{calcCCScapacity}
\alias{calcCCScapacity}
\title{calc CCS capacity}
\usage{
calcCCScapacity(subtype)
}
\arguments{
\item{subtype}{either \code{historical} for data until 2024 or \code{pipeline}
for projections in 2020, 2025 and 2030 (including some redistribution on EU/NEU level)}
}
\description{
Calculate CCS capacity from IEA CCUS data
}
\author{
Anne Merfort, Falk Benke
}
