% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCement.R
\name{calcCement}
\alias{calcCement}
\title{Calculate Historic Cement Production}
\usage{
calcCement()
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x} with
country-level cement production in tonnes, \code{weight}, \code{unit}, \code{description},
and \code{min} fields.
}
\description{
Combines cement production data from \code{\link[=readvanRuijven2016]{readvanRuijven2016()}} and
\code{\link[=readUSGS]{readUSGS(cement)}} into a single data set, using USGS data from
2005 on.
}
\seealso{
\link{calcOutput}
}
\author{
Michaja Pehl
}
