% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCementShare.R
\name{calcCementShare}
\alias{calcCementShare}
\title{Calculate Cement Share in NONMET FE Use}
\usage{
calcCementShare()
}
\value{
A list with a \code{\link[magclass:magclass]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, \code{min}, and \code{max}.
}
\description{
Estimated shares of cement in \code{NONMET} final energy use based on OECD and
Non-OECD figures from IEA 2017 \href{https://www.zotero.org/groups/52011/rd3/items/X8XCUJ5U}{Energy Technology Perspectives}.  Shares are weighted
by GDP for aggregation and converge towards global values by 2100.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
