% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCentralFeedshares.R
\name{calcCentralFeedshares}
\alias{calcCentralFeedshares}
\title{calcCentralFeedshares}
\usage{
calcCentralFeedshares()
}
\value{
Central feed shares and weights as list of two MAgPIE-objects
}
\description{
Calculates future central feed shares for all livestock categories 
based on the results of a non-linear regression between historical 
central feed shares and livestock productivity and using Koeppen-
Geiger climate zones
}
\examples{
\dontrun{ 
calcOutput("CentralFeedshares")

}
}
\author{
Isabelle Weindl, Benjamin Bodirsky, Stephen Wirth, Jan Philipp Dietrich
}
