% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChemicalFeedstocksShare.R
\name{calcChemicalFeedstocksShare}
\alias{calcChemicalFeedstocksShare}
\title{Calculate Chemical Feedstock share projections}
\usage{
calcChemicalFeedstocksShare()
}
\value{
A list with a \code{\link[magclass:magclass]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, \code{min}, and \code{max}.
}
\description{
Calculates the share of \code{CHEMICAL} in \code{CHEMICAL = NECHEM} and converges it
towards the maximum value of either OECD or non-OECD countries by 2050.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
