% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClinker_to_cement_ratio.R
\name{calcClinker_to_cement_ratio}
\alias{calcClinker_to_cement_ratio}
\title{Calculate Clinker-to-Cement Ratio}
\usage{
calcClinker_to_cement_ratio()
}
\value{
A list with a \code{\link[magclass:magclass]{magpie}} object \code{x}, \code{weight},
\code{unit}, and \code{description}.
}
\description{
Calculate Clinker-to-Cement Ratio
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}, \code{\link[=readADVANCE_WP2]{readADVANCE_WP2()}}, \code{\link[=convertADVANCE_WP2]{convertADVANCE_WP2()}}
}
\author{
Michaja Pehl
}
