% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoalLabourProductivity.R
\name{calcCoalLabourProductivity}
\alias{calcCoalLabourProductivity}
\title{CalcCoalLabourProductivity
calculates historical employment in the coal sector and projects current (2020) and future employment factors (Jobs/PJ)}
\usage{
calcCoalLabourProductivity(subtype)
}
\arguments{
\item{subtype}{Either "Employment_factor" or "Employment"}
}
\description{
CalcCoalLabourProductivity
calculates historical employment in the coal sector and projects current (2020) and future employment factors (Jobs/PJ)
}
\author{
Aman Malik
}
