% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCommoditiesEORA.R
\name{calcCommoditiesEORA}
\alias{calcCommoditiesEORA}
\title{calcCommoditiesEORA}
\usage{
calcCommoditiesEORA(subtype)
}
\arguments{
\item{subtype}{- year (1990-2013)}
}
\value{
a list consisting of a MAgPIE object of the EORA data, weight, unit, description and as note a vector of
the lost value and the removed value
}
\description{
removes the distinction between commodities and industries in EORA
}
\note{
For the countries that report both commodities and industries we want to translate the industries into 
commodities (as the commodities are more detailed). So here is what we do:
\tabular{ccccc}{
\tab other countries \tab Industries \tab Commodities \tab other countries \cr
other countries \tab x \tab A \tab C \tab x \cr
Industries \tab  \tab  \tab B \tab  \cr
Commodities \tab x \tab A \tab C \tab x \cr
other countries \tab x \tab A \tab C \tab x\cr
}
\emph{Here we focus on one country and its industries and commodities.  Values flow 
from the inputsectors, which are the rownames, into the outputsectors, which are the columnnames. 
Where nothing is written no numbers are in the table, in all other places every value is possible. 
The columns/rows "Commodities"/"Industries" stand for all commodities/industries of the country.}
 
We want to be able to remove part A and part B. To achieve this we redistribute the 
value of part B by adding it in a certain way onto part C, which is already sufficient to 
remove A as well as B (see below for reason). We use the informaiton of part A to decide how to 
redistribute the value, as it tells us from which sectors the value in 
the industries origanally came from. So we want to use part A to calculate the shares 
describing how much of each value should be added where. Mathematically this ist a matrix multiplication:
mat1 shall be the industrie columns but whith each entry divided by the sum of the corresponding 
column and mat2 shall be part B. Then mat1*mat2 gives us a matrix with the same rows as mat1 but the commodities 
as columns, which we can add onto the commoditie-columns.

Reason: This is the only step in which we want to reduce the value of the total table. Before this 
calculation, the value which is in part A column appears twice in the complete table, once as
 value flowing into the industries (A) and once as value flowing out of the industries (B). As we remove the step in
 which value flows through industries, after the calculation this value sholud appear only once. 
 Therefore we can remove both parts even though we redistribute only one of them.
 
To consider: We also reduce the total value unintentionally, because due to the rounding in readEORA, some industries 
exist as input but not as output (or vice versa) which means that we also have to remove it 
as input to be able to do the matrix multiplication. 

Both lost values are printed out in the end, in order to be able to compare the original value of the 
table with the value of the result to check if everything went right.
}
\seealso{
\code{\link{calcEORA}}
}
\author{
Debbora Leip
}
