% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConsShare.R
\name{calcConsShare}
\alias{calcConsShare}
\title{Calculate the consumption shares for each of the 10 deciles across different regions}
\usage{
calcConsShare()
}
\value{
magpie objects of regional consumption shares for each of the 10 deciles.
weights = 1, such that the aggregation later takes country average
}
\description{
This function calculates the consumption shares for sub-regional deciles within total consumption
across REMIND regions. It assumes a log-normal distribution for sub-regional consumption.
The calculation proceeds in three steps:
1. Calculate 'sigma' (the variance of the distribution) from regional Theil index.
2. Generate regional distribution with 'sigma', and integrate and normalize to compute
the relative share of each decile group.
3. Assign regional values to individual countries within each region and set the
aggregation rule for calcOutput to calculate the average.
}
\examples{
\dontrun{
a <- calcOutput("ConsShare")
}

}
\seealso{
\code{\link{calcOutput}} \code{\link{calcGini},\link{readGini}.\link{calcTheil}}
}
\author{
Jiarui Zhong
}
