% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoolingSharesAll.R
\name{calcCoolingSharesAll}
\alias{calcCoolingSharesAll}
\title{Calculate Cooling Type Shares}
\usage{
calcCoolingSharesAll()
}
\value{
MAgPIE object on cooling type shares per elecricity technology and
REMIND region
}
\description{
This function merges the output of two other functions that calculate REMIND
input data for the shares of cooling types per electricity technology and
REMIND region, using as initial information the Davies (2013) data per
electricity technology and GCAM region. The two other functions separately
calculate data for the base year and for future time steps. The source data
provide most required information but some assumptions on missing data are
also made.
}
\examples{
\dontrun{

a <- calcOutput("CoolingSharesAll")
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readDaviesCooling}},
\code{\link{convertDaviesCooling}},
\code{\link{calcCoolingSharesBase}},\code{\link{calcCoolingSharesFuture}}
}
\author{
Ioanna Mouratiadou
}
