% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoolingSharesBase.R
\name{calcCoolingSharesBase}
\alias{calcCoolingSharesBase}
\title{Calculate Cooling Type Shares for the Base Year}
\usage{
calcCoolingSharesBase()
}
\value{
MAgPIE object on cooling type shares per elecricity technology and
REMIND region
}
\description{
This function calculates REMIND input data for the shares of cooling types
per electricity technology and REMIND region in 2005, using as initial
information the Davies (2013) data per electricity technology and GCAM
region. The source data provide most required information but some
assumptions on missing data are also made.
}
\examples{
\dontrun{
a <- calcOutput("CoolingSharesBase")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readDaviesCooling}},
\code{\link{convertDaviesCooling}},
\code{\link{calcCoolingSharesAll}},\code{\link{calcCoolingSharesFuture}}
}
\author{
Lavinia Baumstark, Ioanna Mouratiadou
}
