% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoolingSharesFuture.R
\name{calcCoolingSharesFuture}
\alias{calcCoolingSharesFuture}
\title{Calculate Cooling Type Shares for Future Timesteps}
\usage{
calcCoolingSharesFuture()
}
\value{
MAgPIE object on cooling type shares per elecricity technology and
REMIND region
}
\description{
This function calculates REMIND input data for the shares of cooling types
per electricity technology and REMIND region in post-2020, using as initial
information the Davies (2013) data per electricity technology and GCAM
region. The source data provide most required information but some
assumptions on missing data are also made.
}
\examples{

\dontrun{ 
a <- calcOutput("CoolingSharesFuture")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readDaviesCooling}},
\code{\link{convertDaviesCooling}},
\code{\link{calcCoolingSharesAll}},\code{\link{calcCoolingSharesBase}}
}
\author{
Ioanna Mouratiadou
}
