% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostOfCapital.R, R/calcCostOfCapital_b.R
\name{calcCostOfCapital}
\alias{calcCostOfCapital}
\title{Calculate capital costs for energy technologies}
\usage{
calcCostOfCapital()

calcCostOfCapital()
}
\description{
Provides REMIND data for 25_WACC Module

Provides REMIND data for 25_WACC Module
}
\examples{
\dontrun{
calcOutput("ETH_WACC")
}

\dontrun{
calcOutput("ETH_WACC")
}

}
\author{
Diamantis Koutsandreas
}
