% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostsTrade.R
\name{calcCostsTrade}
\alias{calcCostsTrade}
\title{Calculate trade costs}
\usage{
calcCostsTrade()
}
\value{
REMIND data forPE tradecosts (energy losses on import) and
corresonding weights (1) as a list of two MAgPIE objects
}
\description{
Provides REMIND data for PE tradecosts (energy losses on import).
}
\examples{
\dontrun{
calcOutput("CostsTrade")
}

}
\author{
Lavinia Baumstark
}
