% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostsTradePeFinancial.R
\name{calcCostsTradePeFinancial}
\alias{calcCostsTradePeFinancial}
\title{Calculate Trade Cost}
\usage{
calcCostsTradePeFinancial()
}
\description{
Provides REMIND data for PE trade cost (energy losses on import, export and use).
}
\examples{
\dontrun{
calcOutput("CostsTradePeFinancial")
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Regina Brecha, Lavinia Baumstark
}
