% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostsWeathering.R
\name{calcCostsWeathering}
\alias{calcCostsWeathering}
\title{Calculate costs of transport of enhanced weathering}
\usage{
calcCostsWeathering()
}
\value{
transport costs of spreading rock on the fields
}
\description{
Calculate costs of transport of enhanced weathering
}
\examples{
\dontrun{
calcOutput("CostsWeathering")
}

}
