% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCroparea.R
\name{calcCroparea}
\alias{calcCroparea}
\title{calcCroparea}
\usage{
calcCroparea(sectoral = "kcr", physical = TRUE, cellular = FALSE,
  irrigation = FALSE)
}
\arguments{
\item{sectoral}{for landtype = "area_harvested" this aggregates FAO products. "ProductionItem" for unaggregated ProdSTAT items, "FoodBalanceItem" for Food Balance Sheet categories, "kcr" for MAgPIE items and "lpj" for LPJmL items.}

\item{physical}{if TRUE the sum over all crops agrees with the cropland area per country}

\item{cellular}{if TRUE crop area it calculates the cellular MAgPIE crop area for all magpie croptypes. Crop area from LUH2 crop types (c3ann, c4ann, c3per, c4per, cnfx) 
are mapped to MAgpIE crop types using mappingLUH2cropsToMAgPIEcrops.csv. Harvested areas of FAO weight area within a specific LUH crop type to devide into MAgPIE crop types.}

\item{irrigation}{if true, the cellular areas are returned sperated by irrigated and rainfed. More about irrigation setup in calcLUH2v2}
}
\value{
areas of individual crops from FAOSTAT and weight
}
\description{
Returns harvested areas of individual crops from FAOSTAT. Total harvested areas can be lower or higher than arable land because of multicropping or fallow land.
}
\author{
Ulrich Kreidenweis, Kristine Karstens
}
