% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDemFoodHist.R
\name{calcDemFoodHist}
\alias{calcDemFoodHist}
\title{Calculate food demand based on the output of the Demand Model}
\usage{
calcDemFoodHist()
}
\value{
Historical food demand
}
\description{
This function prepares historical food demand values for use. As the source
data already provides all required information this function purely selects
the needed data.
}
\examples{

\dontrun{ 
a <- calcOutput("DemFoodHist")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readDemandModel}},
\code{\link{convertDemandModel}}
}
\author{
Isabelle Weindl
}
