% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiffInvestCosts.R
\name{calcDiffInvestCosts}
\alias{calcDiffInvestCosts}
\title{Aggregated investment cost data for REMIND regions (based on IEA_WEO)}
\usage{
calcDiffInvestCosts(subtype)
}
\arguments{
\item{subtype}{either "Invest_Costs" or "Efficiency"}
}
\value{
Magpie object with aggregated but differentiated investment costs for some technologies.
}
\description{
Disaggregated investment cost data is aggregated and technologies renamed to REMIND names
}
\details{
REMIND does not have a classification of coal power plants e.g., sub-critical.
Therefore, countries are given coal plant costs assuming what type of coal plants are
expected to develop there. For other technologies, certain assumptions are taken
to change to REMIND convention.
}
\author{
Aman Malik
}
