% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDspvEFmult.R
\name{calcDspvEFmult}
\alias{calcDspvEFmult}
\title{Calculates a multiplier for the employment factor for distributed solar pv}
\usage{
calcDspvEFmult()
}
\description{
Calculates a multiplier for the employment factor for distributed solar pv
}
\author{
Aman Malik
}
