% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEarlyRetirementAdjFactor.R
\name{calcEarlyRetirementAdjFactor}
\alias{calcEarlyRetirementAdjFactor}
\title{calc Early Retirement Adjustment Factor}
\usage{
calcEarlyRetirementAdjFactor()
}
\value{
magpie object of additional adjusment percentage to be added to the fraction of the early retired capital in countries to account for relatively old technologies fleet
}
\description{
provides the extra retirement rate to account for relatively old fleet technologies retirement
}
\examples{

\dontrun{ 
calcOutput(type="EarlyRetirementAdjFactor")
}
 
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Renato Rodrigues
}
