% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEconometricEmiParameter.R
\name{calcEconometricEmiParameter}
\alias{calcEconometricEmiParameter}
\title{Calculate baseline emissions of waste}
\usage{
calcEconometricEmiParameter()
}
\value{
REMIND data for CO2 parameters to calculate baseline emissions of
waste from population and investment and corresponding weights (population)
as a list of two MAgPIE objects
}
\description{
Provides REMIND data for CO2 parameters to calculate baseline emissions of
waste from population and investment.
}
\examples{
\dontrun{
calcOutput("calcEconometricEmiParameter")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
