% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEDGETransport.R
\name{calcEDGETransport}
\alias{calcEDGETransport}
\title{Prepare EDGETransport inputs}
\usage{
calcEDGETransport(subtype)
}
\arguments{
\item{subtype}{REMIND/iterative EDGE-T input data subtypes}
}
\value{
REMIND/iterative EDGE-T input data for all scenario combinations
}
\description{
Prepare EDGETransport inputs
}
\examples{
\dontrun{
a <- calcOutput(type = "EDGETransport", subtype = "CAPEXandNonFuelOPEX", aggregate = F)
}
}
\author{
Johanna Hoppe
}
