% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEDGETrData.R
\name{calcEDGETrData}
\alias{calcEDGETrData}
\title{Runs EDGE-T model}
\usage{
calcEDGETrData()
}
\value{
list of EDGEtransport iterative inputs
}
\description{
Runs EDGE-T model
}
\examples{
\dontrun{ a <- calcOutput(type="EDGETrData", aggregate = F)
}

}
\seealso{
\code{\link{readEDGETransport}}, \code{\link[madrat]{calcOutput}}
}
\author{
Marianna Rottoli
}
