% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEEAGHGEmissions.R
\name{calcEEAGHGEmissions}
\alias{calcEEAGHGEmissions}
\title{Calculate historical GHG Emissions from European Environment Agency}
\usage{
calcEEAGHGEmissions()
}
\description{
Calculate historical GHG Emissions from European Environment Agency
}
\author{
Falk Benke, Renato Rodrigues
}
