% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEEAGHGProjections.R
\name{calcEEAGHGProjections}
\alias{calcEEAGHGProjections}
\title{Calculate EEA emission projections from the two projections sources provided by European Environment Agency}
\usage{
calcEEAGHGProjections()
}
\description{
Calculate EEA emission projections from the two projections sources provided by European Environment Agency
}
\author{
Falk Benke, Renato Rodrigues
}
