% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEEZdistribution.R
\name{calcEEZdistribution}
\alias{calcEEZdistribution}
\title{Calculate distribution of total EEZ size}
\usage{
calcEEZdistribution()
}
\description{
Calculate distribution of total EEZ size
}
\author{
Tabea Dorndorf
}
