% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF3confinement.R
\name{calcEF3confinement}
\alias{calcEF3confinement}
\title{calcEF3confinement}
\usage{
calcEF3confinement(products = "magpie",
  selection = "n_pollutants_direct")
}
\arguments{
\item{products}{Either livestock products in MAgPIE or IPCC products}

\item{selection}{defaults to n_pollutants_direct}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Emission factors for nitrogenous emissions in livestock confinements
}
\examples{

\dontrun{ 
calcOutput("EF3confinement")
}
}
\author{
Benjamin Leon Bodirsky
}
