% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEF3prp.R
\name{calcEF3prp}
\alias{calcEF3prp}
\title{calcEF3prp}
\usage{
calcEF3prp(select_years = "y2005")
}
\arguments{
\item{select_years}{if only one year is selected, years is set to NULL}
}
\value{
list of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Returns emission factor for manure excreted during pasture range and paddock. Differs depending on the share of small ruminants.
}
\examples{

\dontrun{ 
calcOutput("EF3prp")
}
}
\author{
Benjamin Leon Bodirsky
}
