% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEFch4AWMS.R
\name{calcEFch4AWMS}
\alias{calcEFch4AWMS}
\title{calcEFch4AWMS}
\usage{
calcEFch4AWMS()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
emission factors for methane from animal waste management, depending on manure managed in confinements. The emission factors were calculated based on FAOSTAT estimates due to lack of all necessary parameters in the IPCC Guidelines
}
\examples{

\dontrun{ 
calcOutput("EFch4AWMS")
}

}
\seealso{
\code{\link{calcEFch4Rice}},
\code{\link{calcOutput}}
}
\author{
Benjamin Leon Bodirsky
}
