% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEFch4Rice.R
\name{calcEFch4Rice}
\alias{calcEFch4Rice}
\title{calcEFch4Rice}
\usage{
calcEFch4Rice(physical = TRUE)
}
\arguments{
\item{physical}{if true physical area, if false area harvested}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
emission factors for methane from flooded rice fields, depending on phyiscal area or area harvested. The emission factors were calculated based on FAOSTAT estimates due to lack of all necessary parameters in the IPCC Guidelines
}
\examples{

\dontrun{ 
calcOutput("EFch4Rice")
}

}
\seealso{
\code{\link{calcEFch4AWMS}},
\code{\link{calcOutput}}
}
\author{
Benjamin Leon Bodirsky
}
