% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffortSharingRefEmi.R
\name{calcEffortSharingRefEmi}
\alias{calcEffortSharingRefEmi}
\title{calc Effort Sharing Reference Emissions}
\usage{
calcEffortSharingRefEmi(subtype)
}
\arguments{
\item{subtype}{type of reference emissions used to define emission reduction target fo European Effort Sharing Decision: EEA_GHG, Eurostat_GHG, REMIND_GHG (deprecated) or REMIND_CO2.}
}
\value{
2005 reference emissions to calculate effort sharing decision targets
}
\description{
provides region specific Effort Sharing Reference Emissions
}
\examples{

\dontrun{
calcOutput("EffortSharingRefEmi",subtype="Eurostat_GHG")
}

}
\author{
Renato Rodrigues
}
