% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEffortSharingTarget.R
\name{calcEffortSharingTarget}
\alias{calcEffortSharingTarget}
\title{calc Effort Sharing Target}
\usage{
calcEffortSharingTarget()
}
\value{
target data magpie object
}
\description{
provides region specific Effort Sharing Emission target
}
\examples{

\dontrun{ 
calcOutput("EffortSharingTarget")
}

}
\author{
Renato Rodrigues
}
