% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEfNSoil.R
\name{calcEfNSoil}
\alias{calcEfNSoil}
\title{calcEmisNitrogenPast}
\usage{
calcEfNSoil(method = "IPCC")
}
\arguments{
\item{method}{If IPCC, using the ipcc emission factors as share of applied N inputs. If Nloss, as share of cropland budget surplus.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Emission factors from cropland soils.
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenPast")
}

}
\author{
Benjamin Leon Bodirsky
}
