% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmber.R
\name{calcEmber}
\alias{calcEmber}
\title{calc Ember}
\usage{
calcEmber(subtype = "all")
}
\arguments{
\item{subtype}{data subtype. Either "capacity", "generation" or "all"}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
prepare the yearly Ember electricity data set
To use only a part of the Ember data, call calcOutput("Ember", subtype = "...")
and convert to TW if you want to use capacities as input data to REMIND.
}
\seealso{
[`calcOutput()`]
}
\author{
Pascal Weigmann
}
