% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiAirPollLandUse.R
\name{calcEmiAirPollLandUse}
\alias{calcEmiAirPollLandUse}
\title{calcEmiAirPoll
calculate Air Pollution Emissions}
\usage{
calcEmiAirPollLandUse()
}
\value{
magpie object
}
\description{
calcEmiAirPoll
calculate Air Pollution Emissions
}
\examples{
\dontrun{
a <- calcOutput(type = "EmiAirPollLandUse")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser
}
