% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiCO2LandUse.R
\name{calcEmiCO2LandUse}
\alias{calcEmiCO2LandUse}
\title{EmiCO2LandUse
calculate co2 emissions from land use change}
\usage{
calcEmiCO2LandUse()
}
\value{
magpie object
}
\description{
EmiCO2LandUse
calculate co2 emissions from land use change
}
\examples{

\dontrun{ a <- calcOutput(type="EmiCO2LandUse")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser
}
