% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiFossilFuelExtr.R
\name{calcEmiFossilFuelExtr}
\alias{calcEmiFossilFuelExtr}
\title{Fugitive methane emissions from fossil fuel extraction}
\usage{
calcEmiFossilFuelExtr(source)
}
\arguments{
\item{source}{either "EDGAR" or "CEDS2025" (after REMIND ScenarioMIP release)}
}
\value{
Magpie object with CH4 emissions from fossil fuel extraction for coal, oil and gas in 2005, in Mt CH4
}
\description{
Fugitive methane emissions from fossil fuel extraction
}
\details{
REMIND uses historical data on fugitive methane emissions from fossil fuel extraction for coal, oil and gas
to derive emission factors. The data is available from two sources: EDGAR (for base year 2005) and CEDS2025 (2020).
}
\author{
Gabriel Abrahao
}
