% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiLULUCFCountryAcc.R
\name{calcEmiLULUCFCountryAcc}
\alias{calcEmiLULUCFCountryAcc}
\title{calcEmiLULUCFCountryAcc}
\usage{
calcEmiLULUCFCountryAcc()
}
\value{
Magpie object with historical LULUCF emissions
}
\description{
Historical LULUCF CO2 emissions data following the national
accounting approach. Where available data reported to the UNFCCC is taken.
If not available, data estimated by Forsell et al. (2025),
https://pure.iiasa.ac.at/id/eprint/20368/1/advpub_24.001.pdf, is taken.
Note that Forsell et al. data reports total GHG LULUCF emissions. However, as
non-CO2 LULULUCF emissions are typically small we neglect the difference and
use them to calculate the offset between book-keeping and national accounting
CO2 emissions in REMIND.
}
\author{
Felix Schreyer, Falk Benke
}
