% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiMac.R
\name{calcEmiMac}
\alias{calcEmiMac}
\title{Calculate baseline emissions for maccs}
\usage{
calcEmiMac()
}
\value{
REMIND data for baseline emissions for maccs and corresonding
weights (NULL) as a list of two MAgPIE objects
}
\description{
Provides REMIND data for baseline emissions for maccs.
}
\examples{

\dontrun{ 
calcOutput("calcEmiMac")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
