% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiMac1990.R
\name{calcEmiMac1990}
\alias{calcEmiMac1990}
\title{Calculate baseline emissions for maccs for 1990}
\usage{
calcEmiMac1990()
}
\value{
REMIND data for baseline emissions for maccs for 1990 and
corresonding weights (NULL) as a list of two MAgPIE objects
}
\description{
Provides REMIND data for baseline emissions for maccs for 1990.
}
\examples{
\dontrun{
calcOutput("EmiMac1990")
}
}
\author{
Lavinia Baumstark
}
