% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiPollutantExo.R
\name{calcEmiPollutantExo}
\alias{calcEmiPollutantExo}
\title{calcEmiPollutantExo
calculate EmiPollutantExo based on RCP data}
\usage{
calcEmiPollutantExo(subtype, aviationshippingsource = "RCP")
}
\arguments{
\item{subtype}{Either 'Waste' or 'AviationShipping'}

\item{aviationshippingsource}{Defines source for aviation and shipping emissions. Either 'RCP' or 'LeeGAINS'.}
}
\value{
magpie object
}
\description{
calcEmiPollutantExo
calculate EmiPollutantExo based on RCP data
}
\examples{

\dontrun{ a <- calcOutput(type="EmiPollutantExo")
}


}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser
}
