% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiReference.R
\name{calcEmiReference}
\alias{calcEmiReference}
\title{calc European Reference Emissions}
\usage{
calcEmiReference()
}
\value{
2030 emission reductions tragets for 40%, 55% and 64% reductions in relation to 2005 values
}
\description{
provides European 2030 emission targets in relation to 1990 and 2005 emissions
}
\examples{
\dontrun{
calcOutput("EmiReference")
}

}
\author{
Falk Benke and Renato Rodrigues
}
