% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmisNitrogenAWMSPast.R
\name{calcEmisNitrogenAWMSPast}
\alias{calcEmisNitrogenAWMSPast}
\title{calcEmisNitrogenAWMSPast}
\usage{
calcEmisNitrogenAWMSPast()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates nitrogenous emissions from animal waste management systems in the historical period
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenAWMSPast")
}

}
\author{
Benjamin Leon Bodirsky
}
