% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmisNitrogenCroplandPast.R
\name{calcEmisNitrogenCroplandPast}
\alias{calcEmisNitrogenCroplandPast}
\title{calcEmisNitrogenCroplandPast}
\usage{
calcEmisNitrogenCroplandPast(method = "IPCC")
}
\arguments{
\item{method}{IPCC: emissions are calculated according the the IPCC 2006 National Guidelines for Greenhouse Gas Inventories. Nsurplus: Emissions in 2005 are calculated according to IPCC, and the scaled with nitrogen losses from croplands.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates nitrogenous emissions from croplands in the historical period.
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenCroplandPast")
}

}
\author{
Benjamin Leon Bodirsky
}
