% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmisNitrogenNonagland.R
\name{calcEmisNitrogenNonaglandPast}
\alias{calcEmisNitrogenNonaglandPast}
\title{calcEmisNitrogenNonaglandPast}
\usage{
calcEmisNitrogenNonaglandPast(method = "Nsurplus")
}
\arguments{
\item{method}{Method for calculating Atmospheric deposition: Nsurplus2 and Nsurplus are based on deposition rates based on own emission calculations after 2 or after 1 iteration, respectively.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates nitrogenous emissions from non-agricultural land for the historical period
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenPasturePast")
}

}
\seealso{
\code{\link{calcEmisNitrogenPast}},
\code{\link{calcExcretion}}
}
\author{
Benjamin Leon Bodirsky
}
