% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmisNitrogenPreagriculture.R
\name{calcEmisNitrogenPreagriculture}
\alias{calcEmisNitrogenPreagriculture}
\title{calcEmisNitrogenPreagriculture}
\usage{
calcEmisNitrogenPreagriculture(cellular = FALSE, deposition = TRUE)
}
\arguments{
\item{cellular}{cellular or country outputs}

\item{deposition}{if TRUE, losses include atmospheric deposition inputs that are lost afterwards. If false, only biological fixation is considered.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates nitrogenous emissions Nitrogen emissions from soils under 100% natural cover (even for crop and urban) assuming a pre-agricultural time.
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenPreagriculture")
}

}
\seealso{
\code{\link{calcEmisNitrogenPast}},
\code{\link{calcExcretion}}
}
\author{
Benjamin Leon Bodirsky
}
