% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmisNitrogenWater.R
\name{calcEmisNitrogenWater}
\alias{calcEmisNitrogenWater}
\title{calcEmisNitrogenWater}
\usage{
calcEmisNitrogenWater(method = "Nsurplus")
}
\arguments{
\item{method}{method for calculating no3_n in groundwater}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Nitrogen Budgets for surface Water on country levels.
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenWater")
}
}
\author{
Benjamin Leon Bodirsky
}
