% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionFactors.R
\name{calcEmissionFactors}
\alias{calcEmissionFactors}
\title{Calc Emission Factors}
\usage{
calcEmissionFactors(
  subtype = "emission_factors",
  sectoral_resolution = "aggregated"
)
}
\description{
Calc Emission Factors
}
