% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionFactorsFeedstocks.R
\name{calcEmissionFactorsFeedstocks}
\alias{calcEmissionFactorsFeedstocks}
\title{Calculate emission factors for feedstocks in the chemicals industry
using emissions from UNFCCC and energy demands from IEA Energy Balances}
\usage{
calcEmissionFactorsFeedstocks()
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}.
}
\description{
Calculate emission factors for feedstocks in the chemicals industry
using emissions from UNFCCC and energy demands from IEA Energy Balances
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Falk Benke, Renato Rodrigues, Simón Moreno Leiva
}
