% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionInventory.R
\name{calcEmissionInventory}
\alias{calcEmissionInventory}
\title{calcEmisNitrogenCroplandPast}
\usage{
calcEmissionInventory(datasource = "CEDS",
  mapping = "mappingCEDS59toSectors.csv", from = "CEDS59",
  to = "Sectors")
}
\arguments{
\item{datasource}{The Inventory that shall be used. Options are CEDS, combined_CEDS_IPCC (including own estimates where available), IPCC(own estimates), Nsurplus (own estimates)}

\item{mapping}{if emission inventory shall be aggregated, this indicates the mapping csv. if NULL, the original resolution will be used.}

\item{from}{column in mapping}

\item{to}{column in mapping}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
providees an emission inventory for the past, either from external data or own estimates.
}
\examples{

\dontrun{ 
calcOutput("EmisNitrogenCroplandPast")
}
}
\author{
Benjamin Leon Bodirsky
}
