% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissions4ReportExtra.R
\name{calcEmissions4ReportExtra}
\alias{calcEmissions4ReportExtra}
\title{Calculates historical emissions needed for estimating emission factors
in extra emissions reporting}
\usage{
calcEmissions4ReportExtra(sectors = "CEDS")
}
\arguments{
\item{sectors}{"CEDS" or "IAMC" to select the sectoral aggregation of the output}
}
\value{
list of magclass with CEDS CH4 and N2O emissionsdata for
CEDS sectors for 2020, in Mt N or Mt CH4 per year
}
\description{
Calculates historical emissions needed for estimating emission factors
in extra emissions reporting
}
\author{
Gabriel Abrahao
}
