% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTarget.R
\name{calcEmiTarget}
\alias{calcEmiTarget}
\title{Output for 2 policy cases}
\usage{
calcEmiTarget(sources = "UNFCCC_NDC", subtype, scenario)
}
\arguments{
\item{sources}{A string: "UNFCCC_NDC"}

\item{subtype}{A string, one of: "Ghgshare2005", "Ghgfactor", "Ghghistshare"}

\item{scenario}{GDP and pop scenarios. Passed to \code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDP()}}.}
}
\description{
Output for 2 policy cases
}
