% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyEffPaths.R
\name{calcEnergyEffPaths}
\alias{calcEnergyEffPaths}
\title{Efficiency paths for CES variables beyond calibration CES nest}
\usage{
calcEnergyEffPaths()
}
\description{
Returns the Efficiency paths for Macro variables
}
\author{
Antoine Levesque
}
