% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyFactor.R
\name{calcEnergyFactor}
\alias{calcEnergyFactor}
\title{calcEnergyFactor}
\usage{
calcEnergyFactor()
}
\value{
List of magpie object with results on global level, empty weight, unit and description.
}
\description{
provides generalizable energy after crops are processed into 1st generation biofuel
}
\examples{

\dontrun{ 
calcOutput("EnergyFactor")
}
}
\author{
Ewerton Araujo
}
