% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyServiceCoeff.R
\name{calcEnergyServiceCoeff}
\alias{calcEnergyServiceCoeff}
\title{Calculate Demand based on the output of the Demand Model}
\usage{
calcEnergyServiceCoeff()
}
\value{
Energy Service Coefficients and corresonding weights as a list of
two MAgPIE objects
}
\description{
This function calculates enery service coefficients based in the Self energy
consumption and Specific energy input for freight transportation, iron and
cement
}
\examples{

\dontrun{ 

a <- calcOutput("EnergyServiceCoeff")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNTNU_LCA}},
\code{\link{convertNTNU_LCA}}
}
\author{
Lavinia Baumstark
}
