% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEORA.R
\name{calcEORA}
\alias{calcEORA}
\title{calcEORA}
\usage{
calcEORA(subtype)
}
\arguments{
\item{subtype}{- year (1990-2013)}
}
\value{
list of MAgPIE object of the harmonized EORA data, weight, unit, description and as note a vector
consisting of the three lost values
}
\description{
harmonizes the sectors of EORA by calculation, aggregation and disaggregation
}
\note{
\strong{Some information about the EORA database in general:}
The database is a multi-region input-output table (MRIO) database. 187 individual countries represented 
by a total of 15,909 sectors. This makes the database detailed but also difficult to handle, as the sectoral
division isn't the same for the different countries. Furthermore, sectors
can be either Commodities or Industries, but some countries even report both.

\strong{What the calcEORA function does:}
The aim was to harmonize the sectors of the EORA database and map them to magpie categories. We did this in
3 main steps: removing the distinction between Commodities and Industries (see \code{\link{calcCommoditiesEORA}}
for an explanation on how to do that), aggregating the sectors and finally disaggregating them to magpie categories
using the FAO production times IniFoodPrice.
The last two steps are done for each input country (if value flows from country A to country B, the inputcountry would
be A according to my use of this word...) individually and the resulting magpie objects are then bound toghether. 
(This means that it should be easy to parallelize that part of the function, if you have some spare time 
feel free to do so!)

At three points in this function the total value of the table is reduced: calcCommoditiesEORA looses some
value due to the rounding in \code{\link{readEORA}} and removes some value intentionally (see 
\code{\link{calcCommoditiesEORA}} for an explanation) and we have to remove ANT, USR and ROW as 
inputcountries as they don't exist in the magpie-countries. Those three lost values are returned in the
note part of the returned list, in order to be able to compare the original value of the 
table with the value of the result to check if everything went right. Sometimes there will still be a difference
of a (small) multiple of 16, this occurs if at one point the readSource function loads the EORA table from the cache
and at one point really reads it in. (If you find out why exactly this makes a difference, please tell me!)

\strong{Structure of returned MAgPIE object:} the 1st dimension is the outputcountry, 2nd is as always the year and the 3rd is divided in three: 
the input country, the input sector and the output sector.
}
\seealso{
\code{\link{calcCommoditiesEORA}}, \code{\link{readEORA}}
}
\author{
Debbora Leip
}
