% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcETSRefEmi.R
\name{calcETSRefEmi}
\alias{calcETSRefEmi}
\title{calc ETS Reference Emissions}
\usage{
calcETSRefEmi(subtype)
}
\arguments{
\item{subtype}{type of reference emissions used to define emission reduction targets for European regulations: EEA_GHG}
}
\value{
2005 reference emissions to calculate ETS targets
}
\description{
provides region specific ETS Reference Emissions
}
\examples{

\dontrun{ 
calcOutput("ETSRefEmi",subtype="EEA_GHG")
}

}
\author{
Renato Rodrigues
}
