% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEuropeanEnergyDatasheets.R
\name{calcEuropeanEnergyDatasheets}
\alias{calcEuropeanEnergyDatasheets}
\title{Calculate REMIND variables from European Energy Datasheets}
\usage{
calcEuropeanEnergyDatasheets(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "EU28" (older data from Jun 2020, including GBR) or "EU27" (data from Jun 2021)}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND variables from European Energy Datasheets
}
\author{
Falk Benke
}
