% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEurostatEmissions.R
\name{calcEurostatEmissions}
\alias{calcEurostatEmissions}
\title{Calculate REMIND emission variables from historical Eurostat (env_air_gge) values}
\usage{
calcEurostatEmissions()
}
\description{
Calculate REMIND emission variables from historical Eurostat (env_air_gge) values
}
\author{
Falk Benke
}
