% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExcretion.R
\name{calcExcretion}
\alias{calcExcretion}
\title{calcExcretion}
\usage{
calcExcretion(cellular = FALSE)
}
\arguments{
\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates excretion during grazing, cropland-grazing, confinement and collected for fuel. Based on MAgPIE Feed baskets, slaughter biomass and simple allocation rules.
}
\examples{

\dontrun{ 
calcOutput("Excretion")
}
}
\seealso{
\code{\link{calcExcretionIPCC}}
}
\author{
Benjamin Leon Bodirsky
}
