% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExcretionIPCC.R
\name{calcExcretionIPCC}
\alias{calcExcretionIPCC}
\title{calcExcretionIPCC}
\usage{
calcExcretionIPCC(products = "IPCC")
}
\arguments{
\item{products}{IPCC: IPCC products. MAgPIE: Magpie products}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates excretion in the year 2005 using the IPCC Guidelines excretion rates.
}
\examples{

\dontrun{ 
calcOutput("ExcretionIPCC")
}

}
\seealso{
\code{\link{calcExcretion}},
\code{\link{calcAnimalStocks}}
}
\author{
Benjamin Leon Bodirsky
}
