% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExogDemScen.R
\name{calcExogDemScen}
\alias{calcExogDemScen}
\title{calculate exogenuous FE and ES demand pathways}
\usage{
calcExogDemScen()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
prepare data for exogenuous FE and ES demand pathways that do not
come from EDGE models but from other sources and/or scenario literature.
REMIND can be fixed to those demand pathways if the switch cm_exogDem_scen is activated.
}
\author{
Felix Schreyer
}
