% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOCrop_aggr.R
\name{calcFAOCrop_aggr}
\alias{calcFAOCrop_aggr}
\title{Calculate FAO Crop aggregated}
\usage{
calcFAOCrop_aggr()
}
\value{
FAO land data and corresponding weights as a list of two MAgPIE
objects
}
\description{
Provides the FAOSTAT Production Crops data aggregated to magpie kcr.
}
\examples{

\dontrun{ 

a <- calcOutput("FAOCrop_aggr")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{readSource}}
}
\author{
Ulrich Kreidenweis
}
