% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOFeed.R
\name{calcFAOFeed}
\alias{calcFAOFeed}
\title{Calculate FAO feed use}
\usage{
calcFAOFeed()
}
\value{
FAO feed use and corresponding weights as a list of two MAgPIE
objects
}
\description{
Provides primary and secondary agricultural products as well as conversion
byproducts used as feed. This is based on feed use values from the Food
Balance Sheets (FBS) and Commodity Balance Sheets from FAOSTAT. Function is
based on average values of the specified years.
}
\examples{

\dontrun{ 
a <- calcFAOFeed()
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}}
}
\author{
Isabelle Weindl
}
