% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOFodder_aggrFEED.R
\name{calcFAOFodder_aggrFEED}
\alias{calcFAOFodder_aggrFEED}
\title{Calculate FAO Fodder Crop aggregated to MAgPIE_FEED sectors}
\usage{
calcFAOFodder_aggrFEED()
}
\value{
FAO fodder data and corresponding weights as a list of two MAgPIE
objects
}
\description{
Provides the FAOSTAT Fodder data aggregated to three MAgPIE_FEED fodder
categories.
}
\examples{

\dontrun{ 

a <- calcOutput("FAOFodder_aggrFEED")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{readSource}}
}
\author{
Isabelle Weindl
}
