% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOForestRelocate.R
\name{calcFAOForestRelocate}
\alias{calcFAOForestRelocate}
\title{calcFAOForestRelocate}
\usage{
calcFAOForestRelocate(selectyears = "past", track = TRUE)
}
\arguments{
\item{selectyears}{default on "past"}

\item{track}{if intermediate results shoul be printed}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates the cellular MAgPIE forest and other land area correction based on FAO forestry data and LUH2v2.
}
\examples{

\dontrun{ 
calcOutput("FAOForestRelocate")
}
}
\author{
Kristine Karstens
}
