% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOForestryDemand.R
\name{calcFAOForestryDemand}
\alias{calcFAOForestryDemand}
\title{calcFAOForestryDemand}
\usage{
calcFAOForestryDemand()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the demand of forestry products from FAO data.
}
\examples{

\dontrun{ 
calcOutput("FAOForestryDemand")
}
}
\seealso{
\code{\link{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
