% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOharmonized.R
\name{calcFAOharmonized}
\alias{calcFAOharmonized}
\title{Calculate harmonized FAO Commodity Balance and Food Supply data}
\usage{
calcFAOharmonized()
}
\value{
FAO harmonized data, weight as NULL, and a description as as a list
of MAgPIE objects
}
\description{
This functions adds the CBCrop, CBLive, FSCrop and FSLive data together
}
\examples{

\dontrun{ 
a <- calcOutput("FAOharmonized")

}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Ulrich Kreidenweis
}
