% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLand.R
\name{calcFAOLand}
\alias{calcFAOLand}
\title{calcFAOLand}
\usage{
calcFAOLand()
}
\value{
land areas from FAOSTAT and weight
}
\description{
Returns either physical land areas from FAOSTAT.
}
\author{
Ulrich Kreidenweis
}
