% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLiveHead.R
\name{calcFAOLiveHead}
\alias{calcFAOLiveHead}
\title{Calculate FAO Living Animals}
\usage{
calcFAOLiveHead()
}
\value{
FAO Living Animals data and corresonding weights as a list of two
MAgPIE objects
}
\description{
Provides the FAO Living Animals data. No changes to the content have been
done, besides changes to the countries list.
}
\examples{

\dontrun{ 
calcOutput("LiveHead")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{readSource}}
}
\author{
Ulrich Kreidenweis
}
