% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLivePrim.R
\name{calcFAOLivePrim}
\alias{calcFAOLivePrim}
\title{Calculate FAO Livestock Primary}
\usage{
calcFAOLivePrim()
}
\value{
FAO Livestock Primary data and corresonding weights as a list of two
MAgPIE objects
}
\description{
Provides the FAO Livestock Primary data. No changes to the content have been
done, besides changes to the countries list. Yield information has been
removed.
}
\examples{

\dontrun{ 
calcOutput("LivePrim")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{readSource}}
}
\author{
Ulrich Kreidenweis
}
