% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLivePrim_aggr.R
\name{calcFAOLivePrim_aggr}
\alias{calcFAOLivePrim_aggr}
\title{Calculate FAO Livestock Primary aggregated}
\usage{
calcFAOLivePrim_aggr()
}
\value{
FAO Livestock Primary data and corresonding weights as a list of two
MAgPIE objects
}
\description{
Provides the FAO Livestock Primary data aggregated to MAgPIE categories.
}
\examples{

\dontrun{ 
calcOutput("FAOLivePrim_aggr")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFAO}},
\code{\link{convertFAO}}, \code{\link{calcFAOLivePrim}},
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
