% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOmassbalance.R
\name{calcFAOmassbalance}
\alias{calcFAOmassbalance}
\title{calcFAOmassbalance}
\usage{
calcFAOmassbalance()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates a massbalance dataset of agricultural production, processing and
use out of the combined data of calcFAOharmonized(). Covers dry matter (DM),
reactive nitrogen (Nr), Phosphorus (P), Generalizable Energy (GE) and wet
matter (WM). New products are added to the Food Balance Sheets, and many processing
conversions are made more explicit using simple assumptions. The first part of this
function is the calcFAOmassbalance_pre.
}
\examples{

\dontrun{ 
calcOutput("FAOmassbalance")
}
}
\seealso{
\code{\link{calcFAOmassbalance_pre}}
}
\author{
Benjamin Leon Bodirsky, Xiaoxi Wang
}
