% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEdemand.R
\name{calcFEdemand}
\alias{calcFEdemand}
\title{Calculates Final Energy Demand for Industry and Buildings}
\usage{
calcFEdemand(scenario)
}
\arguments{
\item{scenario}{GDP and pop scenarios. Passed to \code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDP()}}.}
}
\description{
Calculates Final Energy Demand for Industry and Buildings
}
\author{
Falk Benke
}
