% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandBuildings.R
\name{calcFeDemandBuildings}
\alias{calcFeDemandBuildings}
\title{Returns the EDGE-Buildings data at the REMIND level}
\usage{
calcFeDemandBuildings(subtype)
}
\arguments{
\item{subtype}{Final energy (FE) or Energy service (ES)}
}
\description{
Returns the EDGE-Buildings data at the REMIND level
}
\author{
Robin Hasse, Falk Benke
}
