% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandIndustry.R
\name{calcFeDemandIndustry}
\alias{calcFeDemandIndustry}
\title{Calculates FE demand in industry as REMIND variables}
\usage{
calcFeDemandIndustry(use_ODYM_RECC = FALSE)
}
\arguments{
\item{use_ODYM_RECC}{per-capita pathways for `SDP_xx` scenarios?  (Defaults
to `FALSE`.)}
}
\description{
Calculates FE demand in industry as REMIND variables
}
\author{
Michaja Pehl
}
