% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandONONSPEC.R
\name{calcFeDemandONONSPEC}
\alias{calcFeDemandONONSPEC}
\title{Calculate historic and projected other non-specified energy demand}
\usage{
calcFeDemandONONSPEC()
}
\value{
list with MagPIE object
}
\description{
Project the IEA flow ONONSPEC into the future.
}
\author{
Robin Hasse
}
