% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandTransport.R
\name{calcFeDemandTransport}
\alias{calcFeDemandTransport}
\title{Calculates FE demand in transport as REMIND variables}
\usage{
calcFeDemandTransport()
}
\description{
Calculates FE demand in transport as REMIND variables
}
\author{
Alois Dirnaicher, Johanna Hoppe
}
