% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBalanceflow.R
\name{calcFeedBalanceflow}
\alias{calcFeedBalanceflow}
\title{calcFeedBalanceflow}
\usage{
calcFeedBalanceflow(per_livestock_unit = FALSE, cellular = FALSE,
  products = "kall", future = TRUE)
}
\arguments{
\item{per_livestock_unit}{default false}

\item{cellular}{if TRUE value is calculate on cellular level}

\item{products}{products in feed baskets that shall be reported}

\item{future}{if FALSE, only past years will be reported (reduces memory)}
}
\value{
List of magpie objects with results on country or cellular level, unit and description.
}
\description{
Calculates feed balanceflows from MAgPIE-Feed model to meet FAO data
}
\examples{

\dontrun{ 
calcOutput("FeedBalanceflow")
}

}
\author{
Isabelle Weindl, Kristine Karstens
}
