% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBaskets.R
\name{calcFeedBaskets}
\alias{calcFeedBaskets}
\title{calcFeedBaskets}
\usage{
calcFeedBaskets(non_eaten_food = FALSE)
}
\arguments{
\item{non_eaten_food}{if TRUE, non-eaten food is included in feed baskets, if not it is excluded.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Combines feed baskets of the past with scenario-dependent future feed baskets.
}
\examples{

\dontrun{ 
calcOutput("FeedBaskets")
}
}
\author{
Isabelle Weindl, Benjamin Leon Bodirsky, Stephen Wirth, Jan Philipp Dietrich
}
