% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBasketsPast.R
\name{calcFeedBasketsPast}
\alias{calcFeedBasketsPast}
\title{Calculate historical feed baskets based on output of MAgPIE_FEED model
as DM feed biomass (different types of feed) needed per DM livestock products}
\usage{
calcFeedBasketsPast(non_eaten_food = TRUE)
}
\arguments{
\item{non_eaten_food}{if TRUE, non-eaten food is included in feed baskets, if not it is excluded.}
}
\value{
Historical feed baskets and corresonding weights as a list of two MAgPIE
objects
}
\description{
Calculate historical feed baskets based on output of MAgPIE_FEED model
as DM feed biomass (different types of feed) needed per DM livestock products
}
\examples{

\dontrun{ 
calcOutput("FeedBasketsPast")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFeedModel}}
}
\author{
Isabelle Weindl, Benjamin Bodirsky
}
