% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBasketsSysPast.R
\name{calcFeedBasketsSysPast}
\alias{calcFeedBasketsSysPast}
\title{Calculate historical system-specific feed baskets based on output of MAgPIE_FEED model
as DM feed biomass (different types of feed) needed per DM livestock products of respective systems}
\usage{
calcFeedBasketsSysPast()
}
\value{
Historical system-specific feed baskets and corresonding weights as a list of two MAgPIE
objects
}
\description{
Calculate historical system-specific feed baskets based on output of MAgPIE_FEED model
as DM feed biomass (different types of feed) needed per DM livestock products of respective systems
}
\examples{

\dontrun{ 
calcOutput("FeedBasketsSysPast")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFeedModel}}, \code{\link{calcFeedBasketsPast}}
}
\author{
Isabelle Weindl, Benjamin Bodirsky, Jan Philipp Dietrich
}
