% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedEfficiencyFuture.R
\name{calcFeedEfficiencyFuture}
\alias{calcFeedEfficiencyFuture}
\title{calcFeedEfficiencyFuture}
\usage{
calcFeedEfficiencyFuture()
}
\value{
Central feed shares and weights as list of two MAgPIE-objects
}
\description{
Calculates future central feed shares for all livestock categories 
based on the results of a non-linear regression between historical 
central feed shares and livestock productivity and using Koeppen-
Geiger climate zones
}
\examples{
\dontrun{ 
calcOutput("FeedEfficiencyFuture")

}
}
\author{
Isabelle Weindl, Benjamin Bodirsky, Stephen Wirth
}
