% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedPast.R
\name{calcFeedPast}
\alias{calcFeedPast}
\title{calcFeedPast}
\usage{
calcFeedPast(balanceflow = TRUE, cellular = FALSE, products = "kall",
  nutrients = "all")
}
\arguments{
\item{balanceflow}{if TRUE, non-eaten food is included in feed baskets, if not it is excluded.}

\item{cellular}{if TRUE value is calculate on cellular level with returned datajust in dry matter}

\item{products}{products in feed baskets that shall be reported}

\item{nutrients}{nutrients like dry matter (DM), reactive nitrogen (Nr), Phosphorus (P), Generalizable Energy (GE) and wet
matter (WM).}
}
\value{
List of magpie objects with results on country or cellular level, unit and description.
}
\description{
Combines feed baskets of the past with livestock production to get total feed demand
}
\examples{

\dontrun{ 
calcOutput("FeedPast")
}
}
\author{
Isabelle Weindl, Benjamin Leon Bodirsky, Kristine Karstems
}
