% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFertilizerByCrop.R
\name{calcFertilizerByCrop}
\alias{calcFertilizerByCrop}
\title{calcFertilizerByCrop}
\usage{
calcFertilizerByCrop(indicator = "total", deposition = "Nsurplus2",
  cellular = FALSE)
}
\arguments{
\item{indicator}{total: estimates the inputs per total crop production; by_harvest estimates the inputs per ton harvest; by_area estimates the inputs per area harvested}

\item{deposition}{if FALSE, deposition is not accounted for in the distribution. Use FALSE to avoid circularities in calcNitrogenBudget}

\item{cellular}{cellular disaggreagation or national values}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the crop-specific use of different N inputs
}
\examples{

\dontrun{ 
calcOutput("FertilizerByCrop")
}
}
\seealso{
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
