% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFertP.R
\name{calcFertP}
\alias{calcFertP}
\title{Calculate Fertilizer of Phosphor}
\usage{
calcFertP()
}
\value{
Fertilizer data for Phosphor and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides FertP data for Phosphor.No changes to the content have been done.
}
\examples{

\dontrun{ 
calcOutput("FertP")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readIFA}},
\code{\link{convertIFA}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
