% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFETaxes.R
\name{calcFETaxes}
\alias{calcFETaxes}
\title{Calculate FETaxes}
\usage{
calcFETaxes(subtype = "taxes")
}
\arguments{
\item{subtype}{choose between tax rates ("taxes") or subsidies rate ("subsidies") output}
}
\value{
MAgPIE object
}
\description{
Reads in the data of the source IIASA_subs_taxes, by country. and
calculate taxes at the final energy delivery level to the end-use sectors
(industry, buildings and transport). Regional aggregation is done via the
respective energy quantities as weights.
}
\examples{

\dontrun{
calcOutput("FETaxes")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readIIASA_subs_taxes}},
\code{\link{convertIIASA_subs_taxes}}
}
\author{
Christoph Bertram and Renato Rodrigues
}
