% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFGas.R
\name{calcFGas}
\alias{calcFGas}
\title{generate F-Gases based on IMAGE data}
\usage{
calcFGas(subtype = "IMAGE2025", interp = "none")
}
\arguments{
\item{subtype}{"IMAGElegacy" will use the old IMAGE data, "IMAGE2025" will use the new IMAGE2025 data.}

\item{interp}{"interpolate2025" will intepolate from EDGAR historical data from 2025-2050 to account for the very old IMAGE
scenarios. Any other interp value will ignore this step.}
}
\value{
magpie object with F-gases information
}
\description{
generate F-Gases based on IMAGE data
}
\examples{
\dontrun{
x <- calcOutput("FGas")
}
}
\author{
Lavinia Baumstark, Gabriel Abrahao
}
