% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestArea.R
\name{calcForestArea}
\alias{calcForestArea}
\title{calcForestArea}
\usage{
calcForestArea(selectyears = "past")
}
\arguments{
\item{selectyears}{defaults to past}
}
\value{
List of magpie object with results on country level, weight, unit and description.
}
\description{
Calculates consistent forest area and its subcategories based on FAO_FRA2015 and LanduseInitialisation data.
}
\examples{

\dontrun{ 
calcOutput("ForestArea")
}
}
\author{
Kristine Karstens
}
