% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestChange.R
\name{calcForestChange}
\alias{calcForestChange}
\title{Forest change rates}
\usage{
calcForestChange()
}
\value{
writes a magpie object with the new regions to the output-folder
defined in the file "exampleconfig.cfg"
}
\description{
Calculates average annual forest change rates for 1995, 2000, 2005 and 2010
from source FAO_forestry_extent. Constant after 2010. Positive values
inidicate deforestation, negative vlaues afforestation. A weight for the
aggregation is not implemented
}
\seealso{
\code{\link{setConfig}}, \code{\link{readSource}},
\code{\link{readFAO_forestry}}, \code{\link{calcOutput}}
}
\author{
Florian Humpenoeder
}
