% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestProd.R
\name{calcForestProd}
\alias{calcForestProd}
\title{Aggregate forest area designated for production values from source
FAO_forestry_production}
\usage{
calcForestProd()
}
\value{
writes a magpie object with the new regions to the output-folder
defined in the file "exampleconfig.cfg"
}
\description{
Aggregates forest area designated for production values from source
FAO_forestry_production over the magpie-regions. NAs get replaced by zero. A
weight for the aggregation is not implemented
}
\seealso{
\code{\link{setConfig}}, \code{\link{readSource}},
\code{\link{readFAO_forestry}}, \code{\link{calcOutput}}
}
\author{
Nele Steinmetz
}
