% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestProdArea.R
\name{calcForestProdArea}
\alias{calcForestProdArea}
\title{Aggregate forest area designated for production values from source
FAO_forestry_production}
\usage{
calcForestProdArea()
}
\value{
writes a magpie object with the new regions to the output-folder
defined in the file "exampleconfig.cfg"
}
\description{
Aggregates forest area designated for production values from source
FAO_forestry_production over the magpie-regions. A
weight for the aggregation is not implemented
}
\seealso{
\code{\link{setConfig}}, \code{\link{readSource}},
\code{\link{readFAO_forestry}}, \code{\link{calcOutput}}
}
\author{
Ulrich Kreidenweis
}
