% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestProtect.R
\name{calcForestProtect}
\alias{calcForestProtect}
\title{Aggregate forest area with protected areas values from source
FAO_forestry_protection}
\usage{
calcForestProtect()
}
\value{
writes a magpie object with the new regions to the output-folder
defined in the file "exampleconfig.cfg"
}
\description{
Aggregates forest area within protected areas values from source
FAO_forestry_protection over the magpie-regions. NAs get replaced by zero. A
weight for the aggregation is not implemented
}
\seealso{
\code{\link{setConfig}}, \code{\link{readSource}},
\code{\link{readFAO_forestry}}, \code{\link{calcOutput}}
}
\author{
Nele Steinmetz
}
