% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestProtectionShare.R
\name{calcForestProtectionShare}
\alias{calcForestProtectionShare}
\title{calcForestProtectionShare}
\usage{
calcForestProtectionShare(cellular = FALSE)
}
\arguments{
\item{cellular}{cellular (TRUE) or country-level/regional (FALSE) data? For country-level vs regional data, remember to set "aggregate" to false.}
}
\value{
List of magpie object with results on country or cellular level, weight on cellular level, unit and description.
}
\description{
calculates the share of protected forests. Data from FAO on protected forests is used to split the primary and secondary forest pool of the LU2v1 dataset.
}
\examples{

\dontrun{ 
calcOutput("ForestProtectionShare")
}
}
\author{
Benjamin Leon Bodirsky
}
