% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFossilExtraction.R
\name{calcFossilExtraction}
\alias{calcFossilExtraction}
\title{calc Fossil Extraction}
\usage{
calcFossilExtraction(subtype = "FossilExtraction")
}
\arguments{
\item{subtype}{Either 'FossilExtraction' or 'UraniumExtraction'}
}
\value{
magpie object of the coefficients for fossil fuels and uranium extraction cost equations
}
\description{
provides coefficients for fossil fuels (oil, gas and coal) and uranium extraction cost equations.
}
\examples{
\dontrun{
calcOutput(type = "FossilExtraction", subtype = "FossilExtraction")
}
}
\author{
Renato Rodrigues, Felix Schreyer
}
