% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGAINS.R
\name{calcGAINS}
\alias{calcGAINS}
\title{calcGAINS}
\usage{
calcGAINS(subtype = "emission_factors", sectoral_resolution = "extended")
}
\arguments{
\item{subtype}{decides whether emissions or emission factors are returned}

\item{sectoral_resolution}{aggreaged or extenden (uses different GAINS input data)}
}
\description{
Calculates air pollutant emissions and emission factors (user can choose)
based on GAINS emissions and activity data. Result is given on GAINS sector level.
User can choose between aggregated and extended sectoral resolution. Results are
given for multiple scenarios. Scenario design is partly taken from the GAINS data
and partly created in this function (particularly the SSPs).
}
