% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGAINS2025.R
\name{calcGAINS2025}
\alias{calcGAINS2025}
\title{Calculate air pollutant emission factors for all scenarios and SSPs
available from GAINS, at the level of GAINS sectors and for 2005-2100.}
\usage{
calcGAINS2025(
  weight_source = "CEDS2025",
  outsectors = "GAINS2025",
  outunit = "Tg/TWa"
)
}
\arguments{
\item{weight_source}{Source of air pollutant reference emissions in 2020 that
is used to derive the weights ("CEDS2025" or "GAINS2025")}

\item{outsectors}{Output sectoral aggregation ("GAINS2025" or "REMIND")}

\item{outunit}{Output unit for emission factors ("kt/PJ" or "Tg/TWa")}
}
\value{
Emission factor timeseries for all scenarios from 2005 to 2100:
magclass object with dimensions region, year, and
ssp.scenario.sectorGAINS.species
}
\description{
This function is meant to be used to clean-up, fill gaps and smoothen
the GAINS data to obtain consistent timeseries of emission factors
from 2005 to 2100.
The actual generation REMIND-specific files happens in calcGAINS2025forREMIND.
}
\author{
Gabriel Abrahao, Laurin Koehler-Schindler
}
