% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGAINS2025forREMIND.R
\name{calcGAINS2025forREMIND}
\alias{calcGAINS2025forREMIND}
\title{Calculates air pollutant emissions, activities and emission factors
for all scenarios and SSPs available from GAINS, including both
legacy and 2025 variants of GAINS data}
\usage{
calcGAINS2025forREMIND(subtype)
}
\arguments{
\item{subtype}{"emission_factors", "emissions","emissions_starting_values"}
}
\value{
Activity levels, emissions or emission factors
}
\description{
Calculates air pollutant emissions, activities and emission factors
for all scenarios and SSPs available from GAINS, including both
legacy and 2025 variants of GAINS data
}
\author{
Gabriel Abrahao
}
