% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGAINSEmi.R
\name{calcGAINSEmi}
\alias{calcGAINSEmi}
\title{Calculate air pollution emissions and emission factors from GAINS data}
\usage{
calcGAINSEmi(subtype = "emissions")
}
\arguments{
\item{subtype}{"emission_factors", "emissions","emissions_starting_values"}
}
\value{
Emissions and emission factors
}
\description{
Provides input data for exoGAINSAirpollutants.R
}
\examples{
\dontrun{
calcOutput("calcGAINSEmi")
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Sebastian Rauner
}
