% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPpc.R
\name{calcGDPpc}
\alias{calcGDPpc}
\title{calcGDPpc}
\usage{
calcGDPpc(gdp = "PPP")
}
\arguments{
\item{gdp}{Can be switched between ppp and MER. Default is ppp.}
}
\value{
list of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates GDP per capita
}
\examples{

\dontrun{ 
calcOutput("GDPpc", gdp="ppp")
}
}
\seealso{
\code{\link{calcPopulation}},
\code{\link{calcGDPppp}}
}
\author{
Benjamin Leon Bodirsky, Abhijeet Mishra
}
