% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPppp.R
\name{calcGDPppp}
\alias{calcGDPppp}
\title{calcGDPppp}
\usage{
calcGDPppp(GDPpppCalib = "past_transition",
  GDPpppPast = "IHME_USD05_PPP_pc_completed",
  GDPpppFuture = "SRES_SSP_completed", FiveYearSteps = TRUE,
  naming = "indicator_scenario")
}
\arguments{
\item{GDPpppCalib}{to what should be calibrated? past, future or a transition?}

\item{GDPpppPast}{GDPppp past data source}

\item{GDPpppFuture}{GDPppp future data source}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
GDP PPP(ICP11) in million USD05 equivalents
}
\description{
Merges time series of GDP in Purchase Power Parity (PPP) of International
Dollars of the year 2005.  The source is selected in the config file. See
\code{\link{calcGDPpppPast}} for past datasets, and
\code{\link{calcGDPpppFuture}} for future datasets. The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point. All data is calibrated either to the "past" or
the "future" dataset as specified in "getConfig()$calc$PopulationCalib".
\itemize{ \item \code{WDI}: The PPP estimate from the World Developmnet
Indicators (WDI) are supplemented by values for Argentina, Syria and Somalia
which are missing in the database. The values were taken from World Bank.
2014. Purchasing Power Parities and the Real Size of World Economies: A
Comprehensive Report of the 2011 International Comparison Program. The World
Bank. http://elibrary.worldbank.org/doi/book/10.1596/978-1-4648-0329-1.
table 2.13 Then, the 2011 estimate is extrapolated with the GDP projection
in local currency units (LCU), as these growth rates should be approximately
what the growth models predict. The price index from 2011 was transformed
into 2005 equivalents using the inflation rate of the United States(US), as
the PPPs are in USDollar equvialents. \item \code{PWP}: Penn World Tables }
}
\seealso{
\code{\link{convertWDI}}
}
\author{
Lavinia Baumstark, Benjamin Bodirsky
}
