% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGEA2012.R
\name{calcGEA2012}
\alias{calcGEA2012}
\title{calcGEA2012}
\usage{
calcGEA2012(subtype, datatype)
}
\arguments{
\item{subtype}{oil, coal, gas, or bounds}

\item{datatype}{extraseed, exportbound, or decoffset for bounds subtype}
}
\value{
MAgPIE object containing regionally aggregated GEA 2012 data
}
\description{
Extracts oil, gas and coal data from the GEA 2012 into a scenario- and time-dependent grade structure
}
\examples{

\dontrun{ a <- calcOutput("GEA2012")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Stephen Bi
}
