% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGHGPrices.R
\name{calcGHGPrices}
\alias{calcGHGPrices}
\title{calcGHGPrices}
\usage{
calcGHGPrices(emissions = "pollutants", datasource = "REMMAG",
  modelinput_format = FALSE)
}
\arguments{
\item{emissions}{which type of emissions shall be returned. ghg just returns n2o, ch4 and co2, pollutants a longer list including also air pollutants}

\item{datasource}{REMMAG or SSPresults}

\item{modelinput_format}{if modelinput_format is TRUE, the points in the scenarionames for models and versions will be replaced by a hyphen.}
}
\value{
list of magpie object with results on country level, weight on country level, unit and description.
}
\description{
reads in GHG prices from past runs
}
\examples{
  

\dontrun{ 
calcOutput("GHGPrices")
}
}
\seealso{
\code{\link{readSSPResults}}
}
\author{
David Chen, Benjamin Leon Bodirsky
}
