% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGlobalEnergyMonitor.R
\name{calcGlobalEnergyMonitor}
\alias{calcGlobalEnergyMonitor}
\title{Calculate REMIND variables from Global Energy Monitor}
\usage{
calcGlobalEnergyMonitor()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND variables from Global Energy Monitor
}
\author{
Falk Benke
}
