% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGMIA.R
\name{calcGMIA}
\alias{calcGMIA}
\title{calcGMIA}
\usage{
calcGMIA()
}
\value{
list of magpie object with data and weight
}
\description{
Filling gaps in the Historical area equipped for irrigation from GMIA.
}
\examples{

\dontrun{ 
a <- calcOutput("GMIA",aggregate="regglo")
}
}
\author{
Stephen Wirth, Anne Biewald
}
