% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGMPReduction.R
\name{calcGMPReduction}
\alias{calcGMPReduction}
\title{Calculates Global Methane Pledge (GMP) pledged CH4 emissions reductions in 2030 for each region,
assuming 30\% reduction relative to 2020 levels and using CEDS2020 emissions as weights.}
\usage{
calcGMPReduction()
}
\value{
MAgPIE object with regional pledged reduction fraction of 2020 emissions by all countries,
weighted by CEDS2020 emissions of pledged countries
}
\description{
Information about the list of countries that have commited to the GMP
can be found under mrremind::readGlobalMethanePledge()
}
\author{
Gabriel Abrahao
}
