% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGTAP.R
\name{calcGTAP}
\alias{calcGTAP}
\title{calcGTAP}
\usage{
calcGTAP(gtap_version = "GTAP7", subtype = NULL, bilateral = FALSE)
}
\arguments{
\item{gtap_version}{type of GTAP data version
\itemize{
\item \code{GTAP7}
\item \code{GTAP8}
}}

\item{subtype}{GTAP subtype}

\item{bilateral}{whether bilateral trade margin should be calculated}
}
\value{
Trade margins as an MAgPIE object
}
\description{
calculate total value of trade margins from GTAP dataset
}
\examples{
    \dontrun{
    x <- calcGTAP("GTAP7","VXMD")
    }
}
\author{
Xiaoxi Wang
}
