% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGTAPTrade.R
\name{calcGTAPTrade}
\alias{calcGTAPTrade}
\title{calcGTAPTrade}
\usage{
calcGTAPTrade(subtype = NULL, bilateral = FALSE)
}
\arguments{
\item{subtype}{GTAP subtype}

\item{bilateral}{whether bilateral trade data should be calculated}
}
\value{
Trade related data as an MAgPIE object
}
\description{
calculate trade data from GTAP dataset
}
\examples{
    \dontrun{
    x <- calcGTAP("GTAP7_VXMD")
    }
}
\author{
Xiaoxi Wang
}
