% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHDD_CDD.R
\name{calcHDD_CDD}
\alias{calcHDD_CDD}
\title{HDD and CDD scenario and historic data}
\usage{
calcHDD_CDD(tlimit = 18)
}
\arguments{
\item{tlimit}{Temperature threshold for computing HDD and CDD: 18, 21 or 25}
}
\description{
smoothes the HDD and CDD scenario data
}
\author{
Antoine Levesque
}
