% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHHFoodWaste.R
\name{calcHHFoodWaste}
\alias{calcHHFoodWaste}
\title{Calculate food waste}
\usage{
calcHHFoodWaste()
}
\value{
food waste and corresponding weights as a list of two MAgPIE
objects
}
\description{
Provides aggregated household food waste in the units Mt DM, GE, K,  NR, P and WM.
}
\examples{

\dontrun{ 
a <- calcOutput("HHFoodWaste")
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFoodWasteRecycle}}
}
\author{
Isabelle Weindl, Benjamin Leon Bodirsky
}
