% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistoricalLUEmissions.R
\name{calcHistoricalLUEmissions}
\alias{calcHistoricalLUEmissions}
\title{Calculate historical landuse emissions}
\usage{
calcHistoricalLUEmissions()
}
\description{
Calculate historical landuse emissions
}
\author{
Falk Benke
}
