% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHRE.R
\name{calcHRE}
\alias{calcHRE}
\title{Calculate Final Energy for the buildings sector from Heat Roadmap Europe scenarios}
\usage{
calcHRE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate Final Energy for the buildings sector from Heat Roadmap Europe scenarios
}
\author{
Pascal Weigmann
}
