% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_EB_directSum.R
\name{calcIEA_EB_directSum}
\alias{calcIEA_EB_directSum}
\title{Calculate simple sums of IEA Energy Balances}
\usage{
calcIEA_EB_directSum()
}
\description{
Calculates \code{PE|Coal}, \code{PE|Oil}, \code{PE|Gas}, and \code{FE} as direct sums of relevant
products for \code{TES} (PEs) or \code{TFC} (FE), minus (negative) bunkers.  Used for
benchmarking in the \code{historical.mif} and for the IIASA Scenario Compass
Initiative vetting.
}
