% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_ETP.R
\name{calcIEA_ETP}
\alias{calcIEA_ETP}
\title{Calculate REMIND emission variables from IEA ETP values}
\usage{
calcIEA_ETP(aggregate, isValidation = FALSE)
}
\arguments{
\item{aggregate}{Boolean indicating whether output data aggregation should be performed or not}

\item{isValidation}{indicates if result will be used in validation (as opposed to generating input data)}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND emission variables from IEA ETP values
}
\author{
Falk Benke
}
