% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_PFU.R
\name{calcIEA_PFU}
\alias{calcIEA_PFU}
\title{calcIEA_PFU}
\usage{
calcIEA_PFU()
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes data for the Primary Final Useful (PFU) resolution
}
\examples{

\dontrun{ a <- calcOutput("IEA_PFU")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Antoine Levesque
}
