% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_WorldEnergyOutlook.R
\name{calcIEA_WorldEnergyOutlook}
\alias{calcIEA_WorldEnergyOutlook}
\title{Calculate REMIND variables from IEA World Energy Outlook data.}
\usage{
calcIEA_WorldEnergyOutlook(version = "latest")
}
\arguments{
\item{version}{of the WEO data, either 'default' (full paid version) or
'latest' (free, restricted, up-to-date dataset)}
}
\description{
Calculate REMIND variables from IEA World Energy Outlook data.
}
\author{
Falk Benke
}
