% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIeaWorldEnergyOutlook.R
\name{calcIeaWorldEnergyOutlook}
\alias{calcIeaWorldEnergyOutlook}
\title{Calculate REMIND variables from IEA World Energy Outlook data.}
\usage{
calcIeaWorldEnergyOutlook()
}
\description{
Calculate REMIND variables from IEA World Energy Outlook data.
}
\author{
Falk Benke
}
