% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcILO.R
\name{calcILO}
\alias{calcILO}
\title{calcILO
ILO data for employment in oil and gas; coal and lignite}
\usage{
calcILO(subtype)
}
\arguments{
\item{subtype}{"Reliable" and "all". ILO classifies its data per country into reliable, unreliable, and break in series (years without data). "Reliable" contains both reliable and break-in-series-data, "all" contains both reliable and unreliable data}
}
\value{
magpie object with employment data for different years
}
\description{
calcILO
ILO data for employment in oil and gas; coal and lignite
}
\author{
Aman Malik
}
