% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustry_CCS_limits.R
\name{calcIndustry_CCS_limits}
\alias{calcIndustry_CCS_limits}
\title{Calculate Limits on Industry CCS Capacities}
\usage{
calcIndustry_CCS_limits(
  annual_growth_factor = 1.2,
  stage_weight = c(Operational = 1, `In construction` = 1, `Advanced development` = 0.5,
    `Early development` = 0.2),
  facility_subsector = c(Cement = "cement", Chemical = "chemicals",
    `Hydrogen / Ammonia / Fertiliser` = "chemicals", Ethan = "chemicals",
    `Iron and Steel Production` = "steel"),
  region_mapping = NULL
)
}
\arguments{
\item{annual_growth_factor}{Annual growth factor of CCS capacity limits,
defaults to 1.2 (20 \% annually).}

\item{stage_weight}{A named vector of weight factors for different lifecycle
stages.  See Details.}

\item{facility_subsector}{A named vector mapping the "Facility Industry" of
CCS projects to REMIND industry subsectors.  See Details.}

\item{region_mapping}{A data frame with columns \code{iso3c} and \code{region} detailing
the regional resolution on which data should be extrapolated.  If \code{NULL}
(the default), extrapolation is done at the country level.}
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, and \code{min}.
}
\description{
Calculate Limits on Industry CCS Capacities
}
\details{
The limits on industry CCS capacities are calculated from data of the
\href{zotero://select/items/3_E5GNNPZ8}{Global Status of CCS 2023} report (through
\code{\link[=readGlobalCCSinstitute]{readGlobalCCSinstitute()}}.  CCS projects are
\itemize{
\item filtered for valid (i.e. not "Under Evaluation") data for "Operation date"
and "CO~2~ capture capacity"
\item assigned to REMIND industry subsectors according to \code{facility_subsector},
which defaults to\tabular{ll}{
   Facility Industry \tab subsector \cr
   Cement \tab cement \cr
   Chemical \tab chemicals \cr
   Hydrogen / Ammonia / Fertiliser \tab chemicals \cr
   Ethan \tab chemicals \cr
   Iron and Steel Production \tab steel \cr
}

\item weighted by lifecycle stage according to \code{stage_weight}, which defaults to\tabular{lr}{
   Lifecycle stage \tab weight \cr
   Operational \tab 100 \% \cr
   In construction \tab 100 \% \cr
   Advanced development \tab 50 \% \cr
   Early development \tab 20 \% \cr
}

}

The resulting project capacities constitute the limits on industry subsector
CCS capacity for 2030.  The limit on CCS capacities for regions (or countries
if \code{region_mapping} is \code{NULL}) is set to a value of total 2030 subsector CCS
capacity, times the regions share in 2030 subsector activity (e.g. cement
production)
\itemize{
\item in 2035 if the region as some CCS capacity in 2030 in a different industry
subsector, or
\item in 2040 if the region has no industry CCS capacity in 2030 at all.
}

CCS capacities are increase by the annual growth factor (default: 20 \%/p.a.).
}
\author{
Michaja Pehl
}
