% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustry_EEK.R
\name{calcIndustry_EEK}
\alias{calcIndustry_EEK}
\title{Industry Energy Efficiency Capital}
\usage{
calcIndustry_EEK(kap)
}
\arguments{
\item{kap}{General internal capital stock, as calculated internally by
`calcCapital()`.}
}
\value{
A list with a [`magpie`][magclass::magclass] object `x`, `weight`,
  `unit`, and `description` fields.
}
\description{
Industry Energy Efficiency Capital
}
