% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcindustry_max_secondary_steel_share.R
\name{calcindustry_max_secondary_steel_share}
\alias{calcindustry_max_secondary_steel_share}
\title{Calculate Maximum Secondary Steel Production Share}
\usage{
calcindustry_max_secondary_steel_share(scenarios = NULL, regions = NULL)
}
\arguments{
\item{scenarios}{A character vector of scenarios to expand data to.}

\item{regions}{A character vector of regions to expand data to.}
}
\value{
A list with a [`magpie`][magclass::magclass] object `x`.
}
\description{
Reads ExpertGuess/industry_max_secondary_steel_share and expands to all
`scenarios`/`regions` using default data.  See [`tool_expand_tibble()`] for
details.
}
