% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcindustry_specific_FE_limits.R
\name{calcindustry_specific_FE_limits}
\alias{calcindustry_specific_FE_limits}
\title{Thermodynamic Limits for Industry Specific FE Demand}
\usage{
calcindustry_specific_FE_limits()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Return \code{readExpertGuess('industry_specific_FE_limits')} in a format usable as
a REMIND input.
}
\author{
Michaja Pehl
}
