% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIniFoodPrice.R
\name{calcIniFoodPrice}
\alias{calcIniFoodPrice}
\title{calcIniFoodPrice}
\usage{
calcIniFoodPrice(datasource = "FAO", year = "y2005",
  products = "kfo")
}
\arguments{
\item{datasource}{The datasource specification. Currently available \code{FAO} and \code{IMPACT3.2.2World_Price}.}

\item{year}{Specifies the year for the initialization of prices in MAgPIE. Default is \code{y2005}.}

\item{products}{subselection of products to be returned}
}
\value{
List with a magpie object with commodity prices on global level in $05/tDM.
}
\description{
provides global prices from the IMPACT model projections for MAgPIE commodities
for the initialization of the flexible demand model.
}
\note{
The IMPACT projections start in 2005 and the prices are taken from that year.
}
\examples{

\dontrun{ 
calcOutput("IniFoodPrice", aggregate=FALSE)
}

}
\seealso{
\code{\link{readIMPACT3.2.2World_Price}}
}
\author{
Mishko Stevanovic, Benjamin Leon Bodirsky
}
