% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntake.R
\name{calcIntake}
\alias{calcIntake}
\title{calcIntake}
\usage{
calcIntake(convert = TRUE, modelinput = FALSE, standardize = FALSE,
  method = "FAO_WHO_UNU1985")
}
\arguments{
\item{convert}{if TRUE, Lutz data is converted (interpolated completed etc)}

\item{modelinput}{if TRUE, data is aggregated to country totals for model input}

\item{standardize}{if FALSE, no standardization. if "recommendations", the US recommendations are used. if BMI, a normal BMI is used.}

\item{method}{method for calculating intake: either FAO_WHO_UNU1985 for estimates based on height and bodyweight, schofield for just bodyweight, or HHS_USDA for recommended values for US-americans}
}
\value{
total "healthy" intake kcal/day procap for each countries divided by sex and 8 age groups.
}
\description{
it computes the total intake kcal/day procapita through the population dataset by Lutz 2014,
height and BMI data of NCDrisc, and phyiscal inactivity levels based on Halal et al.
}
