% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntakeByGroup.R
\name{calcIntakeByGroup}
\alias{calcIntakeByGroup}
\title{calcIntakeByGroup}
\usage{
calcIntakeByGroup(convert = TRUE, method = "schofield")
}
\arguments{
\item{convert}{if TRUE, Lutz data is converted (interpolated completed etc)}

\item{method}{method for calculating intake: either FAO_WHO_UNU1985 for estimates based on height and bodyweight, schofield for just bodyweight, or HHS_USDA for recommended values for US-americans}
}
\value{
total intake kcal/day procap for each countries divided by sex, age groups and BMI groups.
}
\description{
it computes the intake per capita for each BMI group, excluding intake for pregnancy or lactation
height and BMI data of NCDrisc, and phyiscal inactivity levels based on Halal et al.
}
