% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInvestmentHistorical.R
\name{calcInvestmentHistorical}
\alias{calcInvestmentHistorical}
\title{Calculate REMIND investment variables from IEA World Energy Investment Outlook (2024)}
\usage{
calcInvestmentHistorical()
}
\description{
Calculate REMIND investment variables from IEA World Energy Investment Outlook (2024)
}
\author{
Nicolas Bauer, Falk Benke
}
