% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIoRemind.R
\name{calcIoRemind}
\alias{calcIoRemind}
\title{Calc Input Output for REMIND}
\usage{
calcIoRemind(subtype)
}
\arguments{
\item{subtype}{either "input", "output" or "trade"}
}
\description{
A wrapper around calcIO used to generate REMIND input data, created to apply
corrective steps in a cache-conscious manner by avoiding additional subtypes
in calcIO.
}
\author{
Falk Benke
}
