% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPCCClimateRegions.R
\name{calcIPCCClimateRegions}
\alias{calcIPCCClimateRegions}
\title{calcClimateRegionsIPCC}
\usage{
calcIPCCClimateRegions(
  landusetypes = "all",
  cellular = FALSE,
  yearly = FALSE,
  convert = TRUE
)
}
\arguments{
\item{landusetypes}{all or only one (to save computation memory)}

\item{cellular}{FALSE for country level, TRUE for cells}

\item{yearly}{FALSE for normal magpie 5 year time spans, TRUE for yearly}

\item{convert}{fills missing countries for country level aggregation with warm temperate moist (mostly small island nations)}
}
\value{
Country or cellular magpie object with matrix of fraction of each climate region by country or cell
}
\description{
calculates IPCC Climate Regions (IPCC2006 ch.4.3) based on t, ppt, pet from LPJml. elevation dimension not included for tropical montane class
}
\examples{

\dontrun{ 
calcOutput("IPCCClimateRegions")
}
}
\seealso{
\code{\link{readLPJml_rev21}}
}
\author{
David Chen
}
