% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPCCfracLeach.R
\name{calcIPCCfracLeach}
\alias{calcIPCCfracLeach}
\title{calcIPCCfracLeach}
\usage{
calcIPCCfracLeach(cellular = TRUE)
}
\arguments{
\item{cellular}{if true, returned on cell level}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the leaching rate FRAC_LEACH as defined by the IPCC Guidelines for National Greenhouse Gas Inventories 2006. We use the approach used by Canada, see
Velthof, Gerardus Lambertus, and J. Mosquera Losada. 2011. Calculation of Nitrous Oxide Emission from Agriculture in the Netherlands: Update of Emission Factors and Leaching Fraction. Alterra. http://library.wur.nl/WebQuery/wurpubs/406284.
}
\examples{

\dontrun{ 
a<-calcOutput("IPCCfracLeach",cellular=FALSE)
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Benjamin Leon Bodirsky
}
