% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIRENA.R
\name{calcIRENA}
\alias{calcIRENA}
\title{Calculate REMIND variables from historical IRENA capacities.}
\usage{
calcIRENA()
}
\description{
Calculate REMIND variables from historical IRENA capacities.
}
\author{
Falk Benke
}
