% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationInvCosts.R
\name{calcIrrigationInvCosts}
\alias{calcIrrigationInvCosts}
\title{Calculation of IrrigationInvCosts}
\usage{
calcIrrigationInvCosts()
}
\value{
MAgPIE object
}
\description{
This function calculates values for each country until the year 2050. A
linear convergence is implemented, which means that every country will have
the value of Germany (1995) in 2050.
}
\examples{

\dontrun{ 
calcOutput("IrrigationInvCosts")


}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readWBirrigation}},
\code{\link{convertWBirrigation}}
}
\author{
Nele Steinmetz
}
