% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcJRC_IDEES.R
\name{calcJRC_IDEES}
\alias{calcJRC_IDEES}
\title{Calculate selected REMIND energy and emission variables from historical JRC IDEES values}
\usage{
calcJRC_IDEES(subtype)
}
\arguments{
\item{subtype}{one of
\itemize{
\item \code{'Industry'}: calculate REMIND Industry variables
\item \code{'Transport'}: calculate REMIND Transport variables
\item \code{'ResCom'}: calculate REMIND Residential and Commercial variables
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate selected REMIND energy and emission variables from historical JRC IDEES values
}
\author{
Falk Benke
}
