% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKLWdamage.R
\name{calcKLWdamage}
\alias{calcKLWdamage}
\title{write KLW damage parameters (from Kotz et al. 2024) into input data
they are country-specific and should not be aggregated to the regional level at all}
\usage{
calcKLWdamage(subtype)
}
\arguments{
\item{subtype}{"beta1", "beta2", "maxGMT"}
}
\value{
MAgPIE object of damage parameters for KLW damage function on country level and for 1000 boostrapping samples
}
\description{
write KLW damage parameters (from Kotz et al. 2024) into input data
they are country-specific and should not be aggregated to the regional level at all
}
\author{
Franziska Piontek
}
