% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandfillDumpCH4.R
\name{calcLandfillDumpCH4}
\alias{calcLandfillDumpCH4}
\title{calcLandfillDumpCH4}
\usage{
calcLandfillDumpCH4()
}
\value{
Magpie object of CH4 from global landfills and dumps
}
\description{
Mt CH4 from landfills and dumps (potential for biogas), based on What a waste projections and emissions
}
\examples{

\dontrun{ a <- calcOutput(type="LandfillDumpCH4")
}
}
\author{
David Chen
}
