% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduse.R
\name{calcLanduse}
\alias{calcLanduse}
\title{calcLanduse}
\usage{
calcLanduse(cellular = FALSE, landuse_types = "magpie",
  selectyears = "past")
}
\arguments{
\item{cellular}{cellular (TRUE) or country-level/regional (FALSE) data}

\item{landuse_types}{magpie (5 types) or LUH2v2, "ini_old" and "ini_new" for new and old implementation in calcLanduseInitialisation, carbon for CarbonBudget types (crop, past, urban, natveg)}

\item{selectyears}{defaults to past}
}
\value{
List of magpie object with results on country or cellular level, weight, unit and description.
}
\description{
Calculates the cellular MAgPIE landuse area based on LUH2v2 or LanduseInitialisation data.
}
\examples{

\dontrun{ 
calcOutput("Landuse")
}
}
\author{
Kristine Karstens
}
