% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseChange.R
\name{calcLanduseChange}
\alias{calcLanduseChange}
\title{calcLanduseChange}
\usage{
calcLanduseChange(cellular = FALSE, landuse_types = "magpie")
}
\arguments{
\item{cellular}{cellular (TRUE) or country-level/regional (FALSE) data}

\item{landuse_types}{magpie (5 types) or LUH2v2, "ini_old" and "ini_new" for new and old implementation in calcLanduseInitialisation, carbon for CarbonBudget types (crop, past, urban, natveg)}
}
\value{
List of magpie object with results on country or cellular level, weight, unit and description.
}
\description{
Calculates the cellular MAgPIE landuse change area based on LUH2v2 or LanduseInitialisation data.
}
\examples{

\dontrun{ 
calcOutput("LanduseChange")
}
}
\author{
Kristine Karstens
}
