% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseInitialisation.R
\name{calcLanduseInitialisation}
\alias{calcLanduseInitialisation}
\title{calcLanduseInitialisation}
\usage{
calcLanduseInitialisation(cellular = FALSE, land = "new",
  selectyears = "past", input_magpie = FALSE)
}
\arguments{
\item{cellular}{cellular (TRUE) or country-level/regional (FALSE) data? For country-level vs regional data, remember to set "aggregate" to false.}

\item{land}{new default land="fao", if land="new", the set land seperates primary and secondary forests, "old" uses the old land set (including only forest)}

\item{selectyears}{default on "past"}

\item{input_magpie}{uses just 1995 values and apply area fix (set cell with zero area to minimal value to not distrube aggregating to clusters)}
}
\value{
List of magpie object with results on country or cellular level, weight on cellular level, unit and description.
}
\description{
Calculates the cellular MAgPIE landuse initialisation area. Data from FAO on forestry is used to split the secondary forest pool of the LU2v1 dataset into forestry and secd_forest.
}
\examples{

\dontrun{ 
calcOutput("LanduseInitialisation")
}
}
\author{
Benjamin Leon Bodirsky
}
