% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseIntensity.R
\name{calcLanduseIntensity}
\alias{calcLanduseIntensity}
\title{calcLanduseIntensity}
\usage{
calcLanduseIntensity(sectoral = "kcr", rescale = TRUE)
}
\arguments{
\item{sectoral}{"kcr" (default) for MAgPIE crop items and "lpj" for LPJmL crop items, "pasture" for pasture}

\item{rescale}{TRUE (default), if Tau should be rescaled in a way, that 2010 is always 1}
}
\value{
Total tau data and corresonding weights as a list of two MAgPIE
objects
}
\description{
This function prepares total tau values for use. As the source data already
provides all required information this function purely removes unrequired
data and moves the xref values to the weighting object which is required for
aggregation.
}
\examples{

\dontrun{ 
calcOutput("LanduseIntensity")

}

}
\seealso{
\code{\link{calcTauTotal}}, \code{\link{readTau}},
\code{\link{convertTau}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
