% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLDVShares.R
\name{calcLDVShares}
\alias{calcLDVShares}
\title{Calculate LDV Shares using EDGE-Transport}
\usage{
calcLDVShares()
}
\description{
Calculate LDV Shares using EDGE-Transport
}
\author{
Johanna Hoppe, Falk Benke
}
