% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockGridded.R
\name{calcLivestockGridded}
\alias{calcLivestockGridded}
\title{calcLivestockGridded}
\usage{
calcLivestockGridded()
}
\value{
List of magpie objects with results on cellular level, weights on cellular level, unit and description.
}
\description{
Distributes crop, pasture and livestock production in space to 0.5 degree
}
\examples{

\dontrun{ 
calcOutput("calcLivestockGridded")
}

}
\author{
Kristine Karstens
}
