% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockProductivity.R
\name{calcLivestockProductivity}
\alias{calcLivestockProductivity}
\title{Calculate Yields for Livestock}
\usage{
calcLivestockProductivity(future = TRUE)
}
\arguments{
\item{future}{if TRUE calculates Constant future and linear trends based on SSP Expert guesses}
}
\value{
MAgPIE-FEED data for livestock-yields and corresonding weights as a
list of two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for livestock-yields calculated in the regression
for feed (calcRegressionFEED).. No changes to the content have been done,
besides renaming and providing weights.
}
\examples{

\dontrun{ 
calcOutput("LivestockProductivity")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcRegressionFEED}},
\code{\link{readSource}}
}
\author{
Isabelle Weindl, Lavinia Baumstark, Stephen Wirth
}
