% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmlCarbon.R
\name{calcLPJmlCarbon}
\alias{calcLPJmlCarbon}
\title{calcLPJmlCarbon}
\usage{
calcLPJmlCarbon(climatetype = "historical", landtype = "nat_veg",
  subtype = "soilc_layer", selectyears = "past")
}
\arguments{
\item{climatetype}{Switch between different climate scenarios ("historical")}

\item{landtype}{Switch between different land use types ("lu_maize_*", "lu_grass", "nat_veg")}

\item{subtype}{Switch between diffrent input ("soil_layer", "mrh_*", "lit*", "vegc") and also processed data (soilc_0-30, rh_0-30)}

\item{selectyears}{defaults to past}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle LPJmL data for carbon module
}
\examples{

\dontrun{ 
calcOutput("LPJmlCarbon", landtype="nat_veg", subtype="soilc_layer", aggregate=FALSE)
}
}
\seealso{
\code{\link{readLPJmlCarbon}}
}
\author{
Kristine Karstens
}
