% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmlLandsurface.R
\name{calcLPJmlLandsurface}
\alias{calcLPJmlLandsurface}
\title{calcLPJmlLandsurface}
\usage{
calcLPJmlLandsurface()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Calculates LPJmL land surface area
}
\examples{

\dontrun{ 
calcOutput("LPJmlLandsurface", aggregate=FALSE)
}
}
\seealso{
\code{\link{readLPJml_rev21}}
}
\author{
Kristine Karstens
}
