% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH2MAgPIE.R
\name{calcLUH2MAgPIE}
\alias{calcLUH2MAgPIE}
\title{calcLUH2MAgPIE}
\usage{
calcLUH2MAgPIE(share = "total", bioenergy = "ignore",
  selectyears = "past")
}
\arguments{
\item{share}{total (for total numbers), LUHofMAG (for share of LUH within kcr types), MAGofLUH (for share of kcr within LUH types)}

\item{bioenergy}{ignore (for 0 for share and totals), fix (for fixing betr and begr shares in LUHofMAG to 1 for c3per and c4per)}

\item{selectyears}{defaults to past}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates out of LUH2FAO and FAO2MAgPIE mappings the real aggregation of LUH croptypes to MAgPIE croptypes
}
\examples{

\dontrun{ 
calcOutput("LUH2MAgPIE")
}
}
\author{
Kristine Karstens
}
