% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH2v1.R
\name{calcLUH2v1}
\alias{calcLUH2v1}
\title{calcLUH2v1}
\usage{
calcLUH2v1(landuse_types = "magpie", irrigation = FALSE,
  cellular = FALSE, selectyears = "past")
}
\arguments{
\item{landuse_types}{magpie: magpie landuse classes. LUH2v1: original landuse classes}

\item{irrigation}{if true, the areas are returned sperated by irrigated and rainfed. Irrigation includes flooded area. possible inconsistencies, to be checked.}

\item{cellular}{if true, dataset is returned on 0.5 degree resolution}

\item{selectyears}{defaults to past}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Integrates the LUH2v1 landuse-dataset.
}
\examples{

\dontrun{ 
calcOutput("LUH2v1")
}
}
\seealso{
\code{\link{calcLanduseInitialisation}}
}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder
}
