% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMacBaseLandUse.R
\name{calcMacBaseLandUse}
\alias{calcMacBaseLandUse}
\title{calcMacBase
calculate MacBase}
\usage{
calcMacBaseLandUse(subtype)
}
\arguments{
\item{subtype}{Source of subset of emissions}
}
\value{
magpie object
}
\description{
calcMacBase
calculate MacBase
}
\examples{

\dontrun{ a <- calcOutput(type="MacBaseLandUse")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
David Klein, Julian Oeser
}
