% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMACCsCO2.R
\name{calcMACCsCO2}
\alias{calcMACCsCO2}
\title{Read in abatement potential for CO2 land-use change derived from MAgPIE}
\usage{
calcMACCsCO2()
}
\value{
MAgPIE object
}
\description{
Rrange of possible abatement between maximum and minimum emission level in a year
}
\examples{

\dontrun{ 
calcOutput("MACCsCO2")

}
}
\author{
David Klein
}
