% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMacroInvestments.R
\name{calcMacroInvestments}
\alias{calcMacroInvestments}
\title{Calculate 2005 macroeconomic capital investments}
\usage{
calcMacroInvestments()
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Compute macroeconomic capital investments based on investments shares from the PWT and GDP scenarios from
mrdrivers. The final investments are the product of the two.
}
\seealso{
\itemize{
\item See the vignette \code{vignette("scenarios", "mrdrivers")} for information on the GDP scenarios.
\item \code{\link[=readPWT]{readPWT()}} for information on the PWT version used.
}
}
