% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMacroWACC.R
\name{calcMacroWACC}
\alias{calcMacroWACC}
\title{Calculate the weighted average cost of capital (WACC) markups for REMIND regions.
These values are applied to macroeconomic investments.}
\usage{
calcMacroWACC()
}
\description{
Provides REMIND data for 25_WACC Module
}
\author{
Diamantis Koutsandreas
}
