% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcManagementIntensity.R
\name{calcManagementIntensity}
\alias{calcManagementIntensity}
\title{Calculate total tau}
\usage{
calcManagementIntensity()
}
\value{
Total tau data and corresonding weights as a list of two MAgPIE
objects
}
\description{
This function prepares total tau values for use. As the source data already
provides all required information this function purely removes unrequired
data and moves the xref values to the weighting object which is required for
aggregation.
}
\examples{

\dontrun{ 
calcOutput("ManagementIntensity")

}

}
\seealso{
\code{\link{calcTauTotal}}, \code{\link{readTau}},
\code{\link{convertTau}}
}
\author{
Benjamin Leon Bodirsky
}
