% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcManureFuelShr.R
\name{calcManureFuelShr}
\alias{calcManureFuelShr}
\title{calcManureFuelShr}
\usage{
calcManureFuelShr(products = "magpie")
}
\arguments{
\item{products}{IPCC: IPCC products. MAgPIE: Magpie products}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates the share of Manure excreted during grazing which is collected for fuel. For the future, we assume that with the development, the fuel share reaches 0.
}
\examples{

\dontrun{ 
calcOutput("ManureFuelShr")
}

}
\seealso{
\code{\link{calcExcretion}}
}
\author{
Benjamin Leon Bodirsky
}
