% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcManureRecyclingCroplandPast.R
\name{calcManureRecyclingCroplandPast}
\alias{calcManureRecyclingCroplandPast}
\title{calcManureRecyclingCroplandPast}
\usage{
calcManureRecyclingCroplandPast(cellular = FALSE)
}
\arguments{
\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates manure recycling to cropland based on excretions, animal waste management types (and their shares per country) and emission factors for nitrogenous emissions in livestock confinements
}
\examples{

\dontrun{ 
calcOutput("calcManureRecyclingCroplandPast")
}
}
\seealso{
\code{\link{calcExcretion}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
