% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMilk.R
\name{calcMilk}
\alias{calcMilk}
\title{calcMilk}
\usage{
calcMilk(isocells = FALSE)
}
\arguments{
\item{isocells}{if TRUE iso cells (JPN.2346) instead of global cell (GLO.1) are used}
}
\value{
List of magpie object
}
\description{
test for different cell namings
}
\examples{

\dontrun{ 
calcOutput("calcDump")
}
}
\author{
Kristine Karstens
}
