% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticropping.R
\name{calcMulticropping}
\alias{calcMulticropping}
\title{calcMulticropping}
\usage{
calcMulticropping(selectyears = "time")
}
\arguments{
\item{selectyears}{"time": the full period in 5 year timesteps. "past": only past. "past_all" past with all years. otherwhise, any vector of years.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the ratio between area harvested and physical cropland area. Can be larger or smaller, depending on fallow land and double cropping.
}
\examples{

\dontrun{ 
calcOutput("")
}

}
\seealso{
\code{\link{calcFAOLand}},
\code{\link{calcCroparea}}
}
\author{
Benjamin Leon Bodirsky
}
