% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNDCTargets.R
\name{calcNDCTargets}
\alias{calcNDCTargets}
\title{Calculate NDC Emissions Targets}
\usage{
calcNDCTargets(sources, subtype)
}
\arguments{
\item{sources}{database source, must be 'PBL_NDC'}

\item{subtype}{must be one of
\itemize{
\item 'EmiTargetAbs': absolute emissions targets in MtCO2eq/yr
\item 'Ghgshare': share of emissions covered under NDC target per REMIND region
}}
}
\description{
This function calculates the emissions targets for the NDC scenarios applied in the REMIND module 45_carbonprice realization NDC.
It contains the following steps:
The parameters calculated in 3.) and 4.) are further used in the NDC realization to calculate the region-wide NDC emissions targets
in terms of total GHG emissions excl. bunkers and excl. LULUCF sectors.
}
\author{
Rahel Mandaroux, Falk Benke, Felix Schreyer
}
