% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenBudgetLivestock.R
\name{calcNitrogenBudgetLivestock}
\alias{calcNitrogenBudgetLivestock}
\title{calcNitrogenBudgetLivestock}
\usage{
calcNitrogenBudgetLivestock()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Nitrogen Budgets for the livestock sector on country levels.
}
\examples{

\dontrun{ 
calcOutput("NitrogenBudgetLivestock")
}
}
\author{
Benjamin Leon Bodirsky
}
