% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenBudgetNonagland.R
\name{calcNitrogenBudgetNonagland}
\alias{calcNitrogenBudgetNonagland}
\title{calcNitrogenBudgetNonagland}
\usage{
calcNitrogenBudgetNonagland(deposition = "CEDS", max_nue = 0.95)
}
\arguments{
\item{deposition}{if FALSE, deposition is not accounted for in the distribution. Use FALSE to avoid circularities in calcNitrogenBudget}

\item{max_nue}{NULL or a numeric value. if numeric, an additional N balanceflow is included that takes care that the nitrogen use efficiency does not exceed the numeric value in balanceflow.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Nitrogen Budgets for Non-agricultural land on country levels.
}
\examples{

\dontrun{ 
calcOutput("NitrogenBudgetNonagland")
}
}
\author{
Benjamin Leon Bodirsky
}
