% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenBudgetPasture.R
\name{calcNitrogenBudgetPasture}
\alias{calcNitrogenBudgetPasture}
\title{calcNitrogenBudgetPasture}
\usage{
calcNitrogenBudgetPasture(cellular = FALSE, include_fertilizer = TRUE,
  deposition = "CEDS", max_nue = 0.9)
}
\arguments{
\item{cellular}{cellular disaggreagation or national values}

\item{include_fertilizer}{including fertilizer in budget. Use FALSE to avoid circularities in specific cases}

\item{deposition}{if FALSE, deposition is not accounted for in the distribution. Use FALSE to avoid circularities in calcNitrogenBudget}

\item{max_nue}{NULL or a numeric value. if numeric, an additional N balanceflow is included that takes care that the nitrogen use efficiency does not exceed the numeric value in balanceflow.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates Nitrogen Budgets for Pasture soils on country levels.
}
\examples{

\dontrun{ 
calcOutput("NitrogenBudgetPasture")
}
}
\author{
Benjamin Leon Bodirsky
}
