% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationFreeliving.R
\name{calcNitrogenFixationFreeliving}
\alias{calcNitrogenFixationFreeliving}
\title{calcNitrogenFixationFreeliving}
\usage{
calcNitrogenFixationFreeliving()
}
\value{
List of magpie objects with results on global level, empty weight, unit and description.
}
\description{
calculates fixation rates from freeliving bacteria per area
}
\examples{

\dontrun{ 
calcOutput("NitrogenFixationFreeliving")
}

}
\seealso{
\code{\link{calcNitrogenFixationPast}}
\code{\link{readHerridge}}
}
\author{
Benjamin Leon Bodirsky
}
