% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationRateNatural.R
\name{calcNitrogenFixationRateNatural}
\alias{calcNitrogenFixationRateNatural}
\title{calcNitrogenFixationRateNatural}
\usage{
calcNitrogenFixationRateNatural()
}
\value{
List of magpie objects with results on global level, empty weight, unit and description.
}
\description{
calculates fixation rates from natural ecosystems based on evapostranspiration
}
\examples{

\dontrun{ 
calcOutput("NitrogenFixationRateNatural")
}

}
\seealso{
\code{\link{calcNitrogenFixationPast}}
\code{\link{readHerridge}}
}
\author{
Benjamin Leon Bodirsky
}
