% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenSurplusByCrop.R
\name{calcNitrogenSurplusByCrop}
\alias{calcNitrogenSurplusByCrop}
\title{calcNitrogenSurplusByCrop}
\usage{
calcNitrogenSurplusByCrop(indicator = "total",
  deposition = "Nsurplus2", cellular = FALSE)
}
\arguments{
\item{indicator}{total: estimates the inputs per total crop production; by_area estimates the inputs per area harvested}

\item{deposition}{if FALSE, deposition is not accounted for in the distribution. Use FALSE to avoid circularities in calcNitrogenBudget}

\item{cellular}{cellular disaggreagation or national values}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the crop-specific nitrogen losses and the balanceflow for countries with unrealistically high nitrogen uptake efficiencies
}
\examples{

\dontrun{ 
calcOutput("NitrogenSurplusByCrop")
}
}
\seealso{
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
