% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenWithdrawalByCrop.R
\name{calcNitrogenWithdrawalByCrop}
\alias{calcNitrogenWithdrawalByCrop}
\title{calcNitrogenWithdrawalByCrop}
\usage{
calcNitrogenWithdrawalByCrop(indicator = "total", cellular = FALSE,
  irrigation = FALSE)
}
\arguments{
\item{indicator}{total: estimates the inputs per total crop production; by_area estimates the inputs per area harvested}

\item{cellular}{cellular disaggreagation or national values}

\item{irrigation}{FALSE for the sum of irrigated and rainfed, FALSE for seperated categories, 'rainfed' or 'irrigated for single categories}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the crop-specific withdrawals of nutrients from soils
}
\examples{

\dontrun{ 
calcOutput("NitrogenWithdrawalByCrop")
}
}
\seealso{
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
