% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcnonEnergyIndFE.R
\name{calcnonEnergyIndFE}
\alias{calcnonEnergyIndFE}
\title{Final energy demand for feedstocks (non-energy use)}
\usage{
calcnonEnergyIndFE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Final energy demand for feedstocks (non-energy use)
}
\seealso{
\code{\link[madrat:calcOutput]{calcOutput()}}.
}
\author{
Renato Rodrigues
}
