% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNutrientBudgetFoodProcessing.R
\name{calcNutrientBudgetFoodProcessing}
\alias{calcNutrientBudgetFoodProcessing}
\title{calcNutrientBudgetFoodProcessing}
\usage{
calcNutrientBudgetFoodProcessing()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Fooduse before processing, after processing and food processing loss
}
\examples{

\dontrun{ 
calcOutput("NutrientBudgetFoodProcessing")
}

}
\seealso{
\code{\link{calcFAOmassbalance}}
}
\author{
Benjamin Leon Bodirsky
}
