% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNutrientWasteBudget.R
\name{calcNutrientWasteBudget}
\alias{calcNutrientWasteBudget}
\title{calcNutrientWasteBudget}
\usage{
calcNutrientWasteBudget(nutrient = "nr")
}
\arguments{
\item{nutrient}{The nutrient in which the results shall be reported.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Estimate waste flows of nutrients from differents sources, including Household waste, slaughterwaste and Processingwaste
}
\examples{

\dontrun{ 
calcOutput("NutrientWasteBudget")
}

}
\author{
Benjamin Leon Bodirsky
}
