% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOtherFossilInElectricity.R
\name{calcOtherFossilInElectricity}
\alias{calcOtherFossilInElectricity}
\title{Calculate projected electricity from waste and other fossils
using energy demands from IEA Energy Balances.}
\usage{
calcOtherFossilInElectricity()
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}.
}
\description{
This is used in remind2 reporting as input data to calculate additional
capacity and secondary energy variables.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Falk Benke
}
