% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureYield.R
\name{calcPastureYield}
\alias{calcPastureYield}
\title{Calculate Pasture Yields}
\usage{
calcPastureYield()
}
\value{
Pasture yields and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides Pasture yields defined as ratio of grazed biomass to grazed area
}
\examples{

\dontrun{ 
calcOutput("PastureYield")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}},
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
