% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPE.R
\name{calcPE}
\alias{calcPE}
\title{Computes Primary Energy variables from IEA Energy Balances}
\usage{
calcPE(ieaVersion = "default")
}
\arguments{
\item{ieaVersion}{Release version of IEA data, either 'default' (vetted and used in REMIND)
or 'latest'.}
}
\description{
Computes Primary Energy variables from IEA Energy Balances
}
