% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPHI.R
\name{calcPHI}
\alias{calcPHI}
\title{Calculate capital costs for energy technologies}
\usage{
calcPHI()
}
\description{
Provides REMIND data for realization Investment_Inefficiencies in 01_macro Module
}
\examples{
\dontrun{
calcOutput("IMF_PHI")
}

}
\author{
Diamantis Koutsandreas
}
