% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticsEoL.R
\name{calcPlasticsEoL}
\alias{calcPlasticsEoL}
\title{calculates projections for the end of life fate of plastic waste
in particular, calculates the share that is incinerated}
\usage{
calcPlasticsEoL()
}
\description{
calculates projections for the end of life fate of plastic waste
in particular, calculates the share that is incinerated
}
\author{
Falk Benke, Simón Moreno Leiva
}
