% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcpm_delta_kap_industry.R
\name{calcpm_delta_kap_industry}
\alias{calcpm_delta_kap_industry}
\title{Calculate depreciation rate of industry energy-efficiency capital (EEK).}
\usage{
calcpm_delta_kap_industry()
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{unit}, and
\code{description}.
}
\description{
Calculate depreciation rate of industry energy-efficiency capital (EEK).
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
