% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationPast.R
\name{calcPopulationPast}
\alias{calcPopulationPast}
\title{calcPopulationPast}
\usage{
calcPopulationPast(PopulationPast = "WDI_completed")
}
\arguments{
\item{PopulationPast}{population past data source}
}
\value{
Population in millions.
}
\description{
Calculates a time series of Population. Different sources are available and
can be selected in the moinput config(getConfig()$calc$PopulationPast):
\itemize{ 
  \item \code{WDI}: Source: Worldbank. Taiwan was estimated as the
        difference between all countries and the global total.
  \item \code{UN_PopDiv}: UN Population Division data. Taiwan is estimated
        from "other, non-specified areas". Missing countries have their 
        values set to zero.
}
}
\seealso{
\code{\link{convertWDI}},\code{\link{calcGDPpppPast}}
}
\author{
Lavinia Baumstark, Benjamin Bodirsky
}
