% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialGeologicalStorage.R
\name{calcPotentialGeologicalStorage}
\alias{calcPotentialGeologicalStorage}
\title{Calculate geological storage potential}
\usage{
calcPotentialGeologicalStorage()
}
\value{
geological storage potential data MAgPIE object
}
\description{
Provides geological storage potential
}
\examples{

\dontrun{
calcOutput("PotentialGeologicalStorage")
}

}
\author{
David Klein
}
