% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialGeothermal.R
\name{calcPotentialGeothermal}
\alias{calcPotentialGeothermal}
\title{Calculate Geothermal potential}
\usage{
calcPotentialGeothermal()
}
\value{
geothermal potential data MAgPIE object
}
\description{
Provides geothermal potential data
}
\examples{

\dontrun{ 
calcOutput("PotentialGeothermal")

}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Renato Rodrigues
}
