% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialWeathering.R
\name{calcPotentialWeathering}
\alias{calcPotentialWeathering}
\title{Calculate hydro potential}
\usage{
calcPotentialWeathering()
}
\value{
weathering potential data and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides weathering potential data
}
\examples{
\dontrun{
calcOutput("PotentialWeathering")
}

}
\author{
Lavinia Baumstark
}
