% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialWind.R
\name{calcPotentialWind}
\alias{calcPotentialWind}
\title{Calculate hydro potential}
\usage{
calcPotentialWind()
}
\value{
wind onshore potential data and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides wind onshore potential data
}
\examples{

\dontrun{ 
calcOutput("PotentialWind")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNREL}},
\code{\link{convertNREL}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
