% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialWindOn.R
\name{calcPotentialWindOn}
\alias{calcPotentialWindOn}
\title{Calculate wind onshore potential}
\usage{
calcPotentialWindOn()
}
\value{
wind onshore potential data and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides wind onshore potential data
}
\examples{

\dontrun{
calcOutput("PotentialWindOn")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNREL}},
\code{\link{convertNREL}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
