% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrecipitation.R
\name{calcPrecipitation}
\alias{calcPrecipitation}
\title{calcTemperature}
\usage{
calcPrecipitation(landusetypes = "all", months = FALSE,
  convert = TRUE)
}
\arguments{
\item{landusetypes}{all or only one (to save computation memory)}

\item{months}{FALSE for yearly total, TRUE for monthly values}

\item{convert}{FALSE for raw values, TRUE add temperature of 1000mm for countries without observations or land mass.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates average monthly or total yearly ppt (mm) on different landuse types
}
\examples{

\dontrun{ 
calcOutput("Precipitation")
}
}
\seealso{
\code{\link{readLPJml_rev21}}
}
\author{
David Chen
}
