% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPriceAgriculture.R
\name{calcPriceAgriculture}
\alias{calcPriceAgriculture}
\title{calcPriceAgriculture}
\usage{
calcPriceAgriculture(datasource = "IMPACT3.2.2World_Price")
}
\arguments{
\item{datasource}{Options of the source of data:  \code{IMPACT3.2.2World_Price}
, \code{FAO}, \code{FAOp} and \code{WBGEM}.}
}
\value{
List with a magpie object with commodity prices on global level.
}
\description{
provides global prices from the IMPACT model projections, World Bank observations, and FAO 
obersvations for MAgPIE commodities in $/tDM
}
\examples{

\dontrun{ 
calcOutput("PriceAgriculture", datasource="IMPACT3.2.2World_Price", aggregate=FALSE)
calcOutput("PriceAgriculture", datasource="FAO")
calcOutput("PriceAgriculture",datasource = "WBGEM",aggregate =  FALSE)
}

}
\seealso{
\code{\link{readIMPACT3.2.2World_Price}},
\code{\link{calcWBGEM}},
\code{\link{readWBGEM}}
}
\author{
Mishko Stevanovic, Xiaoxi Wang
}
