% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdShares.R
\name{calcProdShares}
\alias{calcProdShares}
\title{Manufacture production shares for spc and wind}
\usage{
calcProdShares()
}
\value{
magpie object with REMIND-aggregated region
}
\description{
Shares of world manufacture for spv modules and wind turbines for 2018 and 2019
}
\author{
Aman Malik
}
