% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdSysRatioPast.R
\name{calcProdSysRatioPast}
\alias{calcProdSysRatioPast}
\title{Calculate historical distribution of livestock production across different systems based on output of MAgPIE_FEED model}
\usage{
calcProdSysRatioPast()
}
\value{
Historical distribution of livestock production across different systems and corresonding weights as a list of two MAgPIE
objects
}
\description{
Calculate historical distribution of livestock production across different systems based on output of MAgPIE_FEED model
}
\examples{

\dontrun{ 
calcOutput("ProdSysRatioPast")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readFeedModel}}
}
\author{
Isabelle Weindl
}
