% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProduction.R
\name{calcProduction}
\alias{calcProduction}
\title{calcProduction}
\usage{
calcProduction(products = "kcr", cellular = FALSE, calibrated = TRUE,
  attributes = "all", irrigation = FALSE)
}
\arguments{
\item{products}{setname of products ("kcr", "kli", "pasture")}

\item{cellular}{if TRUE production is calculate on cellular level}

\item{calibrated}{if FALSE, lpj yields will be used uncalibrated, if true, calibrated on FAP production on country level}

\item{attributes}{"All" for all crop attributes, or specify e.g. DM (dry matter), Nr (nitrogen) for memory reduction}

\item{irrigation}{if TRUE, additional information on irrigated and rainfed production is provided}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Distributes crop, pasture and livestock production in space to 0.5 degree
}
\examples{

\dontrun{ 
calcOutput("Production")
}

}
\seealso{
\code{\link{calcLanduseInitialisation}},
\code{\link{calcCroparea}}
}
\author{
Benjamin Leon Bodirsky
}
