% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProjectPipelines.R
\name{calcProjectPipelines}
\alias{calcProjectPipelines}
\title{calc Project Pipelines}
\usage{
calcProjectPipelines(subtype)
}
\arguments{
\item{subtype}{choose technology, either `CCS`, `Nuclear`}
}
\description{
Calculate the expected near-term deployment of technologies
}
\author{
Pascal Weigmann
}
