% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPYieldSlope.R
\name{calcPYieldSlope}
\alias{calcPYieldSlope}
\title{calcPYieldSlope}
\usage{
calcPYieldSlope()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
provides slope for calculating pasture intensification
}
\examples{

\dontrun{ 
calcOutput("PYieldSlope")
}
}
\seealso{
\code{\link{readPYieldCoeff}}
}
\author{
Isabelle Weindl
}
