% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRecoveryRate.R
\name{calcRecoveryRate}
\alias{calcRecoveryRate}
\title{Calculate RecoveryRate}
\usage{
calcRecoveryRate()
}
\value{
MAgPIE-FEED data for Recovery Rate and corresonding weights as a
list of two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for Recovery Rate. Aggregated to MAGPIE-sectors.
Usually no weight needed as the data will beused in MAgPIE-FEED model
country based.
}
\examples{

\dontrun{ 
calcOutput("RecoveryRate")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
