% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRegressionFEED.R
\name{calcRegressionFEED}
\alias{calcRegressionFEED}
\title{Calculate FEEDRequirement, NutrientDensity and YieldsLive based on
regression model}
\usage{
calcRegressionFEED()
}
\value{
Productivity, feed requirements and nutrient requirements as well as corresonding weights as a list of two MAgPIE
objects
}
\description{
This intermediate-function prepares FEEDRequirement, NutrientDensity and
YieldsLive for calc*-functions. As the source data already provides all
required information this function purely selects the needed data.
}
\examples{

\dontrun{ 
calcOutput("RegressionFEED")

}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readWirsenius_FEED}}
}
\author{
Lavinia Baumstark, Isabelle Weindl
}
