% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRenShareTargets.R
\name{calcRenShareTargets}
\alias{calcRenShareTargets}
\title{Calculate Renewable Energy Share Targets}
\usage{
calcRenShareTargets(sources)
}
\arguments{
\item{sources}{"NewClimate"}
}
\description{
This function calculates the renewable energy share targets by aggregating country-level targets to targets of REMIND regions. It calculates
the following types of share targets: 1) share of renewables in electricity, 2) share of non-biomass renewables in electricty,
3) share of non-fossil generation in  electricity, 4) share of renewables in total final energy.
First, targets are aggregated to one of these 4 types. Second, countries without targets are assumed to maintain their renewable share from 2020.
Third, targets are harmonized within a REMIND region and filtered based on how many countries within a REMIND region have a target. Finally, energy share targets are aggregated from country-level
to REMIND region-level using projections of electricity or final energy demand on country-level. These country-level projections are
derived from final energy trends by REMIND regions from the EDGE models as well as GDP trends by country from SSP scenarios (see toolCalcEnergyProj).
}
\author{
Felix Schreyer
}
