% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResBiomass.R
\name{calcResBiomass}
\alias{calcResBiomass}
\title{calcResBiomass}
\usage{
calcResBiomass(cellular = FALSE, plantparts = "both",
  irrigation = FALSE, attributes = "all")
}
\arguments{
\item{cellular}{If TRUE calculation and output on cellular level}

\item{plantparts}{both, ag (aboveground) or belowground (bg). Both can have memory problems for cellular outputs}

\item{irrigation}{if TRUE, distinguishes irrigated and non-irrigated crops}

\item{attributes}{in dm, wm, ge, nr, p, k}
}
\value{
MAgPIE-FEED data for ProdResAg and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for aboveground and belowground residues biomass
}
\examples{

\dontrun{ 
calcOutput("ResBiomass")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark, Isabelle Weindl, Benjamin Bodirsky
}
