% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResFieldBalancePast.R
\name{calcResFieldBalancePast}
\alias{calcResFieldBalancePast}
\title{calcResFieldBalancePast}
\usage{
calcResFieldBalancePast(cellular = FALSE, products = "sum")
}
\arguments{
\item{cellular}{If TRUE calculation and output on cellular level}

\item{products}{"sum" (default) or "kres"}
}
\value{
data
}
\description{
Calculates data for aboveground and belowground residues production with other usage
}
\examples{

\dontrun{ 
calcOutput("ResFieldBalancePast")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
