% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResFor2ndBioengery.R
\name{calcResFor2ndBioengery}
\alias{calcResFor2ndBioengery}
\title{calcResFor2ndBioengery}
\usage{
calcResFor2ndBioengery(products = "all", product_aggr = TRUE,
  add_off = FALSE, extrapolation = FALSE)
}
\arguments{
\item{products}{categorie (set) that should be reported, switch between "kres", "res_crop" (sum over all "kres"), "res_wood" and "all"}

\item{product_aggr}{boolean, if product set should be summed up}

\item{add_off}{add a column with empty supply for no residues available for 2nd gen BE}

\item{extrapolation}{Extrapolate missing years (where values are zero)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the supply potential of Crop Residues for 2nd generation bioenergy for future and different ssp scenarios
}
\examples{

\dontrun{ 
calcOutput("ResFor2ndBioengery")
}
}
\seealso{
\code{\link{calcResFor2ndBioengery}}
}
\author{
Kristine Karstens
}
