% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResiduesShare.R
\name{calcResiduesShare}
\alias{calcResiduesShare}
\title{calcResiduesShare
calculate residues share}
\usage{
calcResiduesShare()
}
\value{
magpie object
}
\description{
calcResiduesShare
calculate residues share
}
\examples{

\dontrun{ a <- calcOutput(type="ResiduesShare")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser, David Klein
}
