% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResUseMinShr.R
\name{calcResUseMinShr}
\alias{calcResUseMinShr}
\title{Calculate UseMin}
\usage{
calcResUseMinShr()
}
\value{
MAgPIE-FEED data for UseMin and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides MAgPIE-FEED data for UseMin. No changes to the content have been
done. Usually no weight needed as the data will beused in MAgPIE-FEED model
country based.
}
\examples{

\dontrun{ 
calcOutput("ResUseMinShr")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
