% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRLDCCoefficients.R
\name{calcRLDCCoefficients}
\alias{calcRLDCCoefficients}
\title{calc RLDC Coefficients}
\usage{
calcRLDCCoefficients(subtype = "LoB")
}
\arguments{
\item{subtype}{Either 'LoB' or 'Peak'}
}
\value{
magpie object of the RLDC coefficients data
}
\description{
provides RLDC coefficients values
}
\examples{

\dontrun{ 
calcOutput(type="RLDCCoefficients",subtype='LoB')
}

}
\seealso{
\code{\link{calcOutput}}

\code{\link{calcOutput}}
}
\author{
Renato Rodrigues
}
