% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calcSE}
\alias{calcSE}
\title{Calculate Secondary Energy variables
Currently this function only uses data from Ember, however other sources for SE are available and could be integrated
here using different \code{datasource} arguments.}
\usage{
calcSE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate Secondary Energy variables
Currently this function only uses data from Ember, however other sources for SE are available and could be integrated
here using different \code{datasource} arguments.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Pascal Weigmann
}
