% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSecondary_steel_limits.R
\name{calcSecondary_steel_limits}
\alias{calcSecondary_steel_limits}
\title{Calculate Secondary Steel Limits}
\usage{
calcSecondary_steel_limits()
}
\value{
A [`MAgPIE`][magclass::magclass] object.
}
\description{
Upper limit to secondary steel production due to scrap availability.
}
\author{
Michaja Pehl
}
