% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeed.R
\name{calcSeed}
\alias{calcSeed}
\title{calcSeed}
\usage{
calcSeed(cellular = FALSE, products = "kall", irrigation = FALSE,
  attributes = "all")
}
\arguments{
\item{cellular}{cellular or regional level}

\item{products}{kcr or also kall, which includes seeds for eggs and fish}

\item{irrigation}{if TRUE, distinguishes irrigated and non-irrigated crops}

\item{attributes}{in dm, wm, ge, nr, p, k}
}
\value{
List of magpie object with results and weight on country or cellular level, unit and description.
}
\description{
Calculates Seed demand
}
\examples{

\dontrun{ 
calcOutput("Seed")
}
}
\author{
Benjamin Leon Bodirsky
}
