% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeedRate.R
\name{calcSeedRate}
\alias{calcSeedRate}
\title{calcSeedRate}
\usage{
calcSeedRate()
}
\value{
List of magpie objects with results on country level, weight on country level, unit, Max.&Min. values along with description.
}
\description{
Calculates seed demand as a share of area harvested.
}
\examples{

\dontrun{ 
x <- calcOutput("SeedRate")
}
}
\seealso{
\code{\link{calcFAOmassbalance}},
\code{\link{calcCroparea}}
}
\author{
Benjamin L. Bodirsky, Abhijeet Mishra
}
