% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSelfEnergyServiceCoeff.R
\name{calcSelfEnergyServiceCoeff}
\alias{calcSelfEnergyServiceCoeff}
\title{Calculate Demand based on the output of the Demand Model}
\usage{
calcSelfEnergyServiceCoeff()
}
\value{
self energy service coefficants and corresonding weights as a list
of two MAgPIE objects
}
\description{
This function extracts the self energy service data from the source
NTNU_LCA.
}
\examples{

\dontrun{ 
calcOutput("SelfEnergyServiceCoeff")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNTNU_LCA}},
\code{\link{convertNTNU_LCA}}
}
\author{
Lavinia Baumstark
}
