% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeProduction.R
\name{calcSeProduction}
\alias{calcSeProduction}
\title{calc secondary energy production}
\usage{
calcSeProduction()
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
prepare secondary production data (e.g. electricity generation)
to use in historical constraints in the model
}
\author{
Renato Rodrigues
}
