% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSharedTarget.R
\name{calcSharedTarget}
\alias{calcSharedTarget}
\title{calc Shared Target}
\usage{
calcSharedTarget(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "FErenewablesShare", ...}
}
\value{
target data magpie object
}
\description{
provides region specific target
}
\examples{

\dontrun{
calcOutput("SharedTarget",subtype="FErenewablesShare")
}

}
\author{
Renato Rodrigues
}
