% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShareIndFE.R
\name{calcShareIndFE}
\alias{calcShareIndFE}
\title{Share of Industry Subsectors in FE Use}
\usage{
calcShareIndFE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculates industry subsector shares in final energy carrier use for the
\code{fixed_shares} realisation of the \code{industry} module.
}
\details{
For the region mapping \code{regionmapping_21_EU11.csv}, these are based on IEA
data from \code{calcOutput(type = 'FEdemand')}, for all other
region mappings on vintage data which is ultimately based on Enerdata data.
}
\note{
There is a discrepancy between the shares calculated from these two
sources, that will affect REMIND emission reporting.
}
\seealso{
\code{\link[madrat:calcOutput]{calcOutput()}}.
}
\author{
Lavinia Baumstark

Michaja Pehl
}
