% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSNUpE.R
\name{calcSNUpE}
\alias{calcSNUpE}
\title{calcSNUpE}
\usage{
calcSNUpE(max_snupe = 0.85, cellular = FALSE)
}
\arguments{
\item{max_snupe}{Maximum realistic SNUPE. All values above will be limited to this value. Only holds for past values; future scneario values can exceed this number.}

\item{cellular}{disaggregated to 0.5 degree grid}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the soil nitrogen uptake efficiency. This is the nitrogen taken up from the soil (N in crop biomass minus biological fixation minus seed N) divided by the soil N inputs (fertilizer, manure etc). For the future, SNuPE scenarios are added.
}
\examples{

\dontrun{ 
calcOutput("SNUpE")
}

}
\seealso{
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
