% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelStock.R
\name{calcSteelStock}
\alias{calcSteelStock}
\title{Calculate Steel Stock from Mueller steel stock per capita and WDI population}
\usage{
calcSteelStock()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate Steel Stock from Mueller steel stock per capita and WDI population
}
\author{
Falk Benke
}
