% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStorageFactor.R
\name{calcStorageFactor}
\alias{calcStorageFactor}
\title{calc Capacity Factor}
\usage{
calcStorageFactor()
}
\value{
magpie object of the capacity factor data
}
\description{
provides capacity factor values
}
\examples{
\dontrun{
calcOutput("StorageFactor")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Lavinia Baumstark
}
