% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSubsStationary.R
\name{calcSubsStationary}
\alias{calcSubsStationary}
\title{Calculate SubsStationary}
\usage{
calcSubsStationary(sector)
}
\arguments{
\item{sector}{subsidies in transport or in the stationary/buildings_industry sector}
}
\value{
MAgPIE object
}
\description{
Reads in the data of the source IIASA_subs_taxes, by country. Regional
aggregation is done via the respective energy quantities as weights.
}
\examples{

\dontrun{ 
calcOutput("SubsStationary")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readIIASA_subs_taxes}},
\code{\link{convertIIASA_subs_taxes}}
}
\author{
Christoph Bertram
}
