% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTaxConvergence.R
\name{calcTaxConvergence}
\alias{calcTaxConvergence}
\title{calc Tax Convergence}
\usage{
calcTaxConvergence()
}
\value{
magpie object of the tax convergence levels
}
\description{
tax convergence levels for specific regions
}
\examples{

\dontrun{ 
calcOutput("TaxConvergence")
}

}
\author{
Renato Rodrigues
}
