% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTaxesStationary.R
\name{calcTaxesStationary}
\alias{calcTaxesStationary}
\title{Calculate TaxesStationary}
\usage{
calcTaxesStationary(sector)
}
\arguments{
\item{sector}{taxes in transport or in the stationary/buildings_industry sector}
}
\value{
MAgPIE object
}
\description{
Reads in the data of the source IIASA_subs_taxes, by country. Regional
aggregation is done via the respective energy quantities as weights.
}
\examples{

\dontrun{ 
calcOutput("TaxesStationary")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readIIASA_subs_taxes}},
\code{\link{convertIIASA_subs_taxes}}
}
\author{
Christoph Bertram
}
