% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTaxLimits.R
\name{calcTaxLimits}
\alias{calcTaxLimits}
\title{calc Tax Limits}
\usage{
calcTaxLimits(subtype)
}
\arguments{
\item{subtype}{Name of the subsidy data type limit, e.g. "maxFeSubsidy" for
maximum final energy subsidy or "propFeSubsidy" for proportional cap for
final energy subsidy}
}
\value{
magpie object of the subtype tax limit
}
\description{
tax and subsidies maximum levels. The tax limits serve as a work
around to avoid excess of subsidy levels that could cause problems on the REMIND
model solution. These files should be removed or replaced once a better way
to handle this issue is introduced to the REMIND model formulation or once
better yearly and country subsidy level data is available for the primary
and final energies.
}
\examples{
\dontrun{
calcOutput("TaxLimits")
}

}
\author{
Renato Rodrigues
}
