% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTCdamage.R
\name{calcTCdamage}
\alias{calcTCdamage}
\title{write TC damage parameters into input data
they are country-specific and should not be aggregated to the regional level at all}
\usage{
calcTCdamage(subtype)
}
\arguments{
\item{subtype}{"const", "tasK"}
}
\value{
MAgPIE object of damage parameters for country level tropical cyclone damage function
}
\description{
write TC damage parameters into input data
they are country-specific and should not be aggregated to the regional level at all
}
\author{
Franziska Piontek
}
