% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTheil.R
\name{calcTheil}
\alias{calcTheil}
\title{Calculate regional Theil-T index}
\usage{
calcTheil()
}
\value{
magpie objects of unweighted contribution to Theil, weights (= country shares of regional GDP)
}
\description{
To calculate the regional Theil-T index (= correction to welfare function for a lognormal income distribution) we do
the following: (1) convert country-level Gini coefficients to Theil (2) calculate contribution to Theil-T index that
includes both between-countries and within-country inequality (see e.g. https://en.wikipedia.org/wiki/Theil_index).
The latter can then be aggregated with calcOutput().
}
\details{
NB 1: the aggregation depends on the region mapping. It is implemented such
that the regionmapping specified in getConfig()$regionmapping is used.

NB 2: the result of calcOutput('Theil', aggregate = FALSE), is NOT the country
Theil-T, but the unweighted contribution from a given country to the regional value.
}
\examples{
\dontrun{
  calcOutput("Theil")
}

}
\seealso{
\code{\link{calcOutput}} \code{\link{convertGini},\link{readGini}}
}
\author{
Bjoern Soergel
}
