% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimberDemand.R
\name{calcTimberDemand}
\alias{calcTimberDemand}
\title{calcTimberDemand}
\usage{
calcTimberDemand()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the demand of timber from FAO data (including intermediate products).
}
\examples{

\dontrun{ 
calcOutput("TimberDemand")
}
}
\seealso{
\code{\link{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
