% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTotalSupply.R
\name{calcTotalSupply}
\alias{calcTotalSupply}
\title{Calculate total supply aggregated over all commodities and assigned to
categories food, feed, other util, seed and waste}
\usage{
calcTotalSupply()
}
\value{
FAO total supply and corresponding weights as a list of two MAgPIE
objects
}
\description{
Provides aggregated total supply differentiated according to food, feed,
other util, seed and waste from the FAOSTAT Food Balance Sheets (FBS).
}
\examples{

\dontrun{ 
a <- calcTotalSupply()
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Isabelle Weindl
}
