% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrade.R
\name{calcTrade}
\alias{calcTrade}
\title{Computes Trade variables based on latest IEA data available}
\usage{
calcTrade()
}
\value{
a magclass object
}
\description{
Computes Trade variables based on latest IEA data available
}
