% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeBalanceflow.R
\name{calcTradeBalanceflow}
\alias{calcTradeBalanceflow}
\title{Calculate global under-/overproduction}
\usage{
calcTradeBalanceflow()
}
\value{
global Domestic Balanceflows as MAgPIE object
}
\description{
Calculate the difference between the global production and the global
domestic_supply. The difference is the result of imports not equaling
exports, and because storage is not considered. The calculated
DomesticBalanceflow assures that production matches domestic_supply. The
goods come from nowhere and go to nowhere. The numbers are usually decreased
linearly and become zero in 2050.
}
\examples{

\dontrun{ 
a <- calcTradeBalanceflow()
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}}
}
\author{
Ulrich Kreidenweis, Xiaoxi Wang
}
