% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeExportShr.R
\name{calcTradeExportShr}
\alias{calcTradeExportShr}
\title{Calculate export shares}
\usage{
calcTradeExportShr()
}
\value{
Export shares
}
\description{
Provides export shares of countries compared to total export. This is based
on export values from FAOSTAT. Function calculates this based on average
values of the specified years.
}
\examples{

\dontrun{ 
a <- calcTradeExportShr()
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}}
}
\author{
Ulrich Kreidenweis, Xiaoxi Wang
}
