% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeMargin.R
\name{calcTradeMargin}
\alias{calcTradeMargin}
\title{calcTradeMargin}
\usage{
calcTradeMargin(gtap_version = "GTAP7", bilateral = FALSE,
  producer_price = "FAO")
}
\arguments{
\item{gtap_version}{type of GTAP data version
\itemize{
\item \code{GTAP7}
\item \code{GTAP8}
}}

\item{bilateral}{whether bilateral trade margin should be calculated}

\item{producer_price}{which producer price should be used}
}
\value{
Trade margins as an MAgPIE object
}
\description{
calculate total value of trade margins from GTAP dataset
}
\examples{
    \dontrun{
    x <- calcTradeMargin("GTAP7")
    }
}
\author{
Xiaoxi Wang
}
