% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeTariff.R
\name{calcTradeTariff}
\alias{calcTradeTariff}
\title{calcTradeTariff}
\usage{
calcTradeTariff(gtap_version = "GTAP7", type_tariff = "total")
}
\arguments{
\item{gtap_version}{type of GTAP data version
\itemize{
\item \code{GTAP7}
\item \code{GTAP8}
}}

\item{type_tariff}{which producer price should be used
\itemize{
\item \code{type_tariff}
}}
}
\value{
Trade tariffs as an MAgPIE object
}
\description{
calculate tarde tariffs from GTAP dataset
}
\examples{
    \dontrun{
    x <- calcTradeTariff("GTAP7")
    }
}
\author{
Xiaoxi Wang
}
