% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTranspEff.R
\name{calcTranspEff}
\alias{calcTranspEff}
\title{calcTranspEff
calculate Transport efficiency}
\usage{
calcTranspEff()
}
\value{
magpie object
}
\description{
calcTranspEff
calculate Transport efficiency
}
\examples{

\dontrun{ a <- calcOutput(type="TranspEff")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser
}
