% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportGDPshare.R
\name{calcTransportGDPshare}
\alias{calcTransportGDPshare}
\title{Get transport share in GDP}
\usage{
calcTransportGDPshare()
}
\value{
magpie object of transport shares in GDP
}
\description{
provides transport share in GDP to resize edge-t transportation purchase costs
}
\examples{

\dontrun{
calcOutput("TransportGDPshare")
}

}
\author{
Renato Rodrigues
}
