% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportSubsidies.R
\name{calcTransportSubsidies}
\alias{calcTransportSubsidies}
\title{calc transport subsidies}
\usage{
calcTransportSubsidies()
}
\value{
magpie object of the transport subsidies data
}
\description{
provides transport subsidies
}
\examples{

\dontrun{
calcOutput("TransportSubsidies")
}

}
\author{
Renato Rodrigues
}
