% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrban.R
\name{calcUrban}
\alias{calcUrban}
\title{calcUrban}
\usage{
calcUrban(UrbanCalib = "past", UrbanPast = "WDI",
  UrbanFuture = "SSP", naming = "indicator_scenario")
}
\arguments{
\item{UrbanCalib}{To what should be calibrated? past or future?}

\item{UrbanPast}{Urban past data source}

\item{UrbanFuture}{Urban future data source}

\item{naming}{naming scheme}
}
\value{
Share of urban population
}
\description{
Merges time series of urban shares for the past and present.  See
\code{\link{calcUrbanPast}} for past datasets, and
\code{\link{calcUrbanFuture}} for future datasets.  The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point.  All data is calibrated either to the "past"
or the "future" dataset.  Currently, the SSP (future) and WDI (past) data
have some inconsistencies, which leads to unrealistic figures if the one is
scaled on the other.
}
\seealso{
\code{\link{convertWDI}},\code{\link{calcGDPpppPast}}
}
\author{
Antoine Levesque
}
