% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUrbanPast.R
\name{calcUrbanPast}
\alias{calcUrbanPast}
\title{calcUrbanPast}
\usage{
calcUrbanPast(UrbanPast = "WDI")
}
\arguments{
\item{UrbanPast}{Urban past data source}
}
\value{
Urban shares in population
}
\description{
Calculates a time series of urban shares
}
\seealso{
\code{\link{convertWDI}},\code{\link{calcGDPpppPast}}
}
\author{
Antoine Levesque
}
