% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteEmissions.R
\name{calcWasteEmissions}
\alias{calcWasteEmissions}
\title{calcWasteEmissions}
\usage{
calcWasteEmissions(treatment = "swds", yearly = FALSE)
}
\arguments{
\item{treatment}{type of waste treatment}

\item{yearly}{yearly or 5 year magpiesets}
}
\value{
Magpie object of emissions from waste treatments
}
\description{
Calculates GHG emissions from solid waste disposal treatments, input from calcWasteProjections,  based on IPCC 2006 SWDS waste model and reporting calculations, in million t CO2eq
}
\examples{

\dontrun{ a <- calcOutput(type="WasteEmissions")
}
}
\author{
David Chen
}
