% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteEnergyUseShares.R
\name{calcWasteEnergyUseShares}
\alias{calcWasteEnergyUseShares}
\title{Calculate waste energy use shares based on IEA World Energy Balances}
\usage{
calcWasteEnergyUseShares()
}
\description{
Calculate waste energy use shares based on IEA World Energy Balances
}
\author{
Robert Pietzcker, Falk Benke
}
