% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteProjections.R
\name{calcWasteProjections}
\alias{calcWasteProjections}
\title{calcWasteProjections}
\usage{
calcWasteProjections(pc = TRUE)
}
\arguments{
\item{pc}{MAgpIE object containing Waste data on country level, total(million tons) or per capita (kg/capita)}
}
\value{
Waste data as complete MAgPIE object on country level
}
\description{
Creates waste projections for composition and treatment
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
