% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterConsCoef.R
\name{calcWaterConsCoef}
\alias{calcWaterConsCoef}
\title{Calculate Water Consumption Coefficients}
\usage{
calcWaterConsCoef()
}
\value{
MAgPIE object on water consumption coefficients per elecricity
technology
}
\description{
This function calculates REMIND input data on water consumption coefficients
per electricity technology, using as initial information the Macknick (2011)
data per electricity technology. The source data provide most required
information but some assumptions on missing data are also made.
}
\examples{

\dontrun{ 
calcOutput("WaterConsCoef")

}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readMacknickIntensities}},
\code{\link{calcWaterWithCoef}}
}
\author{
Ioanna Mouratiadou
}
