% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterWithCoef.R
\name{calcWaterWithCoef}
\alias{calcWaterWithCoef}
\title{Calculate Water Withdrawal Coefficients}
\usage{
calcWaterWithCoef()
}
\value{
MAgPIE object on water withdrawal coefficients per elecricity
technology
}
\description{
This function calculates REMIND input data on water withdrawal coefficients
per electricity technology, using as initial information the Macknick (2011)
data per electricity technology. The source data provide most required
information but some assumptions on missing data are also made.
}
\examples{
\dontrun{
calcOutput("WaterWithCoeff")
}

}
\seealso{
\code{\link{readMacknickIntensities}},
\code{\link{calcWaterConsCoef}}
}
\author{
Ioanna Mouratiadou
}
