% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWBGEM.R
\name{calcWBGEM}
\alias{calcWBGEM}
\title{calcWBGEM}
\usage{
calcWBGEM()
}
\value{
magpie object of time series of world price of commodites
}
\description{
recover world price of commidity in terms of real 2005 USD per metric ton
}
\examples{

\dontrun{
calcOutput("WBGEM")}

}
\seealso{
\code{\link{readSource}},
\code{\link{readWBGEM}}
}
\author{
Xiaoxi Wang
}
