% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYield.R
\name{calcYield}
\alias{calcYield}
\title{calcYield}
\usage{
calcYield(physical = TRUE, attributes = "dm", irrigation = FALSE,
  cellular = FALSE)
}
\arguments{
\item{physical}{physical area or havested area}

\item{attributes}{in dm, wm, ge, nr, p, k}

\item{irrigation}{distinguish irrigation or not}

\item{cellular}{if TRUE value is calculate on cellular level}
}
\value{
MAgPIE object of yields
}
\description{
calculates the yield based on FAO data
}
\author{
Debbora Leip, Jan Philipp Dietrich
}
