% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertACCMIP.R
\name{convertACCMIP}
\alias{convertACCMIP}
\title{convertACCMIP}
\usage{
convertACCMIP(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
MAgPIE object
}
\description{
function to convert ACCMIP data to isocountry resolution
}
\examples{
\dontrun{
readSource("ACCMIP", subtype = "nhx_1850")
}
}
\author{
Roman Popov
}
