% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertADVANCE_WP2.R
\name{convertADVANCE_WP2}
\alias{convertADVANCE_WP2}
\title{Convert ADVANCE WP2 Data}
\usage{
convertADVANCE_WP2(x, subtype)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned by
\code{\link[=readADVANCE_WP2]{readADVANCE_WP2()}}.}

\item{subtype}{One of
\itemize{
\item \code{clinker-to-cement-ratio} for the clinker-to-cement ratios from figure 21
of Edelenbosch, O. \emph{Enhancing the representation of energy demand
developments in IAM models - A Modeling Guide for the Cement Industry}
(2015) \url{zotero://select/items/JP8X2QFK},
which is extended from H12 regions to country level.
}}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert ADVANCE WP2 Data
}
\seealso{
\code{\link[=readSource]{readSource()}}, \code{\link[=readADVANCE_WP2]{readADVANCE_WP2()}}
}
\author{
Michaja Pehl
}
