% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAGEB.R
\name{convertAGEB}
\alias{convertAGEB}
\title{Convert AGEB data}
\usage{
convertAGEB(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readJRC_IDEES]{readJRC_IDEES()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert AGEB data
}
\author{
Falk Benke
}
