% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAriadneDB.R
\name{convertAriadneDB}
\alias{convertAriadneDB}
\title{Convert Ariadne DB data}
\usage{
convertAriadneDB(x)
}
\arguments{
\item{x}{A [`magpie`][magclass::magclass] object returned from
[`readAriadneDB()`].}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
convert Ariadne database data
}
\author{
Felix Schreyer
}
