% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBGR.R
\name{convertBGR}
\alias{convertBGR}
\title{Converts BGR oil, gas, coal and uranium reserves data}
\usage{
convertBGR(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{data subtype. Either "oil", "gas", "coal" or "uranium".}
}
\value{
A MAgPIE object containing BGR (Federal Institute for Geosciences and
Natural Resources) country reserves disaggregated data of oil, gas, coal and uranium.
}
\description{
Converts BGR oil, gas, coal and uranium reserves data
}
\examples{
\dontrun{
a <- convertBGR(x, subtype = "oil")
}

}
\author{
Renato Rodrigues
}
