% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBijl2017.R
\name{convertBijl2017}
\alias{convertBijl2017}
\title{Convert Bijl et. al 2017 data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)}
\usage{
convertBijl2017(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{Bijl data subtype.}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Convert Bijl et. al 2017 data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)
}
\examples{

\dontrun{ a <- readSource("Bijl2017","demand",convert=FALSE)}
}
\seealso{
\code{\link{readSource}},
}
\author{
Abhijeet Mishra
}
