% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBiocharDeploymentData.R
\name{convertBiocharDeploymentData}
\alias{convertBiocharDeploymentData}
\title{Provide Biochar price path data}
\usage{
convertBiocharDeploymentData(x)
}
\arguments{
<<<<<<< HEAD
\item{x}{magpie object}
=======
\item{x}{a magpie object}
>>>>>>> master
}
\description{
Provide Biochar price path data
}
\author{
Tabea Dorndorf
}
