% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBodirsky.R
\name{convertBodirsky}
\alias{convertBodirsky}
\title{Convert Bodirsky data}
\usage{
convertBodirsky(x)
}
\arguments{
\item{x}{MAgPIE object containing Bodirsky data country-region resolution}
}
\value{
Bodirsky data as MAgPIE object aggregated to country level
}
\description{
Convert Bodirsky data on ISO country level.
}
\examples{

\dontrun{ a <- convertBodirsky(x)
}

}
\author{
Lavinia Baumstark
}
