% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBP.R
\name{convertBP}
\alias{convertBP}
\title{Disaggregates and cleans BP data.}
\usage{
convertBP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Either "Capacity" or "Generation"}
}
\value{
A magpie object with historical electricity capacities (in MW for Wind, Solar, and Geothermal),
Generation (in TWh for Nuclear, Hydro, Wind, Solar, and Geo Biomass), AND
Production (in EJ for Coal, Oil, and Gas, additionally in tonnes for coal)
}
\description{
Disaggregates historical - capacity, generation, and production data.
}
\author{
Aman Malik
}
