% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCEMOdb.R
\name{convertCEMOdb}
\alias{convertCEMOdb}
\title{Converts CEMO database data}
\usage{
convertCEMOdb(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
magpie object of the CEMO data
}
\description{
Converts CEMO database data
}
\seealso{
\code{\link{readSource}}
}
\author{
Felix Schreyer
}
