% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertClimateTrace.R
\name{convertClimateTrace}
\alias{convertClimateTrace}
\title{Convert ClimateTrace data}
\usage{
convertClimateTrace(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readClimateTrace]{readClimateTrace()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert ClimateTrace data
}
\author{
Pascal Weigmann
}
