% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertConferenceBoard.R
\name{convertConferenceBoard}
\alias{convertConferenceBoard}
\title{convertConferenceBoard}
\usage{
convertConferenceBoard(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A magpie object with country-level results
}
\description{
convertConferenceBoard
}
\author{
Aman Malik
}
