% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCSIRO.R
\name{convertCSIRO}
\alias{convertCSIRO}
\title{Converts CSIRO data}
\usage{
convertCSIRO(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
magpie object of the CSIRO data
}
\description{
Converts CSIRO data
}
\seealso{
\code{\link{readSource}}
}
\author{
Felix Schreyer
}
