% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDaviesCooling.R
\name{convertDaviesCooling}
\alias{convertDaviesCooling}
\title{Convert Davies Cooling}
\usage{
convertDaviesCooling(x)
}
\arguments{
\item{x}{MAgPIE object containing DaviesCooling data region resolution}
}
\value{
MAgPIE object of the Davies (2013) data disaggregated to country
level
}
\description{
Convert Davies (2013) data on on shares of cooling types using mapping from
GCAM regions to ISO country level.
}
\examples{

\dontrun{ a <- convertDaviesCooling(x)
}

}
\seealso{
\code{\link{readDaviesCooling}}
}
\author{
Lavinia Baumstark, Ioanna Mouratiadou
}
