% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDemandModel.R
\name{convertDemandModel}
\alias{convertDemandModel}
\title{Convert DemandModel data}
\usage{
convertDemandModel(x)
}
\arguments{
\item{x}{MAgPIE object containing DemandModel data country-region resolution}
}
\value{
DemandModel data as MAgPIE object aggregated to country level
}
\description{
Convert DemandModel data on ISO country level.
}
\examples{

\dontrun{ a <- convertDemandModel(x)
}

}
\author{
Lavinia Baumstark
}
