% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDias.R
\name{convertDias}
\alias{convertDias}
\title{convertDias}
\usage{
convertDias(x, subtype)
}
\arguments{
\item{x}{magpie object to be converted}

\item{subtype}{"Employment factors" or "Employment"}
}
\description{
convertDias
}
\author{
Aman Malik
}
