% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDylanAusGasCost.R
\name{convertDylanAusGasCost}
\alias{convertDylanAusGasCost}
\title{Converts Dylan's Australian gas cost to magpie}
\usage{
convertDylanAusGasCost(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
magpie object of the CEMO data
}
\description{
Converts Dylan's Australian gas cost to magpie
}
\seealso{
\code{\link{readSource}}
}
\author{
Felix Schreyer
}
