% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEATLancet.R
\name{convertEATLancet}
\alias{convertEATLancet}
\title{Convert data from the EAT Lancet Comission}
\usage{
convertEATLancet(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing EAT Lancet data at mixed country-region
resolution}

\item{subtype}{Type of EAT Lancet data that should be read. Available types are:
\itemize{ 
\item \code{cons_data}: Consumption analysis ("EAT_Lancet_cons_data.csv")
}}
}
\value{
EAT Lancet data as MAgPIE object at ISO country level
}
\description{
Convert data from the EAT Lancet Comission to ISO country level.
}
\examples{

\dontrun{
a <- readSource(type="EATLancet",subtype="cons_data")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
