% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGAR7Fgases.R
\name{convertEDGAR7Fgases}
\alias{convertEDGAR7Fgases}
\title{convertEDGAR7Fgases}
\usage{
convertEDGAR7Fgases(x)
}
\arguments{
\item{x}{magpie object to be converted}
}
\description{
convertEDGAR7Fgases
}
\author{
Gabriel Abrahao
}
