% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEdgeBuildings.R
\name{convertEDGE}
\alias{convertEDGE}
\title{Convert EDGE Buildings data to data on ISO country level.}
\usage{
convertEDGE(x, subtype = "FE")
}
\arguments{
\item{x}{MAgPIE object containing EDGE values at ISO country resolution}

\item{subtype}{either FE or Floorspace}
}
\value{
EDGE data as MAgPIE object aggregated to country level
}
\description{
Convert EDGE Buildings data to data on ISO country level.
}
\author{
Antoine Levesque, Robin Hasse
}
