% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGE_Industry.R
\name{convertEDGE_Industry}
\alias{convertEDGE_Industry}
\title{Convert EDGE Industry Data}
\usage{
convertEDGE_Industry(x, subtype)
}
\arguments{
\item{x}{A [`magpie`][magclass::magclass] object.}

\item{subtype}{Currently only
* `p29_capitalQuantity_industry` for industry subsector energy efficiency 
  capital trajectories}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Convert EDGE Industry Data
}
\seealso{
[`readSource()`]
}
\author{
Michaja Pehl
}
