% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGEtranspLDV.R
\name{convertEDGEtranspLDV}
\alias{convertEDGEtranspLDV}
\title{Convert EDGE Transport LDV shares}
\usage{
convertEDGEtranspLDV(x)
}
\arguments{
\item{x}{MAgPIE object containing EDGE Transport LDV shares values at ISO country resolution}
}
\value{
EDGE Transport LDV shares data as MAgPIE object aggregated to country level
}
\description{
Convert EDGE Transport LDV shares
}
\author{
Renato Rodrigues
}
