% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGETransport.R
\name{convertEDGETransport}
\alias{convertEDGETransport}
\title{Convert EDGEtransport}
\usage{
convertEDGETransport(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing EDGE-T values in 21 region resolution}

\item{subtype}{REMIND/iterative EDGE-T input data subtypes}
}
\value{
REMIND/iterative EDGE-T input data as MAgPIE object disaggregated to ISO level
}
\description{
Convert EDGEtransport
}
\author{
Johanna Hoppe
}
