% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEEA_EuropeanEnvironmentAgency.R
\name{convertEEA_EuropeanEnvironmentAgency}
\alias{convertEEA_EuropeanEnvironmentAgency}
\title{Convert European Environment Agency (EEA) data}
\usage{
convertEEA_EuropeanEnvironmentAgency(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{data subtype. Either "ETS", "historical", "projections", or "projections-detailed"}
}
\value{
magpie object of European Environment Agency (EEA) ETS emissions (GtCO2)
}
\description{
Read-in European Environment Agency (EEA) data on ETS emissions as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "EEA_EuropeanEnvironmentAgency", subtype = "ETS")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
