% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEmber.R
\name{convertEmber}
\alias{convertEmber}
\title{Convert Ember data}
\usage{
convertEmber(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readEmber]{readEmber()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert Ember data
}
\author{
Pascal Weigmann
}
