% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEndCoal.R
\name{convertEndCoal}
\alias{convertEndCoal}
\title{Data on Coal Plants (MW) from EndCoal.org , July 2018}
\usage{
convertEndCoal(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
Magpie object with country level data on coal plants (MW) in various stages
}
\description{
Data on Coal Plants (MW) from EndCoal.org , July 2018
}
\author{
Aman Malik
}
