% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEORA.R
\name{convertEORA}
\alias{convertEORA}
\title{convertEORA}
\usage{
convertEORA(x)
}
\arguments{
\item{x}{MAgpIE object containing EORA data on country level}
}
\value{
EORA data as complete MAgPIE object on country level
}
\description{
Converts readEORA output to complete MAgPIE object
}
\seealso{
\code{\link{readSource}}
}
\author{
Debbora Leip
}
