% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEU_ReferenceScenario.R
\name{convertEU_ReferenceScenario}
\alias{convertEU_ReferenceScenario}
\title{Convert EU Reference Scenario}
\source{
EU Reference Scenario public database http://data.europa.eu/euodp/en/data/dataset/energy-modelling
}
\usage{
convertEU_ReferenceScenario(x)
}
\arguments{
\item{x}{EU Reference Scenario magpie object derived from readEU_ReferenceScenario function}
}
\value{
converted EU Reference Scenario magpie object
}
\description{
Converts EU Reference Scenario magpie object into appropriate form for the REMIND model
}
\examples{
\dontrun{test <- readSource("EU_ReferenceScenario",convert=TRUE)}
}
\author{
Renato Rodrigues
}
