% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanEnergyDatasheets.R
\name{convertEuropeanEnergyDatasheets}
\alias{convertEuropeanEnergyDatasheets}
\title{Convert European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/country
}
\usage{
convertEuropeanEnergyDatasheets(x)
}
\arguments{
\item{x}{European Energy Datasheets magpie object derived from readEuropeanEnergyDatasheets function}
}
\value{
converted European Energy Datasheets magpie object
}
\description{
Converts European Energy Datasheets magpie object into appropriate form for the REMIND model
}
\examples{
\dontrun{test <- readSource("EuropeanEnergyDatasheets",convert=TRUE)}
}
\author{
Renato Rodrigues and Atreya Shankar
}
