% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostat_EffortSharing.R
\name{convertEurostat_EffortSharing}
\alias{convertEurostat_EffortSharing}
\title{Converts EU Effort Sharing targets and historical emissions}
\usage{
convertEurostat_EffortSharing(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{data subtype. Either "target" or "emissions"}
}
\value{
A MAgPIE object containing the EU Effort Sharing targets (\%) or Effort Sharing historical emissions (MtCO2)
}
\description{
Converts EU Effort Sharing targets and historical emissions
}
\examples{

\dontrun{ a <- convertEurostat_EffortSharing(x,subtype="target")
}

}
\author{
Renato Rodrigues
}
