% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostat_REShare.R
\name{convertEurostat_REShare}
\alias{convertEurostat_REShare}
\title{Converts Eurostat renewable energy share data}
\usage{
convertEurostat_REShare(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{data subtype}
}
\value{
A MAgPIE object containing Eurostat renewable energy shares
}
\description{
Converts Eurostat renewable energy share data
}
\author{
Felix Schreyer
}
