% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExpertGuess.R
\name{convertExpertGuess}
\alias{convertExpertGuess}
\title{convertExpertGuess}
\usage{
convertExpertGuess(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{Type of data that are converted.}
}
\description{
Converts data from expert guess
}
