% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExpertGuessSSPLivestockProduction.R
\name{convertExpertGuessSSPLivestockProduction}
\alias{convertExpertGuessSSPLivestockProduction}
\title{convertExpertGuessSSPLivestockProduction}
\usage{
convertExpertGuessSSPLivestockProduction(x)
}
\arguments{
\item{x}{MAgPIE-Object contaiing data to convert}
}
\value{
magpie object containing converted expert guesses
}
\description{
{convert the Expert Guesses for future Livestock Production for the SSP Scenarios
}
}
\examples{
\dontrun{ 
a <- readSource("ExperGuessSSPLivestockProduction", "ssp1")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
