% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO.R
\name{convertFAO}
\alias{convertFAO}
\title{Convert FAO data}
\usage{
convertFAO(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{The FAO file type, e.g.: CBCrop}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Converts FAO data to fit to the common country list and removes or converts
relative values where possible. Yields (Hg/ha) are for instance removed
since they can later easily be calculated from production and area but might
be problematic in the spatial aggregation. Per capita demand values are
transformed into absolute values using population estimates from the 
calcPopulationPast function.
}
\details{
Update 23-Jan-2017 - Added FAO Forestry production and trade data (Abhi)
}
\examples{

\dontrun{ a <- readSource("FAO","Crop", convert=TRUE)}
}
\seealso{
\code{\link{readFAO}}, \code{\link{readSource}},
}
\author{
Ulrich Kreidenweis, Abhijeet Mishra, Mishko Stevanovic
}
