% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO_forestry.R
\name{convertFAO_forestry}
\alias{convertFAO_forestry}
\title{Convert FAO data for forest}
\usage{
convertFAO_forestry(x)
}
\arguments{
\item{x}{MAgPIE object containing values of the forest data}
}
\value{
Forest within protected areas data as MAgPIE object with common
country list
}
\description{
Convert the FAO data to fit to the common country list. Missing countries
are added and assigned the value zero.
}
\author{
Nele Steinmetz
}
