% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO_FRA2015.R
\name{convertFAO_FRA2015}
\alias{convertFAO_FRA2015}
\title{Convert FRA 2015 data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)}
\usage{
convertFAO_FRA2015(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{The FAO FRA 2015 file type, e.g.: fac, production, biodiversity or anndat.}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Convert FRA 2015 data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)
}
\examples{

\dontrun{ a <- readSource("FRA2015","production",convert=TRUE)}
}
\seealso{
\code{\link{readSource}},
}
\author{
Abhijeet Mishra
}
