% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOLossesWaste.R
\name{convertFAOLossesWaste}
\alias{convertFAOLossesWaste}
\title{Convert data on  food losses and waste from FAO for several commodity groups}
\usage{
convertFAOLossesWaste(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing data on  food losses and waste at mixed country-region
resolution}

\item{subtype}{Steps of the food supply chain where food losses and waste occur. Available types are:
\itemize{ 
\item \code{Consumption}: consumption level
}}
}
\value{
Data on  food losses and waste as MAgPIE object at ISO country level
}
\description{
Convert data on  food losses and waste on ISO country level.
}
\examples{

\dontrun{
a <- readSource(type="FAOLossesWaste",subtype="Consumption")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
