% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFRA2015Doc.R
\name{convertFRA2015Doc}
\alias{convertFRA2015Doc}
\title{Convert FRA2015Doc data}
\usage{
convertFRA2015Doc(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values coming from read function}

\item{subtype}{The data table type, e.g.: forest_area}
}
\value{
Data as MAgPIE object
}
\description{
Convert FRA2015Doc data
}
\examples{

\dontrun{ a <- readSource("FRA2015Doc","forest_area", convert=TRUE)}
}
\seealso{
\code{\link{readFRA2015Doc}}, \code{\link{readSource}},
}
\author{
Abhijeet Mishra
}
