% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGEA2012.R
\name{convertGEA2012}
\alias{convertGEA2012}
\title{convertGEA2012}
\usage{
convertGEA2012(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be disaggregated}

\item{subtype}{Type of fossil fuel (oil, coal or gas)}
}
\value{
MAgPIE object containing country-level disaggregation of GEA 2012 data
}
\description{
Converts oil, gas and coal data from the Global Energy Assessment 2012 to country-level aggregation
}
\examples{
\dontrun{
a <- readSource("GEA2012")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Bi
}
