% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGernaat.R
\name{convertGernaat}
\alias{convertGernaat}
\title{Convert Hydro-Potential data from Gernaat et. al 2017}
\source{
Obtained from Nature Energy volume 2 (2017), pp 821–828, with following source: https://www.nature.com/articles/s41560-017-0006-yy
}
\usage{
convertGernaat(x, subtype)
}
\arguments{
\item{x}{magpie object derived from readGernaat()}

\item{subtype}{Possibilities for subtype are "Technical_Potential" or "Economic_Potential"}
}
\value{
magpie object of converted hydropotential data from Gernaat et. al 2017
}
\description{
Convert magpie object of hydropotential data from Gernaat et. al 2017. Technical potential
defined as <0.5 $/kWh and economic potential as <0.1 $/kWh.
}
\examples{
\dontrun{a <- convertGernaat(x, "Technical_Potential")}
}
\author{
Atreya Shankar
}
