% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGFED.R
\name{convertGFED}
\alias{convertGFED}
\title{convertGFED}
\usage{
convertGFED(x, subtype)
}
\arguments{
\item{x}{MAgPIE object obatined from calcGFED()}

\item{subtype}{emissions or emissionfactors}
}
\value{
magpie object of the GFED data in country level
}
\description{
Transforms GFED data on cellular level to country level
}
\examples{

\dontrun{ x <- readGFED(subtype="emissions")
}
\dontrun{ y <- convertGFED(x)
}
}
\seealso{
\code{\link{readGFED}} {\link{extractGFED}}
}
\author{
Lavinia Baumstark, Abhijeet Mishra
}
