% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGGDC10.R
\name{convertGGDC10}
\alias{convertGGDC10}
\title{Convert GGDC 10-Sector Database - https://www.rug.nl/ggdc/structuralchange/previous-sector-database/10-sector-2014}
\usage{
convertGGDC10(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing GGDC disaggregated data
}
\description{
Convert GGDC 10-Sector Database - https://www.rug.nl/ggdc/structuralchange/previous-sector-database/10-sector-2014
}
\examples{
 \dontrun{
a <- convertGGDC10(x)
}

}
\author{
Renato Rodrigues
}
