% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/convertGidden2025_geological_storage_potential.R
\name{convertGidden2025_geological_storage_potential}
\alias{convertGidden2025_geological_storage_potential}
\title{Convert geological storage potential}
\usage{
convertGidden2025_geological_storage_potential(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A magpie object with geological storage potential
}
\description{
Convert geological storage potential
}
\author{
David Klein
}
