% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGini.R
\name{convertGini}
\alias{convertGini}
\title{Convert Gini}
\usage{
convertGini(x)
}
\arguments{
\item{x}{MAgPIE object containing Gini data with World Bank codes, 2011-2100, in percent (range 0-100)}
}
\value{
MAgPIE object of the Gini data in ISO countries, range 0-1
}
\description{
Converts Gini data from readGini() to ISO country level. 
Countries missing in the original data set will have their Gini set to zero (a very small number for numerical reasons to be precise).
The original data range is 2011-2100 in one-year steps, here we extend it to 2000-2150 in 5-year steps. 
Values before (after) the original range are held fixed at 2011 (2100) levels.
}
\examples{

\dontrun{ a <- convertGini(x)
}
}
\seealso{
\code{\link{readSource}} \code{\link{readGini}}
}
\author{
Bjoern Soergel
}
