% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGlobalCCSinstitute.R
\name{convertGlobalCCSinstitute}
\alias{convertGlobalCCSinstitute}
\title{Convert Global CCS Institute Project Database}
\usage{
convertGlobalCCSinstitute(x, subtype = "08-09-2017")
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned by \code{\link[=readGlobalCCSinstitute]{readGlobalCCSinstitute()}}.}

\item{subtype}{Project Database version to read, one of
- \code{'08-09-2017'}: Data apparently from June 2017.
- \code{'2023-11'}: Data from the Global Status of CCS 2023 report.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert Global CCS Institute Project Database
}
