% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGlobalEnergyMonitor.R
\name{convertGlobalEnergyMonitor}
\alias{convertGlobalEnergyMonitor}
\title{Convert Global Energy Monitor data}
\usage{
convertGlobalEnergyMonitor(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readGlobalEnergyMonitor]{readGlobalEnergyMonitor()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert Global Energy Monitor data
}
\author{
Falk Benke
}
