% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGMIA.R
\name{convertGMIA}
\alias{convertGMIA}
\title{convertGMIA}
\usage{
convertGMIA(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing IrrigatedArea data on Country level}

\item{subtype}{: No subtype needed}
}
\value{
Global Map on Irrigation data as MAgPIE object on country level Missing values are added as NA
}
\description{
Convert Global Map on Irrigated Area Data

Convert subtypes on ISO country level.
}
\examples{

\dontrun{
 a <- readSource("GlobalIrrigationMap", "all_data_national")
a <- readSource ("GMIA", "aei_pct", convert=F)
a <- readSource ("GMIA", "aei_pct", convert="correctonly")
}



}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
