% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGPFADB_2005.R
\name{convertGPFADB_2005}
\alias{convertGPFADB_2005}
\title{Convert Global plantation forest data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)}
\usage{
convertGPFADB_2005(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{Available subtypes are Plantations, SNForest, PlantedSNForest, TotPlanted, GroParInit, AgeClassInit, Ownership and EndUse}
}
\value{
Data as MAgPIE object
}
\description{
Convert Global plantation forest data
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)
}
\examples{

\dontrun{ a <- readSource("GPFADB_2005","Plantations",convert=TRUE)}
}
\seealso{
\code{\link{readSource}},
}
\author{
Abhijeet Mishra
}
