% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGTAP.R
\name{convertGTAP}
\alias{convertGTAP}
\title{convertGTAP}
\usage{
convertGTAP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object contains GTAP data}

\item{subtype}{The GTAP subtype: VIWS, VIMS VXWD, VXMD, VOA, VOM}
}
\value{
Converted GTAP Data
}
\description{
Converts GTAP data to fit to the common country list. Weighting is done by using the Imports and Exports from FAO
}
\examples{

  \dontrun{
    x <- ReadSource("GTAP", "GTAP7_VIMS")
  }
}
\author{
Xiaoxi Wang
}
