% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHDD_CDD.R
\name{convertHDD_CDD}
\alias{convertHDD_CDD}
\title{convert HDD_CDD}
\usage{
convertHDD_CDD(x)
}
\arguments{
\item{x}{MAgPIE object containing HDD CDD values at ISO country resolution}
}
\value{
HDD CDD data as MAgPIE object aggregated to country level
}
\description{
Convert historic and scenario based HDD CDD data to data on ISO country
level.
}
\author{
Antoine Levesque
}
