% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHRE.R
\name{convertHRE}
\alias{convertHRE}
\title{Convert HRE data}
\usage{
convertHRE(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readHRE]{readHRE()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert HRE data
}
\author{
Pascal Weigmann
}
