% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIAEA.R
\name{convertIAEA}
\alias{convertIAEA}
\title{Nuclear data from world-nuclear.org}
\usage{
convertIAEA(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Data on currently operating and under-construction nuclear power plants, reactors planned and proposed, 
electricity generation from nuclear
}
\author{
Christoph Bertram
}
