% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIAEA_PRIS.R
\name{convertIAEA_PRIS}
\alias{convertIAEA_PRIS}
\title{convert IAEA Power Reactor Information System}
\usage{
convertIAEA_PRIS(x)
}
\arguments{
\item{x}{a magclass object returned from `readIAEA_PRIS()`}
}
\description{
convert IAEA Power Reactor Information System
}
\author{
Pascal Weigmann
}
