% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA.R
\name{convertIEA}
\alias{convertIEA}
\title{Convert IEA}
\usage{
convertIEA(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing IEA values at IEA mixed country-region
resolution}

\item{subtype}{data subtype. Either "EnergyBalances" or "CHPreport"}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Convert IEA energy data to data on ISO country level.
}
\author{
Anastasis Giannousakis, Renato Rodrigues
}
