% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_CCUS.R
\name{convertIEA_CCUS}
\alias{convertIEA_CCUS}
\title{Convert IEA CCUS data}
\usage{
convertIEA_CCUS(x)
}
\arguments{
\item{x}{A magclass object returned from readIEA_CCUS().}
}
\value{
A magclass object.
}
\description{
Convert IEA CCUS data
}
\author{
Anne Merfort, Falk Benke
}
