% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_ETP.R
\name{convertIEA_ETP}
\alias{convertIEA_ETP}
\title{Convert IEA ETP projections}
\usage{
convertIEA_ETP(x, subtype)
}
\arguments{
\item{x}{IEA ETP projection magpie object derived from readIEA_ETP function}

\item{subtype}{data subtype. Either "main" or "industry_subsectors"}
}
\description{
Convert IEA ETP projections
}
\author{
Falk Benke
}
