% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_EV.R
\name{convertIEA_EV}
\alias{convertIEA_EV}
\title{Convert IEA_EV}
\usage{
convertIEA_EV(x)
}
\arguments{
\item{x}{MAgPIE object containing IEA_EV country resolution}
}
\value{
Krausmann data as MAgPIE object aggregated/disaggregated to country
level
}
\description{
Convert IEA_EV on ISO country level.
}
\examples{

\dontrun{ a <- convertIEA_EV(x)
}
}
\seealso{
\code{\link{readIEA_EV}}
}
\author{
Lavinia Baumstark, Christoph Bertram
}
