% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_EVOutlook.R
\name{convertIEA_EVOutlook}
\alias{convertIEA_EVOutlook}
\title{Convert IEA EV Outlook}
\usage{
convertIEA_EVOutlook(x)
}
\arguments{
\item{x}{a magclass object returned from \code{readIEA_EVOutlook()}}
}
\description{
Convert IEA EV Outlook
}
\author{
Falk Benke
}
