% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_HSMR.R
\name{convertIEA_HSMR}
\alias{convertIEA_HSMR}
\title{convert IEA Hydro Special Market Report}
\usage{
convertIEA_HSMR(x)
}
\arguments{
\item{x}{a magclass object returned from \code{readIEA_HSMR()}}
}
\description{
convert IEA Hydro Special Market Report
}
\author{
Pascal Weigmann
}
