% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_HydrogenProduction.R
\name{convertIEA_HydrogenProduction}
\alias{convertIEA_HydrogenProduction}
\title{Convert IEA hydrogen capacities}
\usage{
convertIEA_HydrogenProduction(x)
}
\arguments{
\item{x}{A magclass object returned from readIEA_HydrogenCapacities().}
}
\value{
A magclass object.
}
\description{
Convert IEA hydrogen capacities
}
\author{
Simon Krogmann
}
