% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_PRIS.R
\name{convertIEA_PRIS}
\alias{convertIEA_PRIS}
\title{convert IEA Power Reactor Information System}
\usage{
convertIEA_PRIS(x)
}
\arguments{
\item{x}{a magclass object returned from `readIEA_PRIS()`}
}
\description{
convert IEA Power Reactor Information System
}
\author{
Pascal Weigmann
}
