% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_REN.R
\name{convertIEA_REN}
\alias{convertIEA_REN}
\title{Reads the distributed solar pv capacity from IEA Renewables report (2019).}
\usage{
convertIEA_REN(x)
}
\arguments{
\item{x}{input magpie object}
}
\value{
magpie object with country-wise distributed solar pv capacity
}
\description{
Reads the distributed solar pv capacity from IEA Renewables report (2019).
}
\details{
Capacity in GW. Distributed solar, defined in the IEA Renewables (2019), includes 
rooftop residential (0-10 kW, grid-connected), rooftop and ground-mounted commercial
 and industrial (10-1000kW, grid-connected), and off-grid (8W - 100 kW)
}
\author{
Aman Malik
}
