% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_WEO.R
\name{convertIEA_WEO}
\alias{convertIEA_WEO}
\title{Converts IEA World Energy Outlook data}
\usage{
convertIEA_WEO(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{data subtype. Either "Capacity" or "Invest_Costs"}
}
\description{
Converts IEA World Energy Outlook data
}
\author{
Renato Rodrigues and Aman Malik
}
