% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_WEO_2021.R
\name{convertIEA_WEO_2021}
\alias{convertIEA_WEO_2021}
\title{Disaggregates IEA WEO 2021 Data}
\usage{
convertIEA_WEO_2021(x, subtype = "GLO")
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Either "GLO" or "regional"}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Disaggregates IEA WEO 2021 Data
}
\author{
Falk Benke
}
