% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIFA.R
\name{convertIFA}
\alias{convertIFA}
\title{Convert IFA}
\usage{
convertIFA(x)
}
\arguments{
\item{x}{MAgPIE object containing IFA data region resolution}
}
\value{
MAgPIE object of the IFA data disaggregated to country level
}
\description{
Convert IFADavies (2013) data to ISO country level.
}
\examples{

\dontrun{ a <- convertIFA(x)
}

}
\author{
Lavinia Baumstark
}
