% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIIASA_subs_taxes.R
\name{convertIIASA_subs_taxes}
\alias{convertIIASA_subs_taxes}
\title{Convert IIASA_subs_taxes data}
\usage{
convertIIASA_subs_taxes(x)
}
\arguments{
\item{x}{MAgPIE object containing IIASA subsidies and taxes data in country
resolution}
}
\value{
IIASA_subs_taxes data as MAgPIE object aggregated to country level
}
\description{
Convert IIASA subsidy and taxes data on ISO country level (removes countries
not part of 249 oficial ISO countries and fills missing with zeros).
}
\examples{

\dontrun{ a <- convertIIASA_subs_taxes(x)
}

}
\author{
Christoph Bertram
}
