% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIIASALanduse.R
\name{convertIIASALanduse}
\alias{convertIIASALanduse}
\title{convert IIASA Land Use Emissions}
\usage{
convertIIASALanduse(x, subtype)
}
\arguments{
\item{x}{a magpie object}

\item{subtype}{one of "historical", "forecast2030", "forecast2035"}
}
\description{
convert IIASA Land Use Emissions
}
\author{
Falk Benke
}
