% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIIASApop.R
\name{convertIIASApop}
\alias{convertIIASApop}
\title{Convert IIASApop}
\usage{
convertIIASApop(x)
}
\arguments{
\item{x}{MAgPIE object containing population data mixed country-region
resolution}
}
\value{
population data as MAgPIE object aggregated to country level
}
\description{
Convert population data to data on ISO country level.
}
\author{
Lavinia Baumstark
}
