% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMAGE2025.R
\name{convertIMAGE2025}
\alias{convertIMAGE2025}
\title{Disaggregates IMAGE 2025 F-gas emissions data,
using EDGAR HFC emissions in 2005 as weights}
\usage{
convertIMAGE2025(x)
}
\arguments{
\item{x}{magpie object}
}
\value{
magpie object
}
\description{
Disaggregates IMAGE 2025 F-gas emissions data,
using EDGAR HFC emissions in 2005 as weights
}
\author{
Gabriel Abrahao
}
