% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIPCC.R
\name{convertIPCC}
\alias{convertIPCC}
\title{Convert subtypes of the IPCC data}
\usage{
convertIPCC(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing IPCC data on region level}

\item{subtype}{data subtype. Either "awmsShr", "awmsEfCh4", "awmsParCh4",
"nExcrRate"}
}
\value{
IPCC data as MAgPIE object for all subtypes on country level
}
\description{
Convert subtypes on ISO country level.
}
\examples{

\dontrun{
a <- readSource("IPCC","awmsShr")
a <- readSource("IPCC","awmsEfCh4")
a <- readSource("IPCC","awmsParCh4")
a <- readSource("IPCC","nExcrRate")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Nele Steinmetz
}
