% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJames.R
\name{convertJames}
\alias{convertJames}
\title{Convert James data}
\usage{
convertJames(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing James data region resolution}

\item{subtype}{subtype of GDP indicator to be selected}
}
\value{
Krausmann data as MAgPIE object aggregated/disaggregated to country
level
}
\description{
Convert James data on ISO country level.
}
\examples{

\dontrun{ a <- convertJames(x,"IHME_USD05_PPP_pc")
}
}
\seealso{
\code{\link{readJames}}
}
\author{
Benjamin Bodirsky
}
