% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJRC_IDEES.R
\name{convertJRC_IDEES}
\alias{convertJRC_IDEES}
\title{Convert JRC IDEES data}
\usage{
convertJRC_IDEES(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readJRC_IDEES]{readJRC_IDEES()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Missing data for EU-28 countries is added, by distributing the difference of
\code{EU28} and the sum of country-values based on countries share in EU-28 GDP.
}
\seealso{
\code{\link[=readJRC_IDEES]{readJRC_IDEES()}}
}
\author{
Michaja Pehl
}
