% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKLWdamage.R
\name{convertKLWdamage}
\alias{convertKLWdamage}
\title{convert KLW damage
fills in countries for which no damage parameters are provided, setting parameters to zero}
\usage{
convertKLWdamage(x)
}
\arguments{
\item{x}{is MAgPIE object containing the damage parameters from KLW}
}
\value{
MAgPIE object containing values for all 249 ISO countries
}
\description{
convert KLW damage
fills in countries for which no damage parameters are provided, setting parameters to zero
}
\author{
Franziska Piontek
}
