% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKoeppenGAMS.R
\name{convertKoeppenGAMS}
\alias{convertKoeppenGAMS}
\title{convertKoeppenGAMS}
\usage{
convertKoeppenGAMS(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{Not available.}
}
\value{
Koeppen geiger climate zone classsification data
}
\description{
koeppen geiger climate data compatible to GAMS usage.
}
\examples{

\dontrun{ 
readSource("readKoeppenGAMS",convert="FALSE")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
