% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKrausmann.R
\name{convertKrausmann}
\alias{convertKrausmann}
\title{Convert Krausmann data}
\usage{
convertKrausmann(x)
}
\arguments{
\item{x}{MAgPIE object containing Krausmann data region resolution}
}
\value{
Krausmann data as MAgPIE object aggregated/disaggregated to country
level
}
\description{
Convert Krausmann data on ISO country level.
}
\examples{

\dontrun{ a <- convertKrausmann(x)
}

}
\author{
Lavinia Baumstark
}
