% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLandfillCH4Capture.R
\name{convertLandfillCH4Capture}
\alias{convertLandfillCH4Capture}
\title{Convert landfill CH4 capture data}
\usage{
convertLandfillCH4Capture(x)
}
\arguments{
\item{x}{MAgPIE object containing original values}
}
\value{
DemandModel data as MAgPIE object aggregated to country level
}
\description{
Convert landfill CH4 capture data
}
\author{
David Chen
}
