% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLassaletta2014.R
\name{convertLassaletta2014}
\alias{convertLassaletta2014}
\title{convertLassaletta2014}
\usage{
convertLassaletta2014(x, subtype)
}
\arguments{
\item{x}{data object that gets provided by wrapper function readSource}

\item{subtype}{budget provides the nr cropland budgets, fert_to_cropland the sahre of inorganic fertilizers being applied to croplands}
}
\value{
Magpie object with results on country level.
}
\description{
converts the dataset of
Lassaletta, L., G. Billen, B. Grizzetti, J. Angalde, and J. Garnier. 2014. 
50 Year Trends in Nitrogen Use Efficiency of World Cropping Systems: The Relationship between Yield and Nitrogen Input to Cropland.
Environmental Research Letters.
into a dataset including all countries. Replacing Soviet Union by Russia and Yugoslavia by Serbia without detailed disaggregation.
}
\examples{

\dontrun{ 
readSource("Lassaletta2014",convert=TRUE)
}
}
\seealso{
\code{\link{readLassaletta2014}},
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky
}
