% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLotzeCampenBiofuel.R
\name{convertLotzeCampenBiofuel}
\alias{convertLotzeCampenBiofuel}
\title{convertLotzeCampenBiofuel}
\usage{
convertLotzeCampenBiofuel(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing future trends in first generation bioenergy demand in
Petajoules as magpie object for each country for biodiesel and ethanol.
}
\description{
Converts the Data from Lotze Campen et al. 2014. to fit the common country list.
Source: Lotze Campen et al. 2014. "Impacts of increased bioenergy demand on global food
markets: an AgMIP economic model intercomparison" Agricultural Economics 45 (103-116).
doi:10.1111/agec.12092.
}
\examples{


  \dontrun{
    x <- readSource("LotzeCampenBiofuel")
  }


}
\author{
Ewerton Araujo
}
