% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLutz2014.R
\name{convertLutz2014}
\alias{convertLutz2014}
\title{convertLutz2014}
\usage{
convertLutz2014(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
It fills the missing values of the output of readLutz2014 through 
the weighted average of the values of two countries with similar carachteristcs
to the one that has na values.
}
\seealso{
\code{\link{readLutz2014}}
}
