% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMarineRegionsOrg.R
\name{convertMarineRegionsOrg}
\alias{convertMarineRegionsOrg}
\title{Convert Exclusive Economic Zone (EEZ) size data}
\usage{
convertMarineRegionsOrg(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Convert Exclusive Economic Zone (EEZ) size data
}
\author{
Tabea Dorndorf
}
