% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMueller.R
\name{convertMueller}
\alias{convertMueller}
\title{Convert Mueller data}
\usage{
convertMueller(x, subtype)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readMueller]{readMueller()}}.}

\item{subtype}{One of:
\itemize{
\item \code{countries}: read table mapping country names use by Müller et al. 2013
to ISO 3166-1 alpha-3 codes.
\item \code{stocks}: read low/medium/high estimates of per-capita steel stocks from
Müller et al. 2013 SI2
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert Mueller data
}
\author{
Falk Benke
}
