% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNCDrisc.R
\name{convertNCDrisc}
\alias{convertNCDrisc}
\title{convertNCDrisc}
\usage{
convertNCDrisc(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{"height" for body height data. Missing data is replaced by non-population weighted global average}
}
\value{
magpie object with a completed dataset.
}
\description{
Converts data from the NCD risc consortium
body height:
Collaboration (NCD-RisC), NCD Risk Factor. 2016. "A Century of Trends in Adult Human Height." ELife 5 (July):e13410. https://doi.org/10.7554/eLife.13410.
}
\seealso{
\code{\link{convertNCDrisc}}
}
