% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNewClimate.R
\name{convertNewClimate}
\alias{convertNewClimate}
\title{Convert policy targets for NPIs from New Climate policy database}
\usage{
convertNewClimate(x, subtype, subset)
}
\arguments{
\item{x}{a magclass object to be converted}

\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets,
RenShareTargets for renewable energy share targets, with YYYY NPI version year}

\item{subset}{A string (or vector of strings) designating the scenario(s) to be returned.}
}
\description{
Converts conditional and unconditional capacity and production targets into total capacity (GW) in target year.
For countries and years without targets, 2020 values from IRENA and BP are used to fill the gaps.
}
\seealso{
\code{\link[=readIRENA]{readIRENA()}}
}
\author{
Rahel Mandaroux, Léa Hayez, Falk Benke
}
