% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertnonEnergyDemand.R
\name{convertnonEnergyDemand}
\alias{convertnonEnergyDemand}
\title{Converts Final energy demand for feedstocks (non-energy use)}
\usage{
convertnonEnergyDemand(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing country disaggregated data
}
\description{
Converts Final energy demand for feedstocks (non-energy use)
}
\examples{

\dontrun{ a <- convertnonEnergyDemand(x)
}

}
\author{
Renato Rodrigues
}
