% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNREL.R
\name{convertNREL}
\alias{convertNREL}
\title{Convert NREL data}
\usage{
convertNREL(x)
}
\arguments{
\item{x}{MAgPIE object containing NREL data country-region resolution}
}
\value{
NRELWirsenius data as MAgPIE object aggregated to country level
}
\description{
Convert NREL data on ISO country level.
}
\examples{

\dontrun{ a <- convertNREL(x,subtype="onshore")
}
}
\author{
Lavinia Baumstark
}
