% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNTNU_LCA.R
\name{convertNTNU_LCA}
\alias{convertNTNU_LCA}
\title{Convert NTNU_LCA data}
\usage{
convertNTNU_LCA(x)
}
\arguments{
\item{x}{MAgPIE object containing NTNU_LCA data country-region resolution}
}
\value{
NTNU_LCA data as MAgPIE object aggregated to country level
}
\description{
Convert NTNU_LCA data on ISO country level.
}
\examples{

\dontrun{ a <- convertNTNU_LCA(x)
}

}
\author{
Lavinia Baumstark
}
