% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECD.R
\name{convertOECD}
\alias{convertOECD}
\title{Convert OECD GDP / ratioPM}
\usage{
convertOECD(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing GDP data mixed country-region resolution}

\item{subtype}{data subtype. Either "gdp" or "ratioPM"}
}
\value{
GDP / ratioPM data as MAgPIE object aggregated to country level
}
\description{
Convert GDP / ratioPM data from OECD to data on ISO country level.
}
\examples{

\dontrun{
a <- readSource("OECD","gdp")
a <- readSource("OECD","ratioPM")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
