% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOpenmod.R
\name{convertOpenmod}
\alias{convertOpenmod}
\title{Converts Openmod capacities data}
\usage{
convertOpenmod(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing openmod EU country disaggregated data with 2010 and 2015 electricity capacities (GW)
}
\description{
Converts Openmod capacities data
}
\examples{

\dontrun{ a <- convertOpenmod(x)
}
 
 
}
\author{
Renato Rodrigues
}
