% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPBL_EFsBaseline.R
\name{convertPBL_EFsBaseline}
\alias{convertPBL_EFsBaseline}
\title{Disaggregates PBL emission factors from the 26 IMAGE regions to ISO3
country level. As these are emission factors, weights are constant,
assuming all countries in a region emit the same amount of each gas
per unit of activity.}
\usage{
convertPBL_EFsBaseline(x)
}
\arguments{
\item{x}{a magpie object}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Disaggregates PBL emission factors from the 26 IMAGE regions to ISO3
country level. As these are emission factors, weights are constant,
assuming all countries in a region emit the same amount of each gas
per unit of activity.
}
