% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPBL_EFsBaseline_2022.R
\name{convertPBL_EFsBaseline_2022}
\alias{convertPBL_EFsBaseline_2022}
\title{Disaggregates PBL emission factors from the 26 IMAGE regions to ISO3
country level. As these are emission factors, weights are constant,
assuming all countries in a region emit the same amount of each gas
per unit of activity.}
\usage{
convertPBL_EFsBaseline_2022(x)
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Disaggregates PBL emission factors from the 26 IMAGE regions to ISO3
country level. As these are emission factors, weights are constant,
assuming all countries in a region emit the same amount of each gas
per unit of activity.
}
