% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPRIMAPhist.R
\name{convertPRIMAPhist}
\alias{convertPRIMAPhist}
\title{convertPRIMAPhist}
\usage{
convertPRIMAPhist(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
MAgPIE object
}
\description{
function to convert PRIMAP-hist data to isocountry resolution
}
\examples{
\dontrun{
readSource("PRIMAPhist", subtype = "hist")
}
}
\author{
Roman Popov
}
