% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertProdShares.R
\name{convertProdShares}
\alias{convertProdShares}
\title{Reads shares of world manufacture for spv modules and wind turbines.}
\usage{
convertProdShares(x)
}
\arguments{
\item{x}{input magpie object}
}
\value{
magpie object with shares
}
\description{
Reads shares of world manufacture for spv modules and wind turbines.
}
\author{
Aman Malik
}
