% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPWT.R
\name{convertPWT}
\alias{convertPWT}
\title{Convert PWT data}
\usage{
convertPWT(x)
}
\arguments{
\item{x}{MAgPIE object containing PWT data country-region resolution}
}
\value{
PWT data as MAgPIE object aggregated to country level
}
\description{
Convert PWT data on ISO country level.
}
\examples{

\dontrun{ a <- convertPWT(x)
}

}
\author{
Lavinia Baumstark
}
