% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPYieldCoeff.R
\name{convertPYieldCoeff}
\alias{convertPYieldCoeff}
\title{Convert PYieldCoeff data to ISO country level.}
\usage{
convertPYieldCoeff(x)
}
\arguments{
\item{x}{MAgPIE object containing data for fixed regional resolution}
}
\value{
data as MAgPIE object disaggregated to country level
}
\description{
Convert PYieldCoeff data to ISO country level.
}
\examples{

\dontrun{ a <- convertPYieldCoeff(x)
}

}
\author{
Isabelle Weindl
}
