% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRam.R
\name{convertRam}
\alias{convertRam}
\title{Regional multiliers from Ram et al. , 2018}
\usage{
convertRam(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Regional multiliers from Ram et al. , 2018
}
