% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRCP.R
\name{convertRCP}
\alias{convertRCP}
\title{convertRCP
convert RCP data}
\usage{
convertRCP(x, subtype)
}
\arguments{
\item{x}{Input object obtained by readSource}

\item{subtype}{Either 'Waste' or 'AviationShipping'}
}
\value{
magpie object of the RCP data
}
\description{
convertRCP
convert RCP data
}
\author{
Julian Oeser
}
