% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertREMIND_11Regi.R
\name{convertREMIND_11Regi}
\alias{convertREMIND_11Regi}
\title{Converts REMIND regional data}
\usage{
convertREMIND_11Regi(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Name of the regional data, e.g. "p4", "biomass", "ch4waste", "tradecost", "pe2se", "xpres_tax", "deltacapoffset", capacityFactorRules", "taxConvergence", "maxFeSubsidy", "maxPeSubsidy", "propFeSubsidy", "fossilExtractionCoeff", "uraniumExtractionCoeff", "RLDCCoefficientsLoB", "RLDCCoefficientsPeak", "earlyRetirementAdjFactor"}
}
\value{
A MAgPIE object containing country disaggregated data
}
\description{
Converts REMIND regional data
}
\examples{

\dontrun{ a <- convertREMIND_11Regi(x,subtype="capacityFactorGlobal")
}

}
\author{
original: not defined - capacity factor, tax, fossil and RLDC changes: Renato Rodrigues
}
