% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRemindCesPrices.R
\name{convertRemindCesPrices}
\alias{convertRemindCesPrices}
\title{Convert RemindCesPrices}
\usage{
convertRemindCesPrices(x, subtype = "ccd632d33a")
}
\arguments{
\item{x}{MAgPIE object containing REMIND prices at the REMIND region resolution}

\item{subtype}{Regional resolution of REMIND data which should be loaded. ccd632d33a corresponds to the REMIND-11,
and 690d3718e1 to REMIND-H12}
}
\value{
magpie object of REMIND prices
}
\description{
Converts CES derivatives/prices from former REMIND runs to ISO level
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque
}
