% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertREN21.R
\name{convertREN21}
\alias{convertREN21}
\title{Policy targets for REN21}
\usage{
convertREN21(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Only "Capacity" as of now}
}
\value{
Magpie object with Total Installed Capacity targets. The target years
differ depending upon the database.
}
\description{
This code aggregates and homogenises different types of
renewable energy targets into total installed capacity targets (in GW).
}
\details{
Policy database accessible in "inputdata/sources/REN21/README"
}
\author{
Aman Malik
}
