% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertResFor2ndBE.R
\name{convertResFor2ndBE}
\alias{convertResFor2ndBE}
\title{convertResFor2ndBE}
\usage{
convertResFor2ndBE(x, subtype = subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{oldReMIND, newAgriSupply}
}
\value{
List of magpie objects with results on country level
}
\description{
Convert old ReMIND use of residues for 2nd generation bioenergy to country level data
}
\examples{

\dontrun{ 
  a <- readSource("ResFor2ndBE", subtype="oldReMIND", convert=TRUE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Kristine Karstens
}
