% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRogelj2017.R
\name{convertRogelj2017}
\alias{convertRogelj2017}
\title{Policy targets for NDCs from Rogelj2017}
\usage{
convertRogelj2017(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{UNCapacity, COCapacity}
}
\value{
Magpie object with Total Installed Capacity (GW) targets. The target years differ depending upon the database.
}
\description{
Converts conditional and unconditional capacity targets into total capacity (GW) in target year
the Generation targets are similar to the capacity targets but include the capacity factors,
the Emissions targets are the total (except land CO2) emissions in the target year
}
\author{
Aman Malik and Christoph Bertram
}
