% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRutovitz2015.R
\name{convertRutovitz2015}
\alias{convertRutovitz2015}
\title{convertRutovitz2015}
\usage{
convertRutovitz2015(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{"oecd_ef" or "regional_mult" or "regional_ef" or "coal_ef" or "gas_ef"}
}
\description{
convertRutovitz2015
}
\author{
Aman Malik
}
