% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSathayeForest.R
\name{convertSathayeForest}
\alias{convertSathayeForest}
\title{Convert Sathaye Forest data}
\usage{
convertSathayeForest(x)
}
\arguments{
\item{x}{MAgPIE object containing Sathaye Forest data region resolution}
}
\value{
Sathaye Forest data as MAgPIE object aggregated/disaggregated to
country level
}
\description{
Convert Sathaye Forest data on ISO country level.
}
\examples{

\dontrun{ a <- convertSathayeForest(x)
}

}
\author{
Lavinia Baumstark
}
