% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSRES.R
\name{convertSRES}
\alias{convertSRES}
\title{Convert SRES data}
\usage{
convertSRES(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing SRES data}

\item{subtype}{data subtype. "pop_a1","pop_a2","pop_b1","pop_b2" or
"gdp_a1","gdp_a2","gdp_b1","gdp_b2"}
}
\value{
SRES data as MAgPIE object aggregated to country level, all models,
all SRES-scenarios, GDP+Population. Units:
}
\description{
Convert SRES data on ISO country level.
}
\examples{

\dontrun{ a <- convertSRES(x,subtype="pop_a1")
}

}
\author{
Benjamin Leon Bodirsky
}
