% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSSP.R
\name{convertSSP}
\alias{convertSSP}
\title{Convert SSP data}
\usage{
convertSSP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing SSP data mixed country-region resolution}

\item{subtype}{data subtype. Either "all" or "ratioPM"}
}
\value{
SSP data as MAgPIE object aggregated to country level, all models,
all SSP-scenarios
}
\description{
Convert SSP data on ISO country level.
}
\examples{

\dontrun{ a <- convertSSP(x,subtype="all")
}
}
\author{
Lavinia Baumstark
}
