% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSSPResults.R
\name{convertSSPResults}
\alias{convertSSPResults}
\title{convertSSPResults}
\usage{
convertSSPResults(x)
}
\arguments{
\item{x}{object coming from read function}
}
\value{
MAgPIE object with ISO countries with all indicators for which disaggregation weight was found
}
\description{
Disagregates from SSP regions to ISO countries
}
\examples{

\dontrun{ 
readSource("SSPResults",aggregate=TRUE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra, Benjamin Leon Bodirsky, Florian Humpenoeder
}
