% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStegmann2022.R
\name{convertStegmann2022}
\alias{convertStegmann2022}
\title{convertStegmann2022}
\usage{
convertStegmann2022(x)
}
\arguments{
\item{x}{unconverted magpie object from read-script}
}
\value{
magpie object with a completed dataset.
}
\description{
Converts data from Stegmann2022
}
