% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStrefler.R
\name{convertStrefler}
\alias{convertStrefler}
\title{convertStrefler}
\usage{
convertStrefler(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}
}
\value{
magpie object with a completed dataset
}
\description{
Converts data on enhanced weathering
}
