% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTCdamage.R
\name{convertTCdamage}
\alias{convertTCdamage}
\title{Convert TCdamage
fills in countries not affected by tropical cyclones (TC), setting parameters to zero}
\usage{
convertTCdamage(x)
}
\arguments{
\item{x}{is MAgPIE object containing the damage parameters for the TC-prone countries}
}
\value{
MAgPIE object containing values for all 249 ISO countries
}
\description{
Convert TCdamage
fills in countries not affected by tropical cyclones (TC), setting parameters to zero
}
\author{
Franziska Piontek
}
