% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTransportSubisidies.R
\name{convertTransportSubsidies}
\alias{convertTransportSubsidies}
\title{Converts transport subsidies data}
\usage{
convertTransportSubsidies(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing transport subsidies per technology
}
\description{
Converts transport subsidies data
}
\examples{
\dontrun{
a <- convertTransportSubsidies(x)
}

}
\author{
Renato Rodrigues
}
