% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUBA.R
\name{convertUBA}
\alias{convertUBA}
\title{Convert UBA data}
\usage{
convertUBA(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass-package]{magpie}} object returned from
\code{\link[=readJRC_IDEES]{readJRC_IDEES()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Convert UBA data
}
\author{
Falk Benke
}
