% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNFCCC.R
\name{convertUNFCCC}
\alias{convertUNFCCC}
\title{Convert UNFCCC data}
\usage{
convertUNFCCC(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readUNFCCC]{readUNFCCC()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Convert UNFCCC data
}
\author{
Falk Benke
}
