% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNFCCC_NDC.R
\name{convertUNFCCC_NDC}
\alias{convertUNFCCC_NDC}
\title{Convert policy targets for NDCs from UNFCCC_NDC}
\usage{
convertUNFCCC_NDC(x, subtype, subset = NULL)
}
\arguments{
\item{x}{a magclass object to be converted}

\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets, Emissions_YYYY_cond or
Emissions_YYYY_uncond for Emissions targets, with YYYY NDC version year}

\item{subset}{String, designating the GDP scenarios to use. Only used for emission targets.}
}
\description{
Converts conditional and unconditional capacity and production targets into total capacity (GW) in target year.
For countries and years without targets, 2015 values from IRENA and BP are used to fill the gaps.
}
\details{
NDC Emissions targets on absolute level for total GHG emissions without bunkers and land-use change emissions are calculated
from country-specific target formulation and land-use change emissions data
}
\seealso{
\code{\link[=readIRENA]{readIRENA()}}
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Sophie Fuchs, Rahel Mandaroux, Falk Benke
}
