% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVanDrecht2009.R
\name{convertVanDrecht2009}
\alias{convertVanDrecht2009}
\title{convertVanDrecht2009}
\usage{
convertVanDrecht2009(x)
}
\arguments{
\item{x}{MAgPIE object containing incomplete country-region resolution}
}
\value{
A MAgPIE object containing sewage quantities and losses
}
\description{
Reads a dataset containing values for sewage
}
\examples{


  \dontrun{
    x <- readSource("VanDerWerf2010")
  }

}
\author{
Benjamin Leon Bodirksy
}
