% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWaste.R
\name{convertWaste}
\alias{convertWaste}
\title{convertWaste}
\usage{
convertWaste(subtype)
}
\arguments{
\item{subtype}{type of waste data, generation composition treatment or special}
}
\value{
Waste data as complete MAgPIE object on country level
}
\description{
Converts readWaste output to complete MAgPIE object containing Waste data on country level (kg/cap)
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
