% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWBirrigation.R
\name{convertWBirrigation}
\alias{convertWBirrigation}
\title{Convert WBirrigation data}
\usage{
convertWBirrigation(x)
}
\arguments{
\item{x}{MAgPIE object containing WBirrigation data country-region
resolution}
}
\value{
WBirrigation data as MAgPIE object aggregated to country level
}
\description{
Convert WorldBank-irrigation data on ISO country level.
}
\examples{

\dontrun{ a <- convertWBirrigation(x)
}

}
\author{
Lavinia Baumstark
}
