% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWFas.R
\name{convertWFaS}
\alias{convertWFaS}
\title{convertWFaS}
\usage{
convertWFaS(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing WFaS data on native level}

\item{subtype}{data subtype. available subtypes are:
\itemize{
\item dom_ww_ssp1_rcp4.5
\item dom_ww_ssp2_rcp6.0
\item dom_ww_ssp3_rcp6.0
\item dom_use_ssp1_rcp4.5
\item dom_use_ssp2_rcp6.0
\item dom_use_ssp3_rcp6.0
\item ind_ww_ssp1_rcp4.5
\item ind_ww_ssp2_rcp6.0
\item ind_ww_ssp3_rcp6.0
\item ind_use_ssp1_rcp4.5
\item ind_use_ssp2_rcp6.0
\item ind_use_ssp3_rcp6.0
\item dom_ww_ssp1_rcp4.5_gridded
\item dom_ww_tech_ssp1_rcp4.5_gridded
\item dom_use_ssp1_rcp4.5_gridded
\item dom_use_tech_ssp1_rcp4.5_gridded
\item ind_ww_ssp1_rcp4.5_gridded
\item ind_ww_tech_ssp1_rcp4.5_gridded
\item ind_use_ssp1_rcp4.5_gridded
\item ind_use_tech_ssp1_rcp4.5_gridded
\item dom_ww_ssp2_rcp2.6_gridded_2100
\item dom_ww_tech_ssp2_rcp2.6_gridded_2100
\item dom_use_ssp2_rcp2.6_gridded_2100
\item dom_use_tech_ssp2_rcp2.6_gridded_2100
\item ind_ww_ssp2_rcp2.6_gridded_2100
\item ind_ww_tech_ssp2_rcp2.6_gridded_2100
\item ind_use_ssp2_rcp2.6_gridded_2100
\item ind_use_tech_ssp2_rcp2.6_gridded_2100
\item dom_ww_ssp2_rcp4.5_gridded_2100
\item dom_ww_tech_ssp2_rcp4.5_gridded_2100
\item dom_use_ssp2_rcp4.5_gridded_2100
\item dom_use_tech_ssp2_rcp4.5_gridded_2100
\item ind_ww_ssp2_rcp4.5_gridded_2100
\item ind_use_ssp2_rcp4.5_gridded_2100
\item ind_use_tech_ssp2_rcp4.5_gridded_2100
\item dom_ww_ssp2_rcp6.0_gridded_2100
\item dom_ww_tech_ssp2_rcp6.0_gridded_2100
\item dom_use_tech_ssp2_rcp6.0_gridded_2100
\item ind_ww_ssp2_rcp6.0_gridded_2100
\item ind_ww_tech_ssp2_rcp6.0_gridded_2100
\item ind_use_ssp2_rcp6.0_gridded_2100
\item ind_use_tech_ssp2_rcp6.0_gridded_2100
\item dom_ww_ssp2_rcp6.0_gridded
\item dom_ww_tech_ssp2_rcp6.0_gridded
\item dom_use_ssp2_rcp6.0_gridded
\item dom_use_tech_ssp2_rcp6.0_gridded
\item ind_ww_ssp2_rcp6.0_gridded
\item ind_ww_tech_ssp2_rcp6.0_gridded
\item ind_use_ssp2_rcp6.0_gridded
\item ind_use_tech_ssp2_rcp6.0_gridded
\item dom_ww_ssp3_rcp6.0_gridded
\item dom_ww_tech_ssp3_rcp6.0_gridded
\item dom_use_ssp3_rcp6.0_gridded
\item dom_use_tech_ssp3_rcp6.0_gridded
\item ind_ww_ssp3_rcp6.0_gridded
\item ind_ww_tech_ssp3_rcp6.0_gridded
\item ind_use_ssp3_rcp6.0_gridded
\item ind_use_tech_ssp3_rcp6.0_gridded
\item dom_use_historic_gridded
\item dom_ww_historic_gridded
\item ind_use_historic_gridded
\item ind_ww_historic_gridded
}}
}
\value{
WFaS data as MAgPIE object for all subtypes on country level
}
\description{
Convert subtypes of the WFaS data

Convert subtypes on ISO country level.
}
\note{
Spatial Information of the gridded Data is lost. For detailed Information on Data the subtypes contain, see help file for readWFaS.
}
\examples{

\dontrun{ 
a <- readSource("WFaS","dom_ww_ssp1_rcp4.5")
a <- readSource("WFaS","dom_ww_ssp1_rcp4.5_gridded")
a <- readSource("WFaS","dom_use_tech_ssp2_rcp6.0_gridded_2100")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
