% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWGBU.R
\name{convertWGBU}
\alias{convertWGBU}
\title{Convert WGBU data}
\usage{
convertWGBU(x)
}
\arguments{
\item{x}{MAgPIE object containing WGBU data country-region resolution}
}
\value{
WGBU data as MAgPIE object aggregated to country level
}
\description{
Convert WGBU data on ISO country level.
}
\examples{
\dontrun{
a <- convertWGBU(x)
}

}
\author{
Lavinia Baumstark
}
