% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWHO.R
\name{convertWHO}
\alias{convertWHO}
\title{convertWHO}
\usage{
convertWHO(x)
}
\arguments{
\item{x}{unconverted magpie object from read-script}
}
\value{
magpie object with a completed dataset.
}
\description{
Converts data from the WHO
}
\seealso{
\code{\link{convertWHO}}
}
