% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWirsenius.R
\name{convertWirsenius}
\alias{convertWirsenius}
\title{Convert Wirsenius data}
\usage{
convertWirsenius(x)
}
\arguments{
\item{x}{MAgPIE object containing Wirsenius data country-region resolution}
}
\value{
Wirsenius data as MAgPIE object aggregated to country level
}
\description{
Convert Wirsenius data on ISO country level.
}
\examples{

\dontrun{ a <- convertWirsenius(x)
}

}
\author{
Lavinia Baumstark
}
