% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWirseniusPHD.R
\name{convertWirseniusPHD}
\alias{convertWirseniusPHD}
\title{Convert WirseniusPHD data}
\usage{
convertWirseniusPHD(x)
}
\arguments{
\item{x}{MAgPIE object containing Wirsenius PHD data country-region
resolution}
}
\value{
WirseniusPHD data as MAgPIE object aggregated to country level
}
\description{
Convert Wirsenius PHD data on ISO country level.
}
\examples{

\dontrun{ a <- convertWirseniusPHD(x)
}

}
\author{
Lavinia Baumstark
}
