% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGMIA.R
\name{correctGMIA}
\alias{correctGMIA}
\title{correctGMIA}
\usage{
correctGMIA(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing Global Map on Irrigaiton data data at 0.5 Degree resolution}

\item{subtype}{: subtypes are the same as in readGMIA}
}
\value{
Global Map on Irrigation data as MAgPIE object at a 0.5 Degree resolution.
}
\description{
Correct Irrigated Area

Correct Irrigated Area to 0.5 Degree x 0.5 Degree Grid.
Change resolution from 5 arcmin to 0.5 Degree by aggregating. Values in ha are summed up, Values in percent are calculated using mean.
}
\examples{

\dontrun{
a <- readSource("IrrigatedArea")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
