% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE-Industry.R
\name{calcSteel_Projections}
\alias{calcSteel_Projections}
\alias{calcIndustry_Value_Added}
\title{EDGE-Industry}
\usage{
calcSteel_Projections(
  subtype = "production",
  match.steel.historic.values = TRUE,
  match.steel.estimates = "none",
  save.plots = NULL,
  China_Production = NULL
)

calcIndustry_Value_Added(
  subtype = "physical",
  match.steel.historic.values = TRUE,
  match.steel.estimates = "none",
  save.plots = NULL,
  China_Production = NULL
)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{production} Returns trajectories of primary and secondary steel
production (\code{calcSteel_Projections()}).
\item \code{secondary.steel.max.share} Returns the maximum share of secondary steel
in total steel production (\code{calcSteel_Projections()}).
\item \code{physical} Returns physical production trajectories for cement
(\code{calcIndustry_Value_Added()}).
\item \code{economic} Returns value added trajectories for all subsectors
(\code{calcIndustry_Value_Added()}).
}}

\item{match.steel.historic.values}{Should steel production trajectories match
historic values?}

\item{match.steel.estimates}{Should steel production trajectories match
exogenous estimates?  \code{NULL} or one of
\itemize{
\item \code{IEA_ETP} IEA 2017 Energy Transition Pathways steel production totals for
OECD and Non-OECD countries from the \emph{Reference Technologies Scenario}
until 2060, and original growth rates after that.
}}

\item{save.plots}{\code{NULL} (default) if no plots are saved, or the path to save
directories to.}

\item{China_Production}{A data frame with columns \code{period} and
\code{total.production} prescribing total production for China to have,
disregarding results from the stock saturation model.}
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, \code{min}, and \code{max}.
}
\description{
Functions for calculating industry activity trajectories.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
