% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportThresholds.R
\name{exportThresholds}
\alias{exportThresholds}
\title{export validation thresholds}
\usage{
exportThresholds(type = "config", years = NULL)
}
\arguments{
\item{type}{choose either "config" to export thresholds as used in the
validationConfig or "full" to export all pipeline data}

\item{years}{choose years to include, currenty only 2025 and 2030 are
available, if NULL uses all available years}
}
\description{
assemble near-term thresholds from project pipelines and potentially other
data sources and export them to a file
}
\author{
Pascal Weigmann
}
