% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGFED.R
\name{extractGFED}
\alias{extractGFED}
\title{extractGFED}
\usage{
extractGFED(datatype = "hdf5", year, ResChange = T)
}
\arguments{
\item{datatype}{Specify which file type has to be read. Only "hdf5" is available at the moment. (Also set as default)}

\item{year}{Specify which year's data has to be read in from GFED database. Currently only years 1997-2015 available.}

\item{ResChange}{Specify if the resolution be changed from 0,25 degree cells to 0,5 degree cells. Default is True.}
}
\value{
magpie object of the GFED data
}
\description{
Aggregate GFED grid from 0,25 to 0,50
}
\examples{
\dontrun{ a <- readSource(type="12regions_baseline")
}
}
\seealso{
\code{\link{readSource}} {\link{readGFED}}
}
\author{
Abhijeet Mishra
}
