% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_historical_mif.R
\name{filter_historical_mif}
\alias{filter_historical_mif}
\title{Filter improperly aggregated regional data from historical.mif}
\usage{
filter_historical_mif(path = NULL, filter_table = NULL)
}
\arguments{
\item{path}{Path to \code{historical.mif} (or any mif for that matter), defaults
to \code{historical.mif} in the currently configured
\code{getConfig('outputfolder')}.}

\item{filter_table}{Data frame with columns \code{scenario}, \code{model}, \code{variable},
\code{include_regions}, and \code{exclude_regions}.  Defaults to the built-in
\code{./inst/extdata/historical_mif_filter_table.csv} of the \code{mrremind} package.}
}
\description{
Load a \code{historical.mif} file, remove designated data, write back to file.
}
\details{
For all combinations of scenario, model, and variable (and all years), either
all regions listed in \code{exclude_regions} in the \code{filter_table} are removed, or
all regions \emph{except} those listed in \code{include_regions} in the \code{filter_table}
are removed.
Defining both \code{include_regions} and \code{exclude_regions} on the same row of
\code{filter_table} will throw an error.
Scenarios, models, and variables are matched precisely, not partially or via
regular expressions.
}
