% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullEDGEBUILDINGS.R
\name{fullEDGEBUILDINGS}
\alias{fullEDGEBUILDINGS}
\title{fullEDGEBUILDINGS}
\usage{
fullEDGEBUILDINGS(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Function that produces the complete ISO data set required for the
EDGE - Buildings model.
}
\examples{

\dontrun{ 
fullEDGEBUILDINGS()
}
}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}}
}
\author{
Antoine Levesque
}
