% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMAGPIE_FEED.R
\name{fullMAgPIE_FEED}
\alias{fullMAgPIE_FEED}
\title{fullMAgPIE_FEED}
\usage{
fullMAgPIE_FEED(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Function that produces the complete country data set required for the
MAgPIE-FEED model (preprocessing).
}
\examples{

\dontrun{ 
fullMAgPIE_FEED()
}

}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Isabelle Weindl, Lavinia Baumstark
}
