% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullREMIND.R
\name{fullREMIND}
\alias{fullREMIND}
\title{fullREMIND}
\usage{
fullREMIND(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Function that produces the complete regional data set required for the
REMIND model.
}
\examples{

\dontrun{ 
fullREMIND()
}

}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}}
}
\author{
Lavinia Baumstark
}
