% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTHRESHOLDS.R
\name{fullTHRESHOLDS}
\alias{fullTHRESHOLDS}
\title{generate validation thresholds}
\usage{
fullTHRESHOLDS(rev = 0, type = "config", years = NULL)
}
\arguments{
\item{rev}{Unused parameter here for the pleasure of [`madrat`].}

\item{type}{choose either "config" to export thresholds as used in the
validationConfig or "full" to export all pipeline data}

\item{years}{choose years to include, currenty only 2025 and 2030 are
available, if NULL uses all available years}
}
\description{
assembles near-term thresholds from project pipelines and potentially other
data sources
}
\author{
Pascal Weigmann
}
