% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullVALIDATIONREMIND.R
\name{fullVALIDATIONREMIND}
\alias{fullVALIDATIONREMIND}
\title{fullVALIDATIONREMIND}
\usage{
fullVALIDATIONREMIND(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).}
}
\description{
Function that generates the historical regional dataset against which the
REMIND model results can be compared.
}
\examples{
\dontrun{
fullVALIDATIONREMIND()
}

}
\seealso{
\code{\link{fullREMIND}},\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}}
}
\author{
David Klein
}
