% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_subsectors_specific.R
\name{readindustry_subsectors_specific}
\alias{readindustry_subsectors_specific}
\alias{calcindustry_subsectors_specific}
\title{industry/subsector change factors}
\usage{
readindustry_subsectors_specific(subtype = NULL)

calcindustry_subsectors_specific(
  subtype = NULL,
  scenarios = NULL,
  regions = NULL,
  direct = NULL
)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{FE} for specific final energy demand change factors
\item \code{material_alpha} for alpha factors and convergence time of specific
material demand decreases relative to the \code{SSP2EU} scenario
\item \code{material_relative} for scaling factors of specific material demand
relative to baseline scenarios
\item \code{material_relative_change} for scaling factors of specific material
demand \emph{change} relative to baseline scenarios
}}

\item{scenarios}{A vector of scenarios for which factors are to be returned.}

\item{regions}{A vector of regions for which factors are to be returned.}

\item{direct}{A data frame as returned by
\code{readindustry_subsectors_specific()} to load debugging/developing data
directly instead of from file.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Change factors of specific FE and material demand for the
\code{industry/subsector} realisation of REMIND.
}
\details{
Factors are read from the files \code{specific_FE.csv},
\code{specific_material_alpha.csv}, \code{specific_material_relative.csv}, and
\code{specific_material_relative_change.csv}, respectively.  \code{NA} is used to mark
defaults for the \code{scenario} and \code{region} columns, and specified values will
overwrite these defaults.

So
\itemize{
\item \verb{NA,NA,cement,1} will be extended to all \code{scenarios} and \code{regions}
\item \verb{scen1,NA,cement,2} will overwrite this default for all \code{regions} in
\code{scen1}
\item \verb{NA,regi1,cement,3} will overwrite this again for all \code{scenarios}
(including \code{scen1}) for \code{regi1}
\item \verb{scen1,regi1,cement,4} will lastly overwrite the value for the \code{scen1},
\code{regi1} combination
}

Replacements occure in this fixed order (\code{NA}/\code{NA}, \code{scenario}/\code{NA},
\code{NA}/\code{region}, \code{scenario}/\code{region}).

Lastly, output is filtered for \code{scenarios} and \code{regions}.

For debugging and development, instead of modifying the .csv files in
\verb{sources/industry_subsectors_specific/} and interfering with production runs,
modify the calling code (e.g. \code{calcFEdemand.R}) to use \code{direct} data (entered
verbatim or loaded from somewhere else.)
}
\author{
Michaja Pehl
}
