% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moinput-package.R
\docType{package}
\name{moinput-package}
\alias{moinput-package}
\alias{moinput}
\title{Model Operations Input Data Library}
\description{
Package contains basic functions of the former landuse library.
}
\details{
\tabular{ll}{ Package: \tab moinput\cr Type: \tab Package\cr Version: \tab
5.55\cr Date: \tab 2016-09-23\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Lavinia Baumstark, Anastasis Giannousakis, Jan Philipp Dietrich,
Benjamin Bodirsky, Mishko Stevanovic, Isabelle Weindl

Maintainer: Anastasis Giannousakis <giannou@pik-potsdam.de>
}
