% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrremind-package.R
\docType{package}
\name{mrremind-package}
\alias{mrremind}
\alias{mrremind-package}
\title{mrremind: MadRat REMIND Input Data Package}
\description{
The mrremind packages contains data preprocessing for the REMIND model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrremind}
}

}
\author{
\strong{Maintainer}: Lavinia Baumstark \email{lavinia@pik-potsdam.de}

Authors:
\itemize{
  \item Renato Rodrigues
  \item Antoine Levesque
  \item Julian Oeser
  \item Christoph Bertram
  \item Ioanna Mouratiadou
  \item Aman Malik
  \item Felix Schreyer
  \item Bjoern Soergel
  \item Marianna Rottoli
  \item Abhijeet Mishra
  \item Alois Dirnaichner
  \item Michaja Pehl
  \item Anastasis Giannousakis
  \item David Klein
  \item Jessica Strefler
  \item Lukas Feldhaus
  \item Regina Brecha
  \item Sebastian Rauner
  \item Jan Philipp Dietrich
  \item Stephen Bi
  \item Falk Benke
  \item Pascal Weigmann
  \item Oliver Richters
  \item Robin Hasse
  \item Sophie Fuchs
  \item Rahel Mandaroux
  \item Johannes Koch
  \item Gabriel Abrahao
  \item Tabea Dorndorf
  \item Simon Krogmann
}

}
