% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readACCMIP.R
\name{readACCMIP}
\alias{readACCMIP}
\title{readACCMIP}
\usage{
readACCMIP(subtype = NULL)
}
\arguments{
\item{subtype}{data subtype. available subtypes are:
\itemize{
\item nhx_1850
\item noy_1850
\item sox_1850
\item nhx_1980
\item noy_1980
\item sox_1980
\item nhx_2000
\item noy_2000
\item sox_2000
\item nhx_26_2030
\item nhx_45_2030
\item nhx_85_2030
\item noy_26_2030
\item noy_45_2030
\item noy_85_2030
\item sox_26_2030
\item sox_45_2030
\item sox_85_2030
\item nhx_26_2100
\item nhx_45_2100
\item nhx_85_2100
\item noy_26_2100
\item noy_45_2100
\item noy_85_2100
\item sox_26_2100
\item sox_45_2100
\item sox_85_2100
}}
}
\value{
magpie object of the ACCMIP data. Unit is t NO2-N  per ha per year, or t NH3-N  per ha per year,...
}
\description{
Read in data from the Atmospheric Chemistry and Climate Model Intercomparison Project
}
\examples{

\dontrun{ 
a <- readACCMIP("ACCMIP","nhx_2000")
a <- readACCMIP("ACCMIP","sox_26_2030")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Roman Popov
}
