% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAmmoniaProductionUSGS.R
\name{readAmmoniaProductionUSGS}
\alias{readAmmoniaProductionUSGS}
\alias{convertAmmoniaProductionUSGS}
\title{readAmmoniaProductionUSGS}
\usage{
readAmmoniaProductionUSGS()
}
\value{
A MAgPIE-Object containing global Ammonia production from the USGS
}
\description{
Function to read the the global Ammonia Production from USGS Website
\url{https://minerals.usgs.gov/minerals/pubs/commodity/nitrogen/}
}
\examples{
\dontrun{
a <- readSouce("AmmoniaProductionUSGS")
}
}
