% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readARIADNE_ReferenceScenario.R
\name{readARIADNE_ReferenceScenario}
\alias{readARIADNE_ReferenceScenario}
\title{Read ARIADNE Reference Scenario}
\usage{
readARIADNE_ReferenceScenario(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "population", "gdp", or "gdp_corona"}
}
\value{
magpie object of ARIADNE reference scenario data by country
}
\description{
Read ARIADNE Reference Scenario data from various .xls files as magpie object
}
\author{
Falk Benke
}
