% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAriadneDB.R
\name{readAriadneDB}
\alias{readAriadneDB}
\title{Ariadne database scenario data}
\usage{
readAriadneDB()
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
This reads in FORECAST industry production data for Germany used in the Ariadne scenarios
}
\author{
Felix Schreyer
}
