% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASTI.R
\name{readASTI}
\alias{readASTI}
\title{Read ASTI R&D investment spreadsheet}
\usage{
readASTI()
}
\value{
ASTI R&D data as MAgPIE object
}
\description{
Read R&D investment data from ASTI database published by IFPRI
}
\examples{

\dontrun{ a <- readSource("ASTI")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Xiaoxi Wang
}
