% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBGR.R
\name{readBGR}
\alias{readBGR}
\title{Read BGR oil, gas, coal and uranium reserves data}
\usage{
readBGR(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "oil", "gas", "coal" or "uranium".}
}
\value{
magpie object of the BGR (Federal Institute for Geosciences and
Natural Resources) data of reserves of oil, gas, coal and uranium per country.
}
\description{
Read-in BGR csv files as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "BGR", subtype = "oil")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
