% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBijl2017.R
\name{readBijl2017}
\alias{readBijl2017}
\title{Read Bijl2017}
\usage{
readBijl2017(subtype)
}
\arguments{
\item{subtype}{data subtype.}
}
\value{
magpie object of the Bijl et. al 2017 data
}
\description{
Read-in the data from Bijl 2017 food demand paper (http://dx.doi.org/10.1016/j.gloenvcha.2017.04.003)
}
\examples{

\dontrun{ a <- readSource("Bijl2017","demand")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
