% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiocharDeploymentData.R
\name{readBiocharDeploymentData}
\alias{readBiocharDeploymentData}
\title{Load data for implementation of biochar as magclass object}
\usage{
readBiocharDeploymentData()
}
\description{
Source:
Assumption for REMIND based on data collected in Dorndorf et al (submitted)
}
\author{
Tabea Dorndorf
}
