% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBodirsky.R
\name{readBodirsky}
\alias{readBodirsky}
\title{Read Bodirsky}
\usage{
readBodirsky()
}
\value{
magpie object of the Bodirsky data
}
\description{
Read-in an Bodirsky data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="Bodirsky")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
