% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBodirsky2018.R
\name{readBodirsky2018}
\alias{readBodirsky2018}
\title{readBodirsky2018}
\usage{
readBodirsky2018(subtype = "bmi_shr")
}
\arguments{
\item{subtype}{bmi_share for the estimated regression paramters for bmi shares.}
}
\value{
magpie object
}
\description{
Reads in regression parameters estimated using mrregression
not yet published
}
\seealso{
\code{\link{readNCDrisc}}
}
