% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBP.R
\name{readBP}
\alias{readBP}
\title{BP Capacity and Generation Data}
\usage{
readBP(subtype)
}
\arguments{
\item{subtype}{Either "Emission", Capacity", "Generation", "Production",
"Consumption", "Trade Oil", "Trade Gas", "Trade Coal" or "Price"}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
BP Capacity and Generation Data
}
\author{
Aman Malik, Falk Benke, Pascal Weigmann
}
