% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBraakhekke2017.R
\name{readBraakhekke2017}
\alias{readBraakhekke2017}
\title{readBraakhekke2017}
\usage{
readBraakhekke2017(subtype = "soilN")
}
\arguments{
\item{subtype}{subtype filename without NC}
}
\value{
magpie object
}
\description{
reads in LPJguess results for nitrogen runs from the publication 
Braakhekke, M. C. et al. Nitrogen leaching from natural ecosystems under global change: a modelling study. Earth Syst. Dynam. 8, 1121–1139 (2017).
}
\seealso{
\code{\link{readLPJml_rev21}}
}
