% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCEEW.R
\name{readCEEW}
\alias{readCEEW}
\title{Read Employment factors and cumulative jobs for RE techs (for India)}
\usage{
readCEEW(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "Employment factors" or "Employment"}
}
\description{
Read Employment factors and cumulative jobs for RE techs (for India)
}
\details{
Reports published by CEEW et al. See README.txt in the source folder for more information.
}
\examples{
 \dontrun{ 
a <- readSource("CEEW",convert=F,subtype="Employment")
}
}
\author{
Aman Malik
}
