% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCEMOdb.R
\name{readCEMOdb}
\alias{readCEMOdb}
\title{read-in power technology parameters for Australia from CEMO model database
Australian contact: Dylan McConnell, dylan.mcconnell(at)unimelb.edu.au}
\usage{
readCEMOdb()
}
\value{
magpie object of the cemo database data
}
\description{
read-in power technology parameters for Australia from CEMO model database
Australian contact: Dylan McConnell, dylan.mcconnell(at)unimelb.edu.au
}
\seealso{
\code{\link{readSource}}
}
\author{
Felix Schreyer
}
