% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConferenceBoard.R
\name{readConferenceBoard}
\alias{readConferenceBoard}
\title{readConferenceBoard}
\usage{
readConferenceBoard()
}
\value{
A magpie object
}
\description{
Employment and GDP-related data from Conference Board https://www.conference-board.org/data/economydatabase/total-economy-database-productivity
}
\author{
Aman Malik
}
