% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSIRO.R
\name{readCSIRO}
\alias{readCSIRO}
\title{read-in CSIRO electricity technology cost for Australia}
\usage{
readCSIRO()
}
\value{
magpie object of the CSIRO technology data
}
\description{
read-in CSIRO electricity technology cost for Australia
}
\seealso{
\code{\link{readSource}}
}
\author{
Felix Schreyer
}
