% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDaviesCooling.R
\name{readDaviesCooling}
\alias{readDaviesCooling}
\title{Read Davies Cooling}
\usage{
readDaviesCooling(subtype)
}
\arguments{
\item{subtype}{Type of Davies data that should be read. Available types are:
\itemize{ \item \code{dataBase}: The Davies source data for the base year
\item \code{dataFuture}: The Daves source data for the future }}
}
\value{
MAgPIE object of the Davies (2013) data
}
\description{
Read in Davies (2013) data on shares of cooling types per electricity
technology and GCAM region
}
\examples{
\dontrun{
a <- readSource(type = "DaviesCooling")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark, Ioanna Mouratiadou
}
