% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDemandModel.R
\name{readDemandModel}
\alias{readDemandModel}
\title{Read DemandModel}
\usage{
readDemandModel()
}
\value{
magpie object of the DemandModel data
}
\description{
Read-in an csv that contains data from a DemandModel as magclass object
}
\examples{

\dontrun{ a <- readSource(type="DemandModel")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
