% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDylanAusGasCost.R
\name{readDylanAusGasCost}
\alias{readDylanAusGasCost}
\title{read-in power Australian gas extraction cost curve based on Dylan's data
Australian contact: Dylan McConnell, dylan.mcconnell(at)unimelb.edu.au}
\usage{
readDylanAusGasCost()
}
\value{
magpie object of the cemo database data
}
\description{
read-in power Australian gas extraction cost curve based on Dylan's data
Australian contact: Dylan McConnell, dylan.mcconnell(at)unimelb.edu.au
}
\author{
Felix Schreyer
}
