% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR7Fgases.R
\name{readEDGAR7Fgases}
\alias{readEDGAR7Fgases}
\title{Read EDGAR7 emissions data for F-gases per species, in kt of each gas}
\usage{
readEDGAR7Fgases()
}
\value{
A magpie object with F-gases emissions per gas species and per country
}
\description{
Read EDGAR7 emissions data for F-gases per species, in kt of each gas
}
\author{
Gabriel Abrahao
}
