% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGE.R
\name{readEDGE}
\alias{readEDGE}
\title{Read EDGE}
\usage{
readEDGE(subtype = "FE_stationary")
}
\arguments{
\item{subtype}{FE for final energy or Capital for capital projections}
}
\value{
magpie object of EDGE
}
\description{
Read-in an EDGE csv file as magclass object
}
\examples{
\dontrun{ a <- readSource(type="EDGE")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque
}
