% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGE_Industry.R
\name{readEDGE_Industry}
\alias{readEDGE_Industry}
\title{Read EDGE Industry Data}
\usage{
readEDGE_Industry(subtype = NULL)
}
\arguments{
\item{subtype}{for now only \code{projections_VA_iso3c}}
}
\value{
a data frame
}
\description{
Read EDGE Industry data
}
\seealso{
\code{\link{readSource}}
}
\author{
Michaja Pehl
}
