% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEdgeBuildings.R
\name{readEdgeBuildings}
\alias{readEdgeBuildings}
\title{Load an EDGE Buildings file as magclass object.}
\usage{
readEdgeBuildings(subtype = c("FE", "Floorspace"), subset)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}

\item{subset}{A string (or vector of strings) designating the scenario(s)
to be returned (needed in 'convertEdgeBuildings').}
}
\value{
magclass object
}
\description{
Load an EDGE Buildings file as magclass object.
}
\author{
Antoine Levesque, Robin Hasse
}
