% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGEtranspLDV.R
\name{readEDGEtranspLDV}
\alias{readEDGEtranspLDV}
\title{Read EDGE Transport LDV shares}
\usage{
readEDGEtranspLDV()
}
\value{
magpie object of EDGE_TranspLDV
}
\description{
Read-in EDGE Transport LDV shares csv file as magclass object
}
\examples{
\dontrun{ a <- readSource(type="EDGEtranspLDV")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
