% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEEA_EuropeanEnvironmentAgency.R
\name{readEEA_EuropeanEnvironmentAgency}
\alias{readEEA_EuropeanEnvironmentAgency}
\title{Read European Environment Agency (EEA) data}
\usage{
readEEA_EuropeanEnvironmentAgency(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "ETS", "ESR", "total", "sectoral", "projections", or "projections-detailed"}
}
\value{
magpie object of European Environment Agency (EEA) ETS emissions (GtCO2)
}
\description{
Read-in European Environment Agency (EEA) data on ETS emissions as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "EEA_EuropeanEnvironmentAgency", subtype = "ETS")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues, Falk Benke
}
