% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEmber.R
\name{readEmber}
\alias{readEmber}
\title{Read Ember Yearly Electricity Data}
\source{
https://ember-climate.org/data-catalogue/yearly-electricity-data/
}
\usage{
readEmber()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Read Ember Yearly Electricity Data
}
\seealso{
[`readSource()`]
}
\author{
Pascal Weigmann
}
