% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEndCoal.R
\name{readEndCoal}
\alias{readEndCoal}
\title{Data on Coal Plants by country (in MW) from EndCoal.org (Actual source: Coal Swarm
and Global Plant Tracker) July 2018}
\usage{
readEndCoal()
}
\description{
Data on Coal Plants by country (in MW) from EndCoal.org (Actual source: Coal Swarm
and Global Plant Tracker) July 2018
}
\author{
Aman Malik
}
