% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEORA.R
\name{readEORA}
\alias{readEORA}
\title{readEORA}
\usage{
readEORA(subtype)
}
\arguments{
\item{subtype}{- year (1990-2013)}
}
\value{
MAgPIE object of the EORA data
}
\description{
Read in EORA data of one year
}
\note{
With about 15000*15000 entries the EORA database is extremely big. Therefore, it wasn't possible
to create a MAgPIE object containing the full information. Instead every value is rounded to the nearest
250 000 USD.
}
\seealso{
\code{\link{readSource}}, \code{\link{calcEORA}}, \code{\link{calcCommoditiesEORA}}
}
\author{
Debbora Leip
}
