% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readETH_COUN.R
\name{readETH_COUN}
\alias{readETH_COUN}
\title{#' This function reads WACC markup data across regions and periods,
computing the average value across technologies for each region and period
as a proxy for the macroeconomic cost of capital for each region}
\usage{
readETH_COUN()
}
\description{
#' This function reads WACC markup data across regions and periods,
computing the average value across technologies for each region and period
as a proxy for the macroeconomic cost of capital for each region
}
\author{
Diamantis Koutsandreas
}
