% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readETH_WACC_b.R
\name{readETH_WACC}
\alias{readETH_WACC}
\title{This function reads WACC markup data across technologies for each of
the REMIND regions}
\usage{
readETH_WACC()
}
\description{
This function reads WACC markup data across technologies for each of
the REMIND regions
}
\author{
Diamantis Koutsandreas
}
