% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanEnergyDatasheets.R
\name{readEuropeanEnergyDatasheets}
\alias{readEuropeanEnergyDatasheets}
\title{Read European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/energy-statistical-pocketbook
}
\usage{
readEuropeanEnergyDatasheets()
}
\value{
magpie object of aggregated energy market data by country. Units follow REMIND report conventions and conversion factor is defined in mapping2REMIND.xlsx file.
}
\description{
Read European Energy Datasheets .xlsx file as magpie object
}
\examples{
\dontrun{test <- readSource("EuropeanEnergyDatasheets",convert=FALSE)}
}
\author{
Renato Rodrigues and Atreya Shankar
}
