% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat_EffortSharing.R
\name{readEurostat_EffortSharing}
\alias{readEurostat_EffortSharing}
\title{Read EU Effort Sharing targets and historical emissions}
\usage{
readEurostat_EffortSharing(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "target" or "emissions"}
}
\value{
magpie object of the EU Effort Sharing targets (%) or Effort Sharing historical historical emissions (GtCO2)
}
\description{
Read-in EU Effort Sharing targets and historical emissions csv files as magclass object
}
\examples{

\dontrun{ a <- readSource(type="Eurostat_EffortSharing",subtype="target")
}
 
}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
