% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat_REShare.R
\name{readEurostat_REShare}
\alias{readEurostat_REShare}
\title{Read Eurostat Historical Renewable FE Share Data}
\usage{
readEurostat_REShare()
}
\value{
magpie object of Eurostat Historical Renewable FE Share Data (share)
}
\description{
Read Eurostat Historical Renewable FE Share Data as magclass object
}
\author{
Felix Schreyer
}
