% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpertGuess.R
\name{readExpertGuess}
\alias{readExpertGuess}
\title{Read ExpertGuess}
\usage{
readExpertGuess(subtype)
}
\arguments{
\item{subtype}{Type of data that should be read.  One of
\itemize{
\item \code{Chinese_Steel_Production}: "Smooth" production estimates by Robert
Pietzcker (2022).
\item \code{industry_max_secondary_steel_share}: Maximum share of secondary steel
production in total steel production and years between which a linear
convergence from historic to target shares is to be applied.
\item \code{cement_production_convergence_parameters}: convergence year and level
(relative to global average) to which per-capita cement demand converges
\item Any of the others nobody cared to document.
}}
}
\value{
magpie object of the data
}
\description{
Read-in data that are based on expert guess
}
\examples{

\dontrun{ a <- readSource(type="ExpertGuess",subtype="ies")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
