% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_forestry.R
\name{readFAO_forestry}
\alias{readFAO_forestry}
\title{Read data of forest}
\usage{
readFAO_forestry(subtype)
}
\arguments{
\item{subtype}{Type of FAO data that should be read. Available types are:
\itemize{ \item \code{protection}: forest within protected areas
(protection.csv) that has been downloaded from the CountrySTAT website. File
is aquired from: http://countrystat.org/home.aspx?c=FOR&tr=2 \item
\code{production}: forest designated for production (production.csv) that
has been downloaded from the CountrySTAT website. File is aquired from:
http://countrystat.org/home.aspx?c=FOR&tr=2 \item \code{extent}: forest
extent (forest_extent.csv) that has been downloaded from the CountrySTAT
website. File is aquired from: http://countrystat.org/home.aspx?c=FOR&tr=1 }}
}
\value{
FAO data of forest as MAgPIE object
}
\description{
Read in data of forest
}
\seealso{
\code{\link{readSource}}
}
\author{
Nele Steinmetz
}
