% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_WHO_UNU1985.R
\name{readFAO_WHO_UNU1985}
\alias{readFAO_WHO_UNU1985}
\title{Read parameters of Schofield equations}
\usage{
readFAO_WHO_UNU1985()
}
\value{
MAgPIE object
}
\description{
Food and Agriculture Organization of the United Nations, World Health Organization, and United Nations University. 1985. "Energy and protein requirements." http://www.who.int/iris/handle/10665/39527.
}
\examples{

\dontrun{ a <- readSource("Schofield") }

}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
