% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOfoodpriceindex.R
\name{readFAOfoodpriceindex}
\alias{readFAOfoodpriceindex}
\title{readFAOfoodpriceindex}
\usage{
readFAOfoodpriceindex()
}
\value{
Magpie object with results on global level.
}
\description{
reads in global food price index from the FAO, downloaded from http://www.fao.org/worldfoodsituation/foodpricesindex/en/
}
\examples{

\dontrun{ 
readSource("FAOfoodpriceindex",convert=FALSE)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky
}
