% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOLossesWaste.R
\name{readFAOLossesWaste}
\alias{readFAOLossesWaste}
\title{Read in data on  food losses and waste from FAO for several commodity groups}
\usage{
readFAOLossesWaste(subtype)
}
\arguments{
\item{subtype}{Steps of the food supply chain where food losses and waste occur. Available types are:
\itemize{ 
\item \code{Consumption}: consumption level
}}
}
\value{
magpie object of food waste percentages for several commodity groups
}
\description{
Data from Annex 4 of the following FAO study:
FAO. 2011. Global food losses and food waste – Extent, causes and prevention. Rome 
(http://www.fao.org/3/a-i2697e.pdf)
}
\examples{

\dontrun{ a <- readSource(type="FAOLossesWaste",subtype="Consumption")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
