% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeedModel.R
\name{readFeedModel}
\alias{readFeedModel}
\title{Read in data of FeedModel}
\usage{
readFeedModel(subtype = "FeedBaskets")
}
\arguments{
\item{subtype}{Available subtypes: "ProdSysRatio", "FeedBaskets" and "FeedBasketsDetailed"}
}
\value{
magpie object of feed basket data
}
\description{
Read in csv files containing data on production system distribution and system-specific feed baskets from the FeedModel "MAgPIE_FEED"
}
\examples{

\dontrun{ a <- readSource(type="FeedModel",subtype="FeedBaskets")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
