% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeedShareReg.R
\name{readFeedShareReg}
\alias{readFeedShareReg}
\title{Read in regression coefficients for central feed shares}
\usage{
readFeedShareReg()
}
\value{
MAgPIE object containing regression coefficients
}
\description{
Read in csv file containing coefficients of non-linear regression analysis
for the calculation of future central feed shares in feed baskets dependent 
on livestock productivity trends
}
\examples{
\dontrun{
a <- readSource("FeedShareReg")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
