% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFRA2015Doc.R
\name{readFRA2015Doc}
\alias{readFRA2015Doc}
\title{Read FRA2015Doc}
\usage{
readFRA2015Doc(subtype)
}
\arguments{
\item{subtype}{data subtype.}
}
\value{
magpie object of the FRA 2015 data
}
\description{
Read-in an FRA data from 2015 (forest resource assessment) pdf file.
}
\examples{

\dontrun{ a <- readSource("FRA2015Doc","forest_area")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
