% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFroehle.R
\name{readFroehle}
\alias{readFroehle}
\title{Read parameters of Froehle equations}
\usage{
readFroehle()
}
\value{
MAgPIE object
}
\description{
Publication Froehle, Andrew W. 2008 "Climate Variables as Predictors of Basal Metabolic Rate: New Equations" American Journal of Human Biology: The Official Journal of the Human Biology Council 20 (5): 510-29. doi: 10.1002/ajhb.20769.
}
\examples{

\dontrun{ a <- readSource("Froehle") }

}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
