% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGAINS2025.R
\name{readGAINS2025}
\alias{readGAINS2025}
\title{Read air pollution emissions, activities and emission factors from GAINS data}
\usage{
readGAINS2025(subtype, subset = "baseline.det")
}
\arguments{
\item{subtype}{"emission_factors", "emissions","emissions_starting_values"}

\item{subset}{Subset of the data to read, split with a dot (".") into
scenario and aggregation level.
scenario can be "baseline" for the historical baseline, "cle_rev" (current legislation),
"mid" (stronger legislation) or "mtfr" (maximum technically feasible reduction).
The aggregation level can be "det" (detailed) or "agg" (aggregated).}
}
\value{
Activity levels, emissions or emission factors. Alternatively,
a Govenrnment Capacity Index (GCI) used for deriving some scenario
extensions.
}
\description{
There's no associated convert function, as the disaggregation takes
a combination of subtypes, and it is easier to carry out most
calculations at the GAINS regional level first and then disaggregate
the results
}
\author{
Gabriel Abrahao
}
