% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGAINS2025final.R
\name{readGAINS2025final}
\alias{readGAINS2025final}
\title{Read and combine air pollution emissions, activities and emission factors from GAINS data}
\usage{
readGAINS2025final(subtype)
}
\arguments{
\item{subtype}{"emifacs", "emissions", "activities"}
}
\value{
Activity levels, emissions or emission factors at the level of
25 GAINS regions, 35 GAINS sectors and 7 species:
magclass object with dimensions region, year, and ssp.scenario.sectorGAINS.species
}
\description{
There's no associated convert function, as the disaggregation takes
a combination of subtypes, and it is easier to carry out most
calculations at the GAINS regional level first and then disaggregate
the results
}
\author{
Gabriel Abrahao, Laurin Koehler-Schindler
}
