% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCPT.R
\name{readGCPT}
\alias{readGCPT}
\alias{readEndCoal}
\title{Data from the Global Coal Plant Tracker January 2021 release by Global Energy Monitor (formerly EndCoal/CoalSwarm)}
\usage{
readGCPT(subtype)
}
\arguments{
\item{subtype}{Options are status, historical, future, lifespans, comp_rates and emissions}
}
\description{
Historical data of operating, under-construction, planned and announced Coal Plants by country (in MW) 
from the Global Energy Monitor's Global Coal Plant Tracker, and extrapolations for 2025 capacity scenarios
}
\author{
Stephen Bi
}
