% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCPT2020_extra.R
\name{readGCPT2020_extra}
\alias{readGCPT2020_extra}
\title{Data from the Global Coal Plant Tracker July 2020 release by Global Energy Monitor (formerly EndCoal/CoalSwarm)}
\usage{
readGCPT2020_extra(subtype)
}
\arguments{
\item{subtype}{Options are status, historical, future, retirements}
}
\description{
Historical data of operating, under-construction, planned and announced Coal Plants by country (in MW) 
from the Global Energy Monitor's Global Coal Plant Tracker, and extrapolations for 2025 capacity scenarios
}
\author{
Stephen Bi
}
