% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEA2012.R
\name{readGEA2012}
\alias{readGEA2012}
\title{read GEA 2012}
\usage{
readGEA2012(subtype)
}
\arguments{
\item{subtype}{Type of fossil fuel and type of data (oil, coal, or gas + costs, qtys, or dec)}
}
\value{
MAgPIE object of the GEA data
}
\description{
Read in datafiles comprising fossil fuel data from the Global Energy Assessment 2012
}
\examples{
\dontrun{
a <- readSource("GEA2012", "coal")
}

}
\author{
Stephen Bi
}
