% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGernaat.R
\name{readGernaat}
\alias{readGernaat}
\title{Read Hydro-Potential data from Gernaat et. al 2017}
\source{
Obtained from Nature Energy volume 2 (2017), pp 821–828, with following source: https://www.nature.com/articles/s41560-017-0006-y
}
\usage{
readGernaat()
}
\value{
magpie object of aggregated hydropotential data from Gernaat et. al 2017
}
\description{
Read privileged csv file sent by David Gernaat for global hydropotential data. Technical potential 
defined as <0.5 $/kWh and economic potential as <0.1 $/kWh.
}
\examples{
\dontrun{a <- readGernaat()}
}
\author{
Atreya Shankar
}
