% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFED.R
\name{readGFED}
\alias{readGFED}
\title{Read GFED}
\usage{
readGFED(subtype)
}
\arguments{
\item{subtype}{Currently available "12regions_baseline", "emissions" and "emissionfactors". "emissionfactors" reads the different chemical ploutants provided in the GFED data.}
}
\value{
magpie object of the GFED data
}
\description{
Read-in an GFED file as magclass object
}
\examples{

\dontrun{ a <- readSource("GFED","emissions")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra, Lavinia Baumstark
}
