% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGGDC10.R
\name{readGGDC10}
\alias{readGGDC10}
\title{Read GGDC 10-Sector Database - https://www.rug.nl/ggdc/structuralchange/previous-sector-database/10-sector-2014}
\usage{
readGGDC10()
}
\description{
Read GGDC 10-Sector Database - https://www.rug.nl/ggdc/structuralchange/previous-sector-database/10-sector-2014
}
\examples{
 \dontrun{
a <- readSource("GGDC10",convert=F)
}

}
\author{
Renato Rodrigues
}
