% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGidden2025_geological_storage_potential.R
\name{readGidden2025_geological_storage_potential}
\alias{readGidden2025_geological_storage_potential}
\title{Read geological storage potential}
\usage{
readGidden2025_geological_storage_potential()
}
\value{
A magpie object with geological storage potential. Off = offshore; On = onshore; potTech = technical potential
without any exclusion layers applied; potLim = applying all exclusion layers described in Table S1 (e.g. protected
areas, population centers, max and min depth, etc.)
}
\description{
Read geological storage potential
}
\author{
David Klein
}
