% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGini.R
\name{readGini}
\alias{readGini}
\alias{convertGini}
\title{Read Gini}
\usage{
readGini()

convertGini(x)
}
\arguments{
\item{x}{MAgPIE object returned from readGini}
}
\value{
magpie object of the Gini data
}
\description{
Read Gini coefficients for SSP scenarios from Rao et al., Futures, 2018. Data has been provided by the authors, but
will be made publicly available as well. This contains data for 184 countries and from 2011 onwards.
}
\details{
Copied from the documentation provided by the authors:
This sheet contains the original Gini projections for 43 countries from the underlying empirical model
(See reference to RSP 2016 in the main paper) and the extrapolations to all countries using the methodology
described in the article. The country codes are the World Bank codes.
}
\examples{
\dontrun{
a <- readSource(type="Gini")
}
}
\seealso{
\code{\link{readSource}}
}
