% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalCCSinstitute.R
\name{readGlobalCCSinstitute}
\alias{readGlobalCCSinstitute}
\title{Read Global CCS Institute Project Database}
\usage{
readGlobalCCSinstitute(subtype = "08-09-2017")
}
\arguments{
\item{subtype}{Project Database version to read, one of
- \code{'08-09-2017'}: Data apparently from June 2017.
- \code{'2023-11'}: Data from the
\href{zotero://select/items/3_E5GNNPZ8}{Global Status of CCS 2023} report.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read Global CCS Institute Project Database
}
