% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalMethanePledge.R
\name{readGlobalMethanePledge}
\alias{readGlobalMethanePledge}
\title{Countries that commited to the Global Methane Pledge by 23.02.2025}
\usage{
readGlobalMethanePledge()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Data scraped from the map at https://www.globalmethanepledge.org/
}
\author{
Gabriel Abrahao
}
