% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMIA.R
\name{readGMIA}
\alias{readGMIA}
\title{{readGMIA }}
\usage{
readGMIA(subtype = NULL)
}
\arguments{
\item{subtype}{: Available subtypes are:
\itemize{
\item all_data_national : National Data on AEI (including differentiation by source (Groundwater, Surface water and nonconventional)), AAI
\item aei_pct : AEI in Percent
\item aei_ha : AEI in ha
\item aai_pct_aei : AAI as percentage of AEI
\item aeigw_pct_aei AEI from Groundwater sources as Percentage of total AEI
\item aeisw_pct_aei : AEI from Surface water sources as Percentage of total AEI
\item aeinc_pct_aei : AEI from nonconventional sources as Percentage total AEI
}}
}
\value{
magpie object of the Irrigated Area  data
}
\description{
Read Global Irrigation Map 

Read the Data from Siebert et.al on Irrigated Areas for each Country. Data contains total valuesas well as values from groundwater, surface water and non convential water sources for the following Categories:
\itemize{
\item Area equipped for Irrigation (AEI)
\item Area acutally irrigated (AAI)
\item Consumption Irrigatoin Water use (ICU)
}
The following Data is also available spatialy explicit with a resolution of 5 arcmin:
\itemize{
\item AEI in Percent
\item AEI in Ha
\item AAI as percent of AEI
\item AEI from groundwater sources as percent of total AEI
\item AEI from surface water sources as percent of total AEI
\item AEI from non conventional sources as percent of total AEI
}
There also exists a correct function to aggregate the data to 0.5 degree resolution, set convert="correctonly" to run.
}
\examples{

\dontrun{ 
a <- readSource("GlobalIrrigationMap", "all_data_national")
a <- readSource ("GMIA", "aei_pct", convert=F)
a <- readSource ("GMIA", "aei_pct", convert="correctonly")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
