% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPFADB_2005.R
\name{readGPFADB_2005}
\alias{readGPFADB_2005}
\title{Read GPFADB_2005}
\usage{
readGPFADB_2005(subtype)
}
\arguments{
\item{subtype}{data subtype. Available subtypes are Plantations, SNForest, PlantedSNForest, TotPlanted, GroParInit, AgeClassInit, Ownership and EndUse}
}
\value{
magpie object
}
\description{
Read-in an Global plantation forest database (GPFD) data from 2005 (FAO)
}
\examples{

\dontrun{ a <- readSource("GPFADB_2005","Plantations")
}

}
\seealso{
\code{\link{readSource}}

\url{http://foris.fao.org/static/pdf/ipc/GPFADB_2005.zip}

\url{http://www.fao.org/forestry/plantedforests/67507/en/}
}
\author{
Abhijeet Mishra
}
