% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGSOC.R
\name{readGSOC}
\alias{readGSOC}
\title{readGSOC}
\usage{
readGSOC()
}
\value{
magpie object in cellular resolution
}
\description{
This function reads the raw GSOC data (available at http://54.229.242.119/GSOCmap/) 
or if available the preprocessed raster layers.
}
\examples{
\dontrun{ readSource("GSOC") }

}
\author{
Kristine Karstens
}
