% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGTAP.R
\name{readGTAP}
\alias{readGTAP}
\title{readGTAP}
\usage{
readGTAP(subtype = NULL)
}
\arguments{
\item{subtype}{Type of GTAP data that should be read. So far available are:
  \itemize{
  \item GTAP7:
  \itemize{
  \item \code{GTAP7_VIWS}: Trade - Bilateral Imports at World Prices
  \item \code{GTAP7_VIMS}: Trade - Bilateral Imports at Market Prices
  \item \code{GTAP7_VXWD}: Trade - Bilateral Exports at World Prices
  \item \code{GTAP7_VXMS}: Trade - Bilateral Exports at Market Prices
  \item \code{GTAP7_VDFM}: Intermediates - Firms' Domestic Purchases at Market Prices
  \item \code{GTAP7_VIFM}: Intermediates - Firms' Imports at Market Prices
  \item \code{GTAP7_VFM}: Endowments - Firms' Purchases at Market Prices
  \item \code{GTAP7_VOA}: Payment received by producers (fram gtate value)
  \item \code{GTAP7_VOM}: Value of output at dometic market prices
  }
  \item GTAP8:
  \itemize{
  \item \code{GTAP8_VIWS}: Trade - Bilateral Imports at World Prices
  \item \code{GTAP8_VIMS}: Trade - Bilateral Imports at Market Prices
  \item \code{GTAP8_VXWD}: Trade - Bilateral Exports at World Prices
  \item \code{GTAP8_VXMS}: Trade - Bilateral Exports at Market Prices
  \item \code{GTAP8_VDFM}: Intermediates - Firms' Domestic Purchases at Market Prices
  \item \code{GTAP8_VIFM}: Intermediates - Firms' Imports at Market Prices
  \item \code{GTAP8_VFM}: Endowments - Firms' Purchases at Market Prices
  \item \code{GTAP8_VOA}: Payment received by producers (fram gtate value)
  \item \code{GTAP8_VOM}: Value of output at dometic market prices
  }
}}
}
\value{
GTAP data as a MAgPie-Object
}
\description{
Read BaseData and BaseView in GTAP database that has been downlodaded from the GTAP wewbsite.
}
\examples{
  \dontrun{
    a <- readSource("GTAP7", "VIWS")
           }
}
\author{
Stephen Wirth, Xiaoxi Wang
}
