% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGTAPTaxRates.R
\name{readGTAPTaxRates}
\alias{readGTAPTaxRates}
\title{readGTAPTaxRates}
\usage{
readGTAPTaxRates(subtype = NULL)
}
\arguments{
\item{subtype}{Type of GTAP data that should be read. So far available are:
  \itemize{
  \item GTAP7:
  \itemize{
  \item \code{GTAP7_RTMS}: Import tariffs ad valorem [%]
  \item \code{GTAP7_RTXS}: Export tariffs ad valorem [%]
  }
  \item GTAP8:
  \itemize{
  \item \code{GTAP8_RTMS}: Import tariffs ad valorem [%]
  \item \code{GTAP8_RTXS}: Export tariffs ad valorem [%]
  }
}}
}
\value{
GTAP tax rates data as a MAgPie-Object
}
\description{
Read in Tax rates from BaseRate dataset of GTAP databases that has been downlodaded from the GTAP wewbsite.
}
\examples{
  \dontrun{
    a <- readSource("GTAP7", "RTMS")
           }
}
\author{
Xiaoxi Wang
}
