% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHDD_CDD.R
\name{readHDD_CDD}
\alias{readHDD_CDD}
\title{Read HDD CDD}
\usage{
readHDD_CDD(subtype = 18)
}
\arguments{
\item{subtype}{Temperature threshold for computing HDD and CDD: 18, 21 or 25}
}
\value{
magpie object HDD CDD
}
\description{
Read-in HDD and CDD past and scenario data as magclass object
}
\examples{

\dontrun{ a <- readSource(type="HDD_CDD")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque
}
