% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHerridge.R
\name{readHerridge}
\alias{readHerridge}
\title{readHerridge}
\usage{
readHerridge(subtype = NULL)
}
\arguments{
\item{subtype}{a subtype for the calculation}
}
\value{
A MAgPIE object containting the share of Nr derived from
\itemize{
\item ndfa: fixation for each country and each commodity.
\item freeliving: fixation by free living agents
}
}
\description{
Reads a dataset containing values for biological nitrogen fixation in agricultural systems. Source: Herridge D. F., Peoples M. B., Boddey R. M.: Global inputs of biological nitrogen fixation in agricultural systems
}
\details{
Availables Subtypes:
 \itemize{
 \item ndfa:  National values for Plant associated fixation
 \item freeliving: Global values for free living agents
 }
}
\examples{

  \dontrun{
    x <- readSource("Herridge", "ndfa")
    x <-  readSource("Herridge", "freeliving", convert=F)
  }

}
\author{
Stephen Wirth
}
