% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHHS_USDA.R
\name{readHHS_USDA}
\alias{readHHS_USDA}
\title{readHHS_USDA}
\usage{
readHHS_USDA()
}
\value{
Magpie object with results on global level.
}
\description{
reads calory requirement for a standardized population from 
HHS & USDA. 2015. "2015-2020 Dietary Guidelines for Americans." 
8. Dietary Guidelines for Americans. 
https://health.gov/dietaryguidelines/2015/resources/2015-2020_Dietary_Guidelines.pdf.
Appendix 2
}
\examples{

\dontrun{ 
readSource("Lassaletta2014",subtype="budget",convert=FALSE)
}
}
\seealso{
\code{\link{convertLassaletta2014}},
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky
}
