% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIAEA.R
\name{readIAEA}
\alias{readIAEA}
\title{Nuclear data from world-nuclear.org}
\usage{
readIAEA()
}
\description{
Data on currently operating and under-construction nuclear power plants, reactors planned and proposed, 
electricity generation from nuclear
}
\author{
Christoph Bertram
}
