% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIAEA_PRIS.R
\name{readIAEA_PRIS}
\alias{readIAEA_PRIS}
\title{read IAEA Power Reactor Information System}
\usage{
readIAEA_PRIS()
}
\description{
read Nuclear capacities and near-term outlook from data scraped from
https://pris.iaea.org/PRIS/CountryStatistics/CountryStatisticsLandingPage.aspx
}
\author{
Pascal Weigmann
}
