% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_CCUS.R
\name{readIEA_CCUS}
\alias{readIEA_CCUS}
\title{Read IEA CCUS data}
\usage{
readIEA_CCUS(subtype)
}
\arguments{
\item{subtype}{either `historical` for data until 2023,
`projections` for "high" and "low" projections up to 2030 used as input-data
 or `pipeline` separated by status for use in formulating near-term bounds}
}
\description{
Reads in capacities from projects in IEA CCUS database
}
\author{
Anne Merfort, Falk Benke
}
