% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_EV.R
\name{readIEA_EV}
\alias{readIEA_EV}
\title{Read IEA_EV}
\usage{
readIEA_EV(subtype)
}
\arguments{
\item{subtype}{Which data: combSales, combStock, bevSales, bevStock, phevSales, or phevStock}
}
\value{
IEA data on electric vehicles as magpie object
}
\description{
Read-in IEA_EV
}
\examples{

\dontrun{ a <- readSource(type="IEA_EV")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark, Christoph Bertram
}
