% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_HSMR.R
\name{readIEA_HSMR}
\alias{readIEA_HSMR}
\title{read IEA Hydro Special Market Report}
\usage{
readIEA_HSMR()
}
\description{
read Hydro capacities and near-term outlook from data scraped from
https://www.iea.org/data-and-statistics/data-tools/hydropower-data-explorer
}
\author{
Pascal Weigmann
}
