% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_HydrogenProduction.R
\name{readIEA_HydrogenProduction}
\alias{readIEA_HydrogenProduction}
\title{Reads excel sheet of IEA hydrogen project pipeline and builds yearly capacity by country from it.
The data is distinguished by technology}
\usage{
readIEA_HydrogenProduction()
}
\value{
magpie object with Hydrogen capacity data both in MWel and in kT H2/y
}
\description{
Reads excel sheet of IEA hydrogen project pipeline and builds yearly capacity by country from it.
The data is distinguished by technology
}
\author{
Simon Krogmann
}
