% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_PRIS.R
\name{readIEA_PRIS}
\alias{readIEA_PRIS}
\title{read IEA Power Reactor Information System}
\usage{
readIEA_PRIS()
}
\description{
read Nuclear capacities and near-term outlook from data scraped from
https://pris.iaea.org/PRIS/CountryStatistics/CountryStatisticsLandingPage.aspx
}
\author{
Pascal Weigmann
}
