% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_PVPS.R
\name{readIEA_PVPS}
\alias{readIEA_PVPS}
\title{Reads PV investment cost data for 2020 which are based on 2018 data from IEA PVPS report}
\usage{
readIEA_PVPS()
}
\value{
magpie object with PV investment cost data
}
\description{
reads excel sheet with PV investment cost data
}
\author{
Felix Schreyer
}
