% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_REN.R
\name{readIEA_REN}
\alias{readIEA_REN}
\title{Reads the distributed solar pv capacity from IEA Renewables report (2019).}
\usage{
readIEA_REN()
}
\description{
Reads the distributed solar pv capacity from IEA Renewables report (2019).
}
\details{
Capacity in GW. Distributed solar, defined in the IEA Renewables (2019), includes
rooftop residential (0-10 kW, grid-connected), tooftop and ground-mounted commercial
 and industrial (10-1000kW, grid-connected), and off-grid (8W - 100 kW)
}
\author{
Aman Malik
}
