% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEIO.R
\name{readIEA_WEIO}
\alias{readIEA_WEIO}
\title{IEA World Energy Investment Outlook}
\usage{
readIEA_WEIO()
}
\description{
Read 2015-2024 investments statistical data in energy sector (electricity, oil, gas)
IEA World Energy Investment Outlook (2024)
(https://www.iea.org/data-and-statistics/data-product/world-energy-investment-2024-datafile)
}
\author{
Nicolas Bauer, Falk Benke
}
