% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEIO_2014.R
\name{readIEA_WEIO_2014}
\alias{readIEA_WEIO_2014}
\title{IEA World Energy Investment Outlook (2014)}
\usage{
readIEA_WEIO_2014()
}
\value{
A \code{\link[=madrat_mule]{madrat_mule()}} with a list containing the \link{tibble} \code{data} with
2014–20 average annual investments into \verb{Energy intensive} and
\verb{Non-energy intensive} industry, in $bn 2012, and the \link{tibble}
\code{country_groups} with \verb{IEA region}s and corresponding \code{iso3c} country
codes.
}
\description{
Read projected 2014-20 investments into industry energy efficiency from the
\href{http://www.iea.org/publications/freepublications/publication/weo-2014-special-report---investment.html}{IEA World Energy Investment Outlook (2014)}
}
