% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEO.R
\name{readIEA_WEO}
\alias{readIEA_WEO}
\title{Read IEA World Energy Outlook data}
\usage{
readIEA_WEO(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "Capacity" or "Invest_Costs"}
}
\description{
Read-in IEA WEO 2016 data for investment costs of different technologies,
and WEO 2017 data for historical electricity capacities (GW)
}
\author{
Renato Rodrigues, Aman Malik, and Jerome Hilaire
}
