% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEO_2021.R
\name{readIEA_WEO_2021}
\alias{readIEA_WEO_2021}
\title{IEA WEO 2021 Data}
\usage{
readIEA_WEO_2021()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
IEA WEO 2021 Data. See README in input file for more details.
}
\author{
Falk Benke
}
