% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFA.R
\name{readIFA}
\alias{readIFA}
\title{Read IFA}
\usage{
readIFA(subtype)
}
\arguments{
\item{subtype}{Type of IFA data that should be read. Available types are:
\itemize{ \item \code{consumption}: read in fertilizer_consumption.xlsx data
\item \code{production}: read in fertilizer_production.xlsx data }}
}
\value{
magpie object of the IFA data
}
\description{
Read-in IFA (International Fertilizer Association) data .xlsx file as
magclass object
}
\examples{

\dontrun{ a <- readSource(type="IFA",subtype="consumption")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
