% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIIASA_subs_taxes.R
\name{readIIASA_subs_taxes}
\alias{readIIASA_subs_taxes}
\title{Read IIASA subsidies and taxes}
\usage{
readIIASA_subs_taxes(subtype)
}
\arguments{
\item{subtype}{Type of country level data as compiled by IIASA that should
be read in. Available types are: \itemize{ \item \code{tax_rate}: tax rate
pre final energy category \item \code{subsidies_bulk}: subsidy rate per final
energy category \item \code{energy}: final energy quantities per category }}
}
\value{
magpie object of the IIASA_subs_taxes data
}
\description{
Read-in country level data on final energy taxes and subsidies as provided
from IIASA from .csv file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "IIASA_subs_taxes", "tax_rate")
}
}
\author{
Christoph Bertram
}
