% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIIASALanduse.R
\name{readIIASALanduse}
\alias{readIIASALanduse}
\title{read IIASA Land Use Emissions}
\usage{
readIIASALanduse(subtype)
}
\arguments{
\item{subtype}{one of "historical", "forecast2030", "forecast2035"}
}
\description{
read IIASA Land Use Emissions
}
\author{
Falk Benke
}
