% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMAGE2025.R
\name{readIMAGE2025}
\alias{readIMAGE2025}
\title{Read F-gases emissions data from several IMAGE scenarios, obtained in 2025}
\usage{
readIMAGE2025()
}
\value{
magpie object with several F-gas emission variables for several IMAGE scenarios
}
\description{
Read F-gases emissions data from several IMAGE scenarios, obtained in 2025
}
\author{
Gabriel Abrahao
}
