% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{readIMF}
\alias{readIMF}
\title{Read IMF}
\usage{
readIMF()
}
\value{
magpie object of the data
}
\description{
Read-in data that are based on IMF
}
\examples{

\dontrun{ a <- readSource(type="IMF")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
