% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMPACT3.2.2World_Price.R
\name{readIMPACT3.2.2World_Price}
\alias{readIMPACT3.2.2World_Price}
\title{Read IMPACT3.2.2World_Price}
\usage{
readIMPACT3.2.2World_Price()
}
\value{
magpie object of the world prices from the IMPACT model for different SSP scenarios
}
\description{
Read-in world prices data csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="IMPACT3.2.2World_Price")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Mishko Stevanovic
}
