% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJames.R
\name{readJames}
\alias{readJames}
\title{Read James}
\usage{
readJames()
}
\value{
GDP per capita in USD05 in PPP or MER as magpie object
}
\description{
Read-in GDP per-capita data from the publication James, Spencer L., Paul
Gubbins, Christopher JL Murray, and Emmanuela Gakidou. 2012. "Developing a
Comprehensive Time Series of GDP per Capita for 210 Countries from 1950 to
2015." Population Health Metrics 10 (1): 12. doi:10.1186/1478-7954-10-12.
from a .csv file to a magclass object
}
\examples{

\dontrun{ a <- readSource(type="James",subtype="IHME_USD05_PPP_pc")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
