% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJRC_IDEES.R
\name{readJRC_IDEES}
\alias{readJRC_IDEES}
\title{Read JRC IDEES}
\usage{
readJRC_IDEES(subtype)
}
\arguments{
\item{subtype}{one of
\itemize{
\item \code{'Emission'}: read worksheets from the Emission Balance files
\item \code{'Energy'}:   read worksheets from the Energy Balance files
\item \code{'Industry'}: read worksheets from the Industry files
\item \code{'Transport'}: read worksheets from the Transport files
\item \code{'MBunkers'}: read worksheets from the Bunkers files
\item \code{'Residential'}: read worksheets from the Residential files
\item \code{'Tertiary'}: read worksheets from the Tertiary (Services and Agriculture) files
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read the \href{https://data.jrc.ec.europa.eu/dataset/jrc-10110-10001/resource/f590b6f1-60e5-49a6-a972-60bc2b2e34b3}{IDEES data base from JRC}
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Michaja Pehl, Falk Benke
}
