% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKLWdamage.R
\name{readKLWdamage}
\alias{readKLWdamage}
\title{Reads country-specific damage coefficients for the damage function presented in
Kotz et al. (2024). Data has been provided by the authors.
This contains data for all countries and for 1000 boostrapping realizations per country,
capturing uncertainty from climate and empirical modeling.
Subtypes are the temperature and temperature^2 coefficients and the maximum temperature
per country for which the function is defined.}
\usage{
readKLWdamage(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "beta1", "beta2" or "maxGMT"}
}
\value{
KLW damage coefficients
}
\description{
Reads country-specific damage coefficients for the damage function presented in
Kotz et al. (2024). Data has been provided by the authors.
This contains data for all countries and for 1000 boostrapping realizations per country,
capturing uncertainty from climate and empirical modeling.
Subtypes are the temperature and temperature^2 coefficients and the maximum temperature
per country for which the function is defined.
}
\author{
Franziska Piontek
}
