% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKoeppenGAMS.R
\name{readKoeppenGAMS}
\alias{readKoeppenGAMS}
\title{readKoeppenGAMS}
\usage{
readKoeppenGAMS()
}
\value{
Koeppen geiger climate zone classsification data
}
\description{
Reads in the koeppen geiger climate data compatible to GAMS usage.
}
\examples{

\dontrun{ 
readSource("readKoeppenGAMS")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
