% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKrausmann.R
\name{readKrausmann}
\alias{readKrausmann}
\title{Read Krausmann}
\usage{
readKrausmann()
}
\value{
magpie object of the Krausmann data
}
\description{
Read-in an Krausmann data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="Krausmann")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
