% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKuczmarski.R
\name{readKuczmarski}
\alias{readKuczmarski}
\title{readKuczmarski}
\usage{
readKuczmarski()
}
\value{
magpie object
}
\description{
Reads in data on body height recommendations and standard deviations
body height:
https://www.cdc.gov/nchs/data/series/sr_11/sr11_246.pdf
Table 16
Kuczmarski RJ, Ogden CL, Guo SS, et al. 2000 CDC growth charts for the
United States: Methods and development. National Center for Health Statistics.
Vital Health Stat 11(246). 2002
}
\seealso{
\code{\link{readNCDrisc}}
}
