% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandfillCH4Capture.R
\name{readLandfillCH4Capture}
\alias{readLandfillCH4Capture}
\title{readLandfillCH4Capture}
\usage{
readLandfillCH4Capture()
}
\value{
Magpie object with results on global level.
}
\description{
reads percentage of ch4 from landfills that is captured, from World Bank CURB Toolv2.1
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
