% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLassaletta2014.R
\name{readLassaletta2014}
\alias{readLassaletta2014}
\title{readLassaletta2014}
\usage{
readLassaletta2014(subtype = "budget")
}
\arguments{
\item{subtype}{budget provides the nr cropland budgets, fert_to_cropland the sahre of inorganic fertilizers being applied to croplands}
}
\value{
Magpie object with results on country level.
}
\description{
reads nitrogen budgets for a country dataset from 
Lassaletta, L., G. Billen, B. Grizzetti, J. Angalde, and J. Garnier. 2014. 
50 Year Trends in Nitrogen Use Efficiency of World Cropping Systems: The Relationship between Yield and Nitrogen Input to Cropland.
Environmental Research Letters.
}
\examples{

\dontrun{ 
readSource("Lassaletta2014",subtype="budget",convert=FALSE)
}
}
\seealso{
\code{\link{convertLassaletta2014}},
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky
}
