% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLee.R
\name{readLee}
\alias{readLee}
\title{readLee
Read in Aviation emission data from Lee}
\usage{
readLee(subtype)
}
\arguments{
\item{subtype}{Either 'emi' or 'ef'}
}
\value{
magpie object of Aviation emission / emission factors data
}
\description{
readLee
Read in Aviation emission data from Lee
}
\examples{

\dontrun{ a <- readSource(type="", subtype="Waste")
}


}
\seealso{
\code{\link{readSource}}
}
\author{
Julian Oeser
}
