% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLotzeCampenBiofuel.R
\name{readLotzeCampenBiofuel}
\alias{readLotzeCampenBiofuel}
\title{Read LotzeCampenBiofuel}
\usage{
readLotzeCampenBiofuel()
}
\value{
Future trends in first generation bioenergy demand in ExaJoule as magpie object
}
\description{
Read-in Future trends in first generation bioenergy demand from the publication 
Lotze Campen et al. 2014. "Impacts of increased bioenergy demand on global food
markets: an AgMIP economic model intercomparison" Agricultural Economics 45 (103-116).
doi:10.1111/agec.12092.
from a .csv file to a magclass object
}
\examples{

\dontrun{ a <- readSource("LotzeCampenBiofuel")
}


}
\seealso{
\code{\link{readSource}}
}
\author{
Ewerton Araujo
}
