% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJml_rev21.R
\name{readLPJml_rev21}
\alias{readLPJml_rev21}
\title{readLPJml_rev21}
\usage{
readLPJml_rev21(subtype = "soilc")
}
\arguments{
\item{subtype}{Switch between diffrent input}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LPJmL contant
}
\examples{

\dontrun{ 
readSource("LPJml_rev21", subtype="maize_mrh")
}

}
\seealso{
\code{\link{readLPJml_rev21}},
\code{\link{readLPJ}}
}
\author{
Benjamin Leon Bodirsky
}
