% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmlCarbon.R
\name{readLPJmlCarbon}
\alias{readLPJmlCarbon}
\title{readLPJmlCarbon}
\usage{
readLPJmlCarbon(subtype = "historical.nat_veg.soilc_layer")
}
\arguments{
\item{subtype}{Input name coming from calcLPJmlCarbon}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LPJmL contant
}
\examples{

\dontrun{ 
readSource("LPJmlCarbon", subtype="historical.nat_veg.soilc_layer", convert="onlycorrect")
}

}
\seealso{
\code{\link{readLPJmlCarbon}},
\code{\link{readLPJ}}
}
\author{
Kristine Karstens
}
