% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLutz2014.R
\name{readLutz2014}
\alias{readLutz2014}
\title{readLutz2014}
\usage{
readLutz2014()
}
\value{
magpie object with the dataset downloaded. It contains missing values
and it is possible to replace them with the function convertLutz2014.
}
\description{
It reads and clears the dataset of the global population projections 
by age, sex and education, available on the Wittgenstein Centre Data Explorer
and published by Lutz, Butz and K. C. , 2014. "Population and human capital 
in the twenty-first century" Oxford University Press. 
From .csv file to a magclass object
}
\seealso{
\code{\link{convertLutz2014}}
}
