% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMacknickIntensities.R
\name{readMacknickIntensities}
\alias{readMacknickIntensities}
\title{Read Macknic Intensities}
\usage{
readMacknickIntensities(subtype)
}
\arguments{
\item{subtype}{Type of Macknick data that should be read. Available types
are: \itemize{ \item \code{data}: The original Macknick source data \item
\code{missingAssumed}: Additional data to fill gaps }}
}
\value{
MAgPIE object of the Macknick (2011) data
}
\description{
Read in Macknick (2011) data on water consumption and withdrawal
coefficients per electricity technology
}
\examples{
\dontrun{
a <- readSource(type = "MacknickIntensities", convert = FALSE)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Ioanna Mouratiadou
}
