% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAgPIE.R
\name{readMAgPIE}
\alias{readMAgPIE}
\title{Read MAgPIE data}
\usage{
readMAgPIE(subtype)
}
\arguments{
\item{subtype}{Either "EmiAirPoll", "macBase" or "co2tax"}
}
\value{
magpie object
}
\description{
Read-in MAgPIE data
}
\examples{

\dontrun{ a <- readSource(type="MAgPIE",subtype="EmiAPExo")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Julian Oeser
}
