% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMarineRegionsOrg.R
\name{readMarineRegionsOrg}
\alias{readMarineRegionsOrg}
\title{Load Exclusive Economic Zone (EEZ) size data as magclass object.
Areas with overlapping claims or jointly governed areas are split equally
between the respective countries.}
\usage{
readMarineRegionsOrg()
}
\description{
Source:
Flanders Marine Institute (2023). Maritime Boundaries Geodatabase:
Maritime Boundaries and Exclusive Economic' Zones (200NM), version 12.
Available online at https://www.marineregions.org/. https://doi.org/10.14284/632
}
\author{
Tabea Dorndorf
}
