% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMIRCA.R
\name{readMIRCA}
\alias{readMIRCA}
\title{readMIRCA}
\usage{
readMIRCA(subtype = "multicropping")
}
\arguments{
\item{subtype}{'multicropping' (default): multicropping factor on cropped (excluding fallow) land,}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read in data based on MIRCA data (https://www.uni-frankfurt.de/45218023/MIRCA)
}
\examples{

\dontrun{ 
readSource("MIRCA", subtype="multicropping")
}

}
\author{
Kristine Karstens
}
