% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMueller.R
\name{readMueller}
\alias{readMueller}
\title{Read M端ller et al. 2013 data.}
\usage{
readMueller(subtype)
}
\arguments{
\item{subtype}{One of:
\itemize{
\item \code{countries}: read table mapping country names use by M端ller et al. 2013
to ISO 3166-1 alpha-3 codes.
\item \code{stocks}: read low/medium/high estimates of per-capita steel stocks from
M端ller et al. 2013 SI2
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read data from M端ller et al. 2013 (http://dx.doi.org/10.1021/es402618m).
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Michaja Pehl
}
