% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNCDrisc.R
\name{readNCDrisc}
\alias{readNCDrisc}
\title{readNCDrisc}
\usage{
readNCDrisc(subtype)
}
\arguments{
\item{subtype}{"height" for body height data}
}
\value{
magpie object with the dataset downloaded. It contains missing values
and it is possible to replace them with the function convertNCDrisc
}
\description{
Reads in data from the NCD risc consortium
body height:
Collaboration (NCD-RisC), NCD Risk Factor. 2016. "A Century of Trends in Adult Human Height." ELife 5 (July):e13410. https://doi.org/10.7554/eLife.13410.
}
\seealso{
\code{\link{convertNCDrisc}}
}
\author{
Benjamin Leon Bodirsky
}
