% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNewClimate.R
\name{readNewClimate}
\alias{readNewClimate}
\title{Reads NPI policy database with technology capacity target from the Policy data base (v4 August 2024)
by PBL that translate the high impact policies of https://climatepolicydatabase.org/.}
\usage{
readNewClimate(subtype, subset)
}
\arguments{
\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets,
RenShareTargets for renewable energy share targets,
with YYYY NDC version year, determines the database version to be read in}

\item{subset}{A string (or vector of strings) designating the scenario(s) to be returned (only used in convert).}
}
\description{
Reads excel sheet with NPi (National Policies Implemented)
data on different policy targets (capacity, production, emissions) with different variations.
NPI targets only include targets that are based on implemented policy instruments.
}
\author{
Rahel Mandaroux, Léa Hayez, Falk Benke
}
