% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnonEnergyDemand.R
\name{readnonEnergyDemand}
\alias{readnonEnergyDemand}
\title{Read Final energy demand for feedstocks (non-energy use)}
\usage{
readnonEnergyDemand()
}
\value{
magpie object of region dependent data
}
\description{
Read Final energy demand for feedstocks (non-energy use)
}
\examples{
\dontrun{
a <- readSource(type = "nonEnergyDemand")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
