% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNREL.R
\name{readNREL}
\alias{readNREL}
\title{Read NREL}
\usage{
readNREL(subtype)
}
\arguments{
\item{subtype}{type either "onshore" or "offshore"}
}
\value{
magpie object of NREL
}
\description{
Read-in NREL xlsx file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "NREL", subtype = "onshore")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
