% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNTNU_LCA.R
\name{readNTNU_LCA}
\alias{readNTNU_LCA}
\title{Read NTNU_LCA}
\usage{
readNTNU_LCA()
}
\value{
magpie object of the NTNU_LCA data
}
\description{
Read-in an NTNU_LCA data .xlsx file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="NTNU_LCA")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
