% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD.R
\name{readOECD}
\alias{readOECD}
\alias{convertOECD}
\title{Read OECD}
\usage{
readOECD()

convertOECD(x)
}
\arguments{
\item{x}{MAgPIE object returned from readOECD}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Read-in risk premium
}
\examples{
\dontrun{
readSource("OECD")
}
}
\seealso{
[madrat::readSource()]
}
