% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOpenmod.R
\name{readOpenmod}
\alias{readOpenmod}
\title{Read Openmod capacities data}
\usage{
readOpenmod()
}
\value{
magpie object of the LIMES team updated Openmod data on capacities (GW)
}
\description{
Read-in an modified openmod capacities data file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "Openmod")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
