% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPauliuk.R
\name{readPauliuk}
\alias{readPauliuk}
\title{Read Pauliuk et al. 2013 data}
\usage{
readPauliuk(subtype = "lifetime")
}
\arguments{
\item{subtype}{One of:
\itemize{
\item \code{lifetime}: Read estimated lifetime of overall steel stocks (approach b)
in years.
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read data from Pauliuk et al. 2013
(https://dx.doi.org/10.1016/j.resconrec.2012.11.008).
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Michaja Pehl
}
