% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPBL_EFsBaseline.R
\name{readPBL_EFsBaseline}
\alias{readPBL_EFsBaseline}
\title{Read emission factors from a PBL IMAGE baseline (no mitigation) scenario
for a specific sector and gas species (CH4 or N2O)}
\usage{
readPBL_EFsBaseline(subtype)
}
\arguments{
\item{subtype}{gas and subsector combination string. One of:
c("CH4_entf", "CH4_gasp", "CH4_landf", "CH4_manu", "CH4_oilp",
"CH4_rice", "CH4_sewa", "N2O_adip", "N2O_fert", "N2O_manu",
"N2O_nitr", "N2O_sewa", "N2O tran")}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Read emission factors from a PBL IMAGE baseline (no mitigation) scenario
for a specific sector and gas species (CH4 or N2O)
}
