% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPRIMAPhist.R
\name{readPRIMAPhist}
\alias{readPRIMAPhist}
\title{readPRIMAPhist}
\usage{
readPRIMAPhist(subtype)
}
\arguments{
\item{subtype}{data subtype. available subtypes are:
\itemize{
\item hist
\item hist_no_ex
}}
}
\value{
magpie object of the PRIMAP-hist data.
}
\description{
Read in an PRIMAP-hist data csv file as magclass object.
}
\examples{

\dontrun{ 
a <- readPRIMAPhist("PRIMAPhist","hist")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Roman Popov
}
