% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProdShares.R
\name{readProdShares}
\alias{readProdShares}
\title{Reads shares of world manufacture for spv modules and wind turbines.}
\usage{
readProdShares()
}
\description{
Reads shares of world manufacture for spv modules and wind turbines.
}
\author{
Aman Malik
}
