% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPWT.R
\name{readPWT}
\alias{readPWT}
\alias{convertPWT}
\title{Read PWT}
\usage{
readPWT()

convertPWT(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPWT}
}
\description{
Read-in PWT data (version 8.0) as magclass object
}
\examples{
\dontrun{
readSource("PWT")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
