% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPYieldCoeff.R
\name{readPYieldCoeff}
\alias{readPYieldCoeff}
\title{Read in coefficients for calculating pasture intensification}
\usage{
readPYieldCoeff()
}
\value{
MAgPIE object
}
\description{
Read in csv file containing coefficients of linear regression
for the calculation of future pasture intensification dependent 
on animal numbers
}
\examples{
\dontrun{
a <- readSource("PYieldCoeff")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
