% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRCP.R
\name{readRCP}
\alias{readRCP}
\title{Read RCP 
Read in RCP data}
\usage{
readRCP(subtype)
}
\arguments{
\item{subtype}{Either 'Waste' or 'AviationShipping'}
}
\value{
magpie object of the RCP data
}
\description{
Read RCP 
Read in RCP data
}
\examples{

\dontrun{ a <- readSource(type="RCP", subtype="Waste")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Julian Oeser
}
