% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND.R
\name{readREMIND}
\alias{readREMIND}
\title{readREMIND}
\usage{
readREMIND(subtype)
}
\arguments{
\item{subtype}{Either "intensive" or "extensive"}
}
\value{
MAgPIE object with regional aggregation of REMIND H12
}
\description{
Reads in a reporting mif file from REMIND
}
\examples{

\dontrun{ 
readSource("REMIND",aggregate=FALSE)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
David Klein
}
