% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND_11Regi.R
\name{readREMIND_11Regi}
\alias{readREMIND_11Regi}
\title{Read REMIND 11 Regi data}
\usage{
readREMIND_11Regi(subtype)
}
\arguments{
\item{subtype}{Name of the source, e.g. "fossilExtractionCoeff", "uraniumExtractionCoeff"}
}
\description{
Read REMIND 11 Regi data
}
\author{
unknown
}
