% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRemindCesPrices.R
\name{readRemindCesPrices}
\alias{readRemindCesPrices}
\title{Read RemindCesPrices}
\usage{
readRemindCesPrices(subtype = "ccd632d33a")
}
\arguments{
\item{subtype}{Regional resolution of REMIND data which should be loaded. ccd632d33a corresponds to the REMIND-11,
and 690d3718e1 to REMIND-H12}
}
\value{
magpie object of REMIND prices
}
\description{
Read-in CES derivatives/prices from former REMIND runs
}
\examples{
\dontrun{ a <- readSource(type="RemindCesPrices", subtype = "ccd632d33a")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque
}
