% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREN21.R
\name{readREN21}
\alias{readREN21}
\title{Reads policy database from REN21 2017 with capacity targets or regional technology costs}
\usage{
readREN21(subtype)
}
\arguments{
\item{subtype}{Capacity Generation Emissions Share}
}
\value{
magpie object with Total Installed Capacity targets in GW for different target years
}
\description{
Reads excel sheet with data on proposed policies, 
on Renewable energy capacity targets (which are broken down into Total Installed Capacity 
(TIC-Absolute), Additional Installed Capacity (AC-Absolute), and Production Absolute targets) 
or regional technology costs
}
\details{
Country name is ISO coded. Capacity/Additional Capacity targets are in GW. Generation/Production targets are in GWh.
}
\author{
Aman Malik, Lavinia Baumstark
}
