% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readResFor2ndBE.R
\name{readResFor2ndBE}
\alias{readResFor2ndBE}
\title{readResFor2ndBE}
\usage{
readResFor2ndBE(subtype)
}
\arguments{
\item{subtype}{oldReMIND, newAgriSupply}
}
\value{
List of magpie objects with results on old ReMIND regions level
}
\description{
Read in old ReMIND use of residues for 2nd generation bioenergy and newly estimations
}
\examples{

\dontrun{ a <- readSource("ResFor2ndBE", subtype="oldReMIND")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Kristine Karstens
}
