% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRogelj2017.R
\name{readRogelj2017}
\alias{readRogelj2017}
\title{Reads NDC policy database from Rogelj  et al. 2017 with capacity, emission, and share targets}
\usage{
readRogelj2017(subtype)
}
\arguments{
\item{subtype}{Capacity Generation Emissions Share}
}
\value{
magpie object
}
\description{
Reads excel sheet (supplementary sheet from Rogelj et al., 2017 ) with NDC (Nationally Determined Contributions) 
data on different policy targets (capacity, emission, and share targets) with different variations
}
\details{
Country name is ISO coded. Capacity/Additional Capacity targets are in GW. Generation/Production targets are in GWh.
}
\author{
Aman Malik and Christoph Bertram
}
