% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSathayeForest.R
\name{readSathayeForest}
\alias{readSathayeForest}
\title{Read Sathaye Forest}
\usage{
readSathayeForest()
}
\value{
magpie object of the Sathaye Forest data
}
\description{
Read-in an Sathaye Forest data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource("SathayeForest")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
