% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSchofield.R
\name{readSchofield}
\alias{readSchofield}
\title{Read parameters of Schofield equations}
\usage{
readSchofield()
}
\value{
MAgPIE object
}
\description{
University, United Nations. 2004. Human Energy Requirements: Report of a Joint FAO/WHO/UNU Expert Consultation: Rome, 17-24 October 2001. Vol. 1. Food & Agriculture Org. http://books.google.com/books?hl=en&lr=&id=8WW7LP5h6usC&oi=fnd&pg=PR3&dq=%22working+groups.+The+work+of+these+groups+preceded+the+expert+consultation+and+served+as%22+%22for+discussions+and+exchange+during+the+meeting.+Thanks+are+also+due+to+Dr+E.%22+&ots=8Fizk--Hr6&sig=ocPdelHKyX2_npUTh41zZRRHf68.
}
\examples{

\dontrun{ a <- readSource("Schofield") }

}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
