% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilGrids.R
\name{readSoilGrids}
\alias{readSoilGrids}
\title{readSoilGrids}
\usage{
readSoilGrids(subtype)
}
\arguments{
\item{subtype}{Switch between different input. Use predefined ones or any FileName specified in 'SoilGrids/META_GEOTIFF_1B.csv'}
}
\value{
magpie object in cellular resolution
}
\description{
This function reads the raw SoilGrids data (available at https://files.isric.org/soilgrids/data/recent/OCSTHA_M_30cm_250m_ll.tif)
or if available the preprocessed raster layers.
}
\examples{
\dontrun{ readSource("SoilGrids", subtype = "cstock_0_30") }

}
\seealso{
\code{\link{downloadSoilGrids}}
}
\author{
Kristine Karstens
}
