% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSSP.R
\name{readSSP}
\alias{readSSP}
\title{Read SSP}
\usage{
readSSP(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "all" or "ratioPM"}
}
\value{
magpie object of the SSP data
}
\description{
Read-in an SSP data csv.zip file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="SSP",subtype="all")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
