% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSSPResults.R
\name{readSSPResults}
\alias{readSSPResults}
\title{readSSPResults}
\usage{
readSSPResults()
}
\value{
MAgPIE object with regional aggregation of SSP regions, including all indicators
}
\description{
Reads in a reporting mif file from the SSP scenario results
}
\examples{

\dontrun{ 
readSource("SSPresults",aggregate=FALSE)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
