% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStationary.R
\name{readStationary}
\alias{readStationary}
\alias{convertStationary}
\title{Load Stationary File as magclass object}
\usage{
readStationary(subset)

convertStationary(x, subset)
}
\arguments{
\item{subset}{A string, or vector of string designating the entries for the scenario dimension. If possible,
passed to \code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDP()}}. Unspecified scenarios receive SSP2 data.}

\item{x}{MAgPIE object returned from readStationary}
}
\value{
magclass object
}
\description{
Reads in EDGE_TradMod.cs4r, which hast data for the SSP scenarios. In the convert function, the subset argument
is used to specify the desired scenarios. Unspecified scenarios are copies of the SSP2 scenario.
}
\author{
Antoine Levesque, Robin Hasse
}
