% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStockChangeFactors.R
\name{readStockChangeFactors}
\alias{readStockChangeFactors}
\title{readStockChangeFactors}
\usage{
readStockChangeFactors()
}
\value{
List of magpie objects with results on cellular level
}
\description{
Read in IPCC stock cange factors for lpj croptypes
}
\examples{

\dontrun{ 
readSource("StockChangeFactors")
}

}
\seealso{
\code{\link{readStockChangeFactors}},
}
\author{
Kristine Karstens
}
