% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStrefler.R
\name{readStrefler}
\alias{readStrefler}
\title{Get data on enhanced weathering}
\usage{
readStrefler(subtype)
}
\arguments{
\item{subtype}{type of data, one of "weathering_graderegi", "weathering_costs"}
}
\value{
magpie object of region dependent data
}
\description{
Get data on enhanced weathering
}
\examples{

\dontrun{
a <- readSource(type="Strefler", subtype="weathering_graderegi")
}
}
\seealso{
\code{\link{readSource}}
}
