% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCdamageKrichene.R
\name{readTCdamageKrichene}
\alias{readTCdamageKrichene}
\title{Reads country-specific damage coefficients for tropical cyclones from
Krichene et al. (in prep.). Data has been provided by the authors,
but will be made publicly available as well.
This contains data for 41 countries (those exposed to tropical cyclones),
and two coefficients (constant and linear temperature)}
\usage{
readTCdamageKrichene(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "const" or "tasK"}
}
\value{
TC damage coefficients
}
\description{
Reads country-specific damage coefficients for tropical cyclones from
Krichene et al. (in prep.). Data has been provided by the authors,
but will be made publicly available as well.
This contains data for 41 countries (those exposed to tropical cyclones),
and two coefficients (constant and linear temperature)
}
\author{
Franziska Piontek
}
