% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUN_PopDiv.R
\name{readUN_PopDiv}
\alias{readUN_PopDiv}
\title{Read UN Population Division Data}
\usage{
readUN_PopDiv()
}
\value{
\code{magclass} object; population in thousands.
}
\description{
Read past UN population data. Covers 1950 to 2015 yearly and per M.49 area.
See \emph{United Nations, Department of Economicand Social Affairs, 
Population Division} "World Population Prospects: The 2015 Revision" 
(\href{https://esa.un.org/unpd/wpp/}{website}).
}
\seealso{
\code{\link{readSource}}
}
\author{
Michaja Pehl
}
