% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC.R
\name{readUNFCCC}
\alias{readUNFCCC}
\title{Read UNFCCC data}
\usage{
readUNFCCC()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Read UNFCCC data
}
\seealso{
[`readSource()`]
}
\author{
Falk Benke
}
