% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC_NDC.R
\name{readUNFCCC_NDC}
\alias{readUNFCCC_NDC}
\title{Reads NDC policy database with capacity, emission, and share targets, originally based on Rogelj et al. 2017}
\usage{
readUNFCCC_NDC(subtype)
}
\arguments{
\item{subtype}{Capacity_2021_cond (or 2018 or uncond) for capacity target, Emissions_2021_cond (or 2018 or uncond) for Emissions targets}
}
\value{
magpie object
}
\description{
Reads excel sheet with NDC (Nationally Determined Contributions)
data on different policy targets (capacity, emission, and share targets) with different variations
}
\details{
Country name is ISO coded. Capacity/Additional Capacity targets are in GW. Generation/Production targets are in GWh.
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters
}
