% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNIDO.R
\name{readUNIDO}
\alias{readUNIDO}
\title{read UNIDO data}
\usage{
readUNIDO(subtype)
}
\arguments{
\item{subtype}{one of
- \code{INDSTAT2}: read INDSTAT2 data}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read data from United Nations Industrial Organisation.
}
\seealso{
\code{\link[=readSource]{readSource()}},
}
\author{
Michaja Pehl
}
