% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVanDerWerf2010.R
\name{readVanDerWerf2010}
\alias{readVanDerWerf2010}
\title{readVanDerWerf2010}
\usage{
readVanDerWerf2010()
}
\value{
A MAgPIE object containing emissionfactors per burned dry matter for
different causes.
}
\description{
Reads a dataset containing values for global fire emissions. Source: van der
Werf G. R., Randerson J. T., Giglio L., Collatz G. J., Mu M., Kasibhatla P.
S., Morton D. C., DeFries R. S., Jin Y., van Leeuwen T. T.: Global fire
emissions and the contribution of deforestation, savanna, forest,
agricultural, and peat fires (1997-2009)
}
\examples{


  \dontrun{
    x <- readSource("VanDerWerf2010")
  }

}
\author{
Stephen Wirth
}
