% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVanDrecht2009.R
\name{readVanDrecht2009}
\alias{readVanDrecht2009}
\title{readvanDrecht2009}
\usage{
readVanDrecht2009()
}
\value{
A MAgPIE object containing sewage quantities and losses
}
\description{
Reads a dataset containing values for sewage
}
\examples{


  \dontrun{
    x <- readSource("VanDerWerf2010")
  }

}
\author{
Benjamin Leon Bodirksy
}
