% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvanRuijven2016.R
\name{readvanRuijven2016}
\alias{readvanRuijven2016}
\title{Read van Ruijven et al. (2016) data.}
\usage{
readvanRuijven2016()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read data from van Ruijven et al. 2016,
(http://dx.doi.org/10.1016/j.resconrec.2016.04.016,
https://www.zotero.org/groups/52011/rd3/items/itemKey/6QMNBEHQ), obtained
through personal communication (e-mail to Michaja Pehl).
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Michaja Pehl
}
