% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWaste.R
\name{readWaste}
\alias{readWaste}
\title{Read WhataWaste2.0 World Bank data}
\usage{
readWaste(subtype)
}
\arguments{
\item{subtype}{data subtype. "Generation" "Composition" and "Treatment" in kg/capita}
}
\value{
magpie object of the WhataWaste data with Generation, Disposal, or Composition
}
\description{
Read-in a xlsx file as magclass object
kg/cap
}
\examples{

\dontrun{ a <- readSource(type="Waste",subtype="Generation")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
