% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBGEM.R
\name{readWBGEM}
\alias{readWBGEM}
\title{readWBGEM}
\usage{
readWBGEM()
}
\value{
magpie object of global price of commodities
}
\description{
read raw data of world commodity prices from the dataset of World Bank global economic monitor
}
\examples{

\dontrun{ readSource(type = "WBGEM")}

}
\seealso{
\code{\link{readSource}}
}
\author{
Xiaoxi Wang
}
