% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBirrigation.R
\name{readWBirrigation}
\alias{readWBirrigation}
\title{Read WBirrigation}
\usage{
readWBirrigation()
}
\value{
magpie object of the WBirrigation data
}
\description{
Read-in an WBirrigation data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="WBirrigation")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
