% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWDIextended.R
\name{readWDIextended}
\alias{readWDIextended}
\title{Read dataset with population numbers completed for small islands states}
\usage{
readWDIextended(subtype = "gdppast")
}
\arguments{
\item{subtype}{Type: gdppast,gdpfuture,poppast,popfuture}
}
\value{
magpie object of the completed WDI data
}
\description{
Based on data review of werton Arauju
}
\examples{

\dontrun{ a <- readSource(type="WDI",subtype="SP.POP.TOTL")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Ewerton Arauju
}
