% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWFaS.R
\name{readWFaS}
\alias{readWFaS}
\title{readWFaS}
\usage{
readWFaS(subtype = NULL)
}
\arguments{
\item{subtype}{data subtype. available subtypes are:
\itemize{
\item dom_ww_ssp1_rcp4.5
\item dom_ww_ssp2_rcp6.0
\item dom_ww_ssp3_rcp6.0
\item dom_use_ssp1_rcp4.5
\item dom_use_ssp2_rcp6.0
\item dom_use_ssp3_rcp6.0
\item ind_ww_ssp1_rcp4.5
\item ind_ww_ssp2_rcp6.0
\item ind_ww_ssp3_rcp6.0
\item ind_use_ssp1_rcp4.5
\item ind_use_ssp2_rcp6.0
\item ind_use_ssp3_rcp6.0
\item dom_ww_ssp1_rcp4.5_gridded
\item dom_ww_tech_ssp1_rcp4.5_gridded
\item dom_use_ssp1_rcp4.5_gridded
\item dom_use_tech_ssp1_rcp4.5_gridded
\item ind_ww_ssp1_rcp4.5_gridded
\item ind_ww_tech_ssp1_rcp4.5_gridded
\item ind_use_ssp1_rcp4.5_gridded
\item ind_use_tech_ssp1_rcp4.5_gridded
\item dom_ww_ssp2_rcp2.6_gridded_2100
\item dom_ww_tech_ssp2_rcp2.6_gridded_2100
\item dom_use_ssp2_rcp2.6_gridded_2100
\item dom_use_tech_ssp2_rcp2.6_gridded_2100
\item ind_ww_ssp2_rcp2.6_gridded_2100
\item ind_ww_tech_ssp2_rcp2.6_gridded_2100
\item ind_use_ssp2_rcp2.6_gridded_2100
\item ind_use_tech_ssp2_rcp2.6_gridded_2100
\item dom_ww_ssp2_rcp4.5_gridded_2100
\item dom_ww_tech_ssp2_rcp4.5_gridded_2100
\item dom_use_ssp2_rcp4.5_gridded_2100
\item dom_use_tech_ssp2_rcp4.5_gridded_2100
\item ind_ww_ssp2_rcp4.5_gridded_2100
\item ind_use_ssp2_rcp4.5_gridded_2100
\item ind_use_tech_ssp2_rcp4.5_gridded_2100
\item dom_ww_ssp2_rcp6.0_gridded_2100
\item dom_ww_tech_ssp2_rcp6.0_gridded_2100
\item dom_use_tech_ssp2_rcp6.0_gridded_2100
\item ind_ww_ssp2_rcp6.0_gridded_2100
\item ind_ww_tech_ssp2_rcp6.0_gridded_2100
\item ind_use_ssp2_rcp6.0_gridded_2100
\item ind_use_tech_ssp2_rcp6.0_gridded_2100
\item dom_ww_ssp2_rcp6.0_gridded
\item dom_ww_tech_ssp2_rcp6.0_gridded
\item dom_use_ssp2_rcp6.0_gridded
\item dom_use_tech_ssp2_rcp6.0_gridded
\item ind_ww_ssp2_rcp6.0_gridded
\item ind_ww_tech_ssp2_rcp6.0_gridded
\item ind_use_ssp2_rcp6.0_gridded
\item ind_use_tech_ssp2_rcp6.0_gridded
\item dom_ww_ssp3_rcp6.0_gridded
\item dom_ww_tech_ssp3_rcp6.0_gridded
\item dom_use_ssp3_rcp6.0_gridded
\item dom_use_tech_ssp3_rcp6.0_gridded
\item ind_ww_ssp3_rcp6.0_gridded
\item ind_ww_tech_ssp3_rcp6.0_gridded
\item ind_use_ssp3_rcp6.0_gridded
\item ind_use_tech_ssp3_rcp6.0_gridded
\item dom_use_historic_gridded
\item dom_ww_historic_gridded
\item ind_use_historic_gridded
\item ind_ww_historic_gridded
}}
}
\value{
magpie object of the WFas data
}
\description{
Read in IPCC emissions

Read in Water Consumption data from WFaS: 
\strong{Historical} and data from \strong{PCRGLOBWOB} are available
\itemize{ 
\item  Industrial Water Use
\item Industrial Waste Water
\item Domestic Water Use
\item Domestic Waste Water
 }
Data is available \strong{with and without technological change} and for the following SSP and RCP Scenarios:
\itemize{
 \item SSP1RCP4.5
 \item SSP2RCP2.6
 \item SSP2RCP4.5
 \item SSP2RCP6.0
 \item SSP3RCP6.0
}
Data is available spacially explicit and countrywise (not for all RCPs and not with technological change).
All available subtypes are listed in the param section.
SUbtypes are structured as follows:
"Sector"_"use or ww"_"tech, historic or nothing"_ssp Scenarios"_rcp Scenario"_"gridded or nothing"_"2100 or nothing"
 Examples:
 \itemize{
 \item dom_ww_ssp1_rcp4.5 : Returns data for \strong{dom}estic \strong{w}aste \strong{w}ater volume \strong{for all countries}, from 2000 to 2050 for \strong{SSP1} and \strong{RCP4.5} without technological change
 \item dom_ww_ssp1_rcp4.5_gridded : Returns \strong{dom}estic \strong{w}aste \strong{w}ater volume on a \strong{0.5?x0.5? grid} from 2000 to 2050 for \strong{SSP1} and \strong{RCP4.5} without technological change
 \item dom_ww_tech_ssp2_rcp2.6_gridded_2100 : Returns \strong{dom}estic \strong{w}aste \strong{w}ater volume on a \strong{0.5 Degree x0.5 Degree grid} from 2000 to \strong{2100}, for \strong{SSP2} and \strong{RCP2.6} with \strong{tech}nological change
 \item ind_use_historic_gridded : Returns \strong{ind}ustrial water \strong{use} on a \strong{0.5 Degreex0.5 Degree grid}, from \strong{1960 to 1999}
 }
}
\note{
Data is missing for domestic use SSP2RCP6.0 and industrial Wastewater with technological change for SSP2RCP4.5.
}
\examples{

\dontrun{ 
a <- readSource("WFaS","dom_ww_ssp1_rcp4.5")
a <- readSource("WFaS","dom_ww_ssp1_rcp4.5_gridded")
a <- readSource("WFaS","dom_use_tech_ssp2_rcp6.0_gridded_2100")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Stephen Wirth
}
