% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWGBU.R
\name{readWGBU}
\alias{readWGBU}
\title{Read WGBU}
\usage{
readWGBU()
}
\value{
magpie object of WGBU
}
\description{
Read-in an WGBU xlsx file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "WGBU")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
