% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWHO.R
\name{readWHO}
\alias{readWHO}
\title{Read WHO}
\usage{
readWHO(subtype)
}
\arguments{
\item{subtype}{Type of WHO data that should be read. Includes
physical_inactivity_adults and physical_inactivity_underaged}
}
\value{
magpie object of the WHO data
}
\description{
Read-in WHO (World health organization) data files as magpie object.
The files contain information on physical inactivity
}
\examples{

\dontrun{ a <- readSource(type="WHO",subtype="physical_activity_adults")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Bodirsky
}
