% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWHObmi.R
\name{readWHObmi}
\alias{readWHObmi}
\title{readWHObmi}
\usage{
readWHObmi()
}
\value{
magpie object
}
\description{
Reads in data on body mass index (BMI) recommendations from WHO
http://www.who.int/childgrowth/standards/bmi_for_age/en/
http://www.who.int/growthref/who2007_bmi_for_age/en/
}
\seealso{
\code{\link{readNCDrisc}}
}
