% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWirsenius_FEED.R
\name{readWirsenius_FEED}
\alias{readWirsenius_FEED}
\title{Read Wirsenius_FEED}
\usage{
readWirsenius_FEED()
}
\value{
magpie object of the Wirsenius_FEED data
}
\description{
Read-in an Wirsenius_FEED data .csv files as one magclass object
}
\examples{

\dontrun{ a <- readSource("Wirsenius_FEED")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
