% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWirseniusPHD.R
\name{readWirseniusPHD}
\alias{readWirseniusPHD}
\title{Read Wirsenius PHD}
\usage{
readWirseniusPHD()
}
\value{
magpie object of the WirseniusPHD data
}
\description{
Read-in Wirsenius PHD data .csv file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="WirseniusPHD")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
