% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWirseniusSubsystems.R
\name{readWirseniusSubsystems}
\alias{readWirseniusSubsystems}
\title{Read Wirsenius Livestock Subsystems}
\usage{
readWirseniusSubsystems(subtype)
}
\arguments{
\item{subtype}{Type of subsystem information that should be read in. 
Available types are: "cattlemeat2milk", "meat2egg" and "feed_beefsys_ratio"}
}
\value{
magpie object of selected livestock subsystem data
}
\description{
Read in data on livestock subsystems structure. Source: Wirsenius.
}
\examples{

\dontrun{ 
a <- readSource("WirseniusSubsystems","cattlemeat2milk")
a <- readSource("WirseniusSubsystems","meat2egg")
a <- readSource("WirseniusSubsystems","feed_beefsys")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl
}
