% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWISE.R
\name{readWISE}
\alias{readWISE}
\title{readWISE}
\usage{
readWISE()
}
\value{
magpie object in cellular resolution
}
\description{
This function reads the preprocessed data from WISE_30sec
}
\examples{
\dontrun{ readSource("WISE") }

}
\author{
Kristine Karstens
}
