% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_expand_tibble.R
\name{tool_expand_tibble}
\alias{tool_expand_tibble}
\title{Expand tibble across scenarios and regions with default values}
\usage{
tool_expand_tibble(d, scenarios, regions, structure.columns = NULL)
}
\arguments{
\item{d}{A data.frame with columns `scenario` and `region`.}

\item{scenarios}{A character vector of scenario names.}

\item{regions}{A character vector of region names.}

\item{structure.columns}{A character vector of column names to be carried
along.}
}
\value{
A `tibble`.
}
\description{
The data.frame `d` is expanded in such a manner that all rows with `NA` in
either the `scenario` or `region` columns are extended to repeat for all
scenarios and regions listed in `scenarios` and `regions`.  Rows with
specified scenarios and/or regions will overwrite extended ones.  Regions are
expanded before scenarios.
}
