% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddDimensions.R
\name{toolAddDimensions}
\alias{toolAddDimensions}
\title{Wrapper around magclass::add_dimension supporting more than one value for the new dimension.
For each value, the input magclass object is copied, extended by the new dimension and
appended to the output.}
\usage{
toolAddDimensions(x, dimVals, dimName, dimCode)
}
\arguments{
\item{x}{a magclass object}

\item{dimVals}{list of values for the new dimension to be added}

\item{dimName}{name of the new dimension}

\item{dimCode}{dimension number of the new dimension (e.g. 3.1)}
}
\value{
the extended magclass object
}
\description{
Wrapper around magclass::add_dimension supporting more than one value for the new dimension.
For each value, the input magclass object is copied, extended by the new dimension and
appended to the output.
}
