% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateCustomRegs.R
\name{toolAggregateCustomRegs}
\alias{toolAggregateCustomRegs}
\title{Aggregate regions without regional aggregations such as global sum}
\usage{
toolAggregateCustomRegs(
  x,
  agg,
  rel,
  to = NULL,
  removeAllAgg = TRUE,
  regs = "GLO"
)
}
\arguments{
\item{x}{a magclass object in country resolution}

\item{agg}{magclass object supplying explicit regional aggregates}

\item{rel}{aggregation mapping for \code{toolAggregate}}

\item{to}{aggregation target for \code{toolAggregate}}

\item{removeAllAgg}{decide whether to exclude all aggregated data or keep
those (variables/periods) that are not overwritten by data from agg object}

\item{regs}{one or multiple names of aggregated regions to be removed/overwritten}
}
\value{
magclass object
}
\description{
Aggregate regional data, but if regional aggregations exist, discard the
automatically aggregated values and replace them with source data.
}
\author{
Falk Benke, Pascal Weigmann
}
