% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateTimeSteps.R
\name{toolAggregateTimeSteps}
\alias{toolAggregateTimeSteps}
\title{aggregate values to 5-year averages to suppress volatility}
\usage{
toolAggregateTimeSteps(x, nYears = 5)
}
\arguments{
\item{x}{magpie object ...}

\item{nYears}{integer ...}
}
\description{
aggregate values to 5-year averages to suppress volatility
}
\author{
Robin Hasse
}
