% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateWithoutGlobal.R
\name{toolAggregateWithoutGlobal}
\alias{toolAggregateWithoutGlobal}
\title{Aggregate regions without global sum}
\usage{
toolAggregateWithoutGlobal(x, glo, rel, to = NULL, removeAllAgg = FALSE)
}
\arguments{
\item{x}{a magclass object in country resolution}

\item{glo}{magclass object supplying explicit global data}

\item{rel}{aggregation mapping for \code{toolAggregate}}

\item{to}{aggregation target for \code{toolAggregate}}

\item{removeAllAgg}{decide whether to exclude all aggregated data or keep
those (variables/periods) that are not overwritten by data from glo object}
}
\value{
magclass object
}
\description{
Aggregate regional data, but if a global region exists, discard the
automatically aggregated global values and replace them with global data.
}
\author{
Falk Benke, Pascal Weigmann
}
