% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolBiomassSupplyAggregate.R
\name{toolBiomassSupplyAggregate}
\alias{toolBiomassSupplyAggregate}
\title{toolBiomassSupplyAggregate
The function aggregates biomass supply curves to regionmapping different from H12.
It only works if all regions are subregions of H12 regions. The offset parameter (a) is taken from the H12 region.
The slope parameter (b) is multiplied by a weight. The weight is the inverse of the share of agricultural area of the subregion in the H12 region.}
\usage{
toolBiomassSupplyAggregate(
  x,
  rel = NULL,
  weight = calcOutput("FAOLand", aggregate = F)[, , "6610", pmatch = TRUE][, "y2010", ]
)
}
\arguments{
\item{x}{magclass object that should be aggregated}

\item{rel}{relation matrix containing a region mapping.}

\item{weight}{aggregation weight}
}
\value{
return: returns region aggregated biomass supply curve data
}
\description{
toolBiomassSupplyAggregate
The function aggregates biomass supply curves to regionmapping different from H12.
It only works if all regions are subregions of H12 regions. The offset parameter (a) is taken from the H12 region.
The slope parameter (b) is multiplied by a weight. The weight is the inverse of the share of agricultural area of the subregion in the H12 region.
}
\author{
Felix Schreyer
}
