% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalcGhgFactor.R
\name{toolCalcGhgFactor}
\alias{toolCalcGhgFactor}
\title{Calculate GHG Emission Factors from GHG emission targets}
\usage{
toolCalcGhgFactor(x, subtype, subset)
}
\arguments{
\item{x}{a magclass object with targets read in from NDC or NPI database}

\item{subtype}{Emissions_YYYY_cond or Emissions_YYYY_uncond}

\item{subset}{String, designating the GDP scenarios to use}
}
\description{
Emission targets are represented by a GHG Emission Factor, which is the quotient of total GHG
emissions in the target year divided by the CEDS GHG emissions in 2005.
}
\seealso{
\code{\link[=convertUNFCCC_NDC]{convertUNFCCC_NDC()}}, \code{\link[=convertNewClimate]{convertNewClimate()}}
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Sophie Fuchs, Rahel Mandaroux, Falk Benke
}
