% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalcGhgFactor.R
\name{toolCalcGhgTarget}
\alias{toolCalcGhgTarget}
\title{Calculate absolute emission targets}
\usage{
toolCalcGhgTarget(x, subtype, subset)
}
\arguments{
\item{x}{a magclass object with targets read in from NDC or NPI database}

\item{subtype}{Emissions_YYYY_cond or Emissions_YYYY_uncond}

\item{subset}{String, designating the GDP scenarios to use}
}
\description{
Calculate absolute emission targets depending on country-specific emissions target formulations.
So far, the function mainly used to calculate NDC emissions targets.
}
\seealso{
\code{\link[=convertUNFCCC_NDC]{convertUNFCCC_NDC()}}, \code{\link[=convertNewClimate]{convertNewClimate()}}
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Sophie Fuchs, Rahel Mandaroux, Falk Benke
}
