% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalcIEAfromStructureMappingPEFE.R
\name{toolCalcIEAfromStructureMappingPEFE}
\alias{toolCalcIEAfromStructureMappingPEFE}
\title{toolCalcIEAfromStructureMappingPEFE}
\usage{
toolCalcIEAfromStructureMappingPEFE(data, structureMapping,
  subtype = "remind")
}
\arguments{
\item{data}{data to map}

\item{structureMapping}{mapping to use}

\item{subtype}{remind (default), edge, pfu or magpie}
}
\value{
MAgPIE object with completed time dimensionality.
}
\description{
mapping IEA reported data to ReMIND-MAgPIE items
}
\author{
Anastasis Giannousakis, Lavinia Baumstark, Isabelle Weindl
}
