% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCell2isoCell.R
\name{toolCell2isoCell}
\alias{toolCell2isoCell}
\title{toolCell2isoCell}
\usage{
toolCell2isoCell(x)
}
\arguments{
\item{x}{magpie object on cellular level}
}
\value{
return changed input data
}
\description{
Sets cell names to [iso country code].[cell number]
}
\author{
Kristine Karstens
}
