% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFAOcheck.R
\name{toolFAOcheck}
\alias{toolFAOcheck}
\title{toolFAOcheck}
\usage{
toolFAOcheck()
}
\value{
should return a report
}
\description{
Checks the FAO datasets for known problems. Among other things it compares
whether production values from ProdSTAT and Commodity Balances agree,
whether values are lost via aggregation.
}
\author{
Ulrich Kreidenweis
}
