% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFAOcombine.R
\name{toolFAOcombine}
\alias{toolFAOcombine}
\title{Combine FAO datasets}
\usage{
toolFAOcombine(..., combine = "Item")
}
\arguments{
\item{...}{two magpie objects with FAO data}

\item{combine}{"Item" to combine datasets that for instance both contain
palm oil data}
}
\value{
MAgPIE object with data from both inputs but dublicates removed
}
\description{
Allows to combine two similar FAO datasets with dublicates being removed.
For instance combine Production:Crops and Production: Crops Processed to one
magpie object
}
\examples{

  \dontrun{ a <- toolFAOcombine(Crop,CropPro, combine="Item")
  }
  
}
\seealso{
\code{\link{readSource}}
}
\author{
Ulrich Kreidenweis
}
