% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFillEU34Countries.R
\name{toolFillEU34Countries}
\alias{toolFillEU34Countries}
\title{Sets values for 6 EU countries not belonging to EU 28 but EU 34 to zero if
they are NA. Used to avoid EUR region not yielding NA because of these countries.}
\usage{
toolFillEU34Countries(x)
}
\arguments{
\item{x}{magpie object with 249 ISO country codes in the spatial dimension}
}
\description{
Sets values for 6 EU countries not belonging to EU 28 but EU 34 to zero if
they are NA. Used to avoid EUR region not yielding NA because of these countries.
}
\author{
Falk Benke
}
