% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFillYearsWithClosest.R
\name{toolFillYearsWithClosest}
\alias{toolFillYearsWithClosest}
\title{Fills years in a magpie object with their closest temporal neighbor
When a NA is in the middle of two valid years, choose the previous one}
\usage{
toolFillYearsWithClosest(inx)
}
\arguments{
\item{inx}{a magclass object with NAs}
}
\value{
the magclass object with some of those NAs filled
}
\description{
Does not do anything for slices where all timesteps are NA
}
