% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetIEAYear.R
\name{toolGetIEAYear}
\alias{toolGetIEAYear}
\title{Returns the year associated with a given ieaVersion}
\usage{
toolGetIEAYear(ieaVersion)
}
\arguments{
\item{ieaVersion}{Release version of IEA data, either 'default' (vetted and used in REMIND)
or 'latest'.}
}
\description{
Returns the year associated with a given ieaVersion
}
\author{
Falk Benke
}
