% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHoldConstantBeyondEnd.R
\name{toolHoldConstantBeyondEnd}
\alias{toolHoldConstantBeyondEnd}
\title{toolHoldConstantBeyondEnd}
\usage{
toolHoldConstantBeyondEnd(x)
}
\arguments{
\item{x}{MAgPIE object to be continued.}
}
\value{
MAgPIE object with completed time dimensionality.
}
\description{
Holds a historical dataset constant for the entire simulation period "time".
}
\author{
Benjamin Leon Bodirsky
}
