% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIntake2.R
\name{toolIntake2}
\alias{toolIntake2}
\title{toolIntake2}
\usage{
toolIntake2(bodyweight, bodyheight = NULL, inactivity, tmean = NULL,
  method = NULL)
}
\arguments{
\item{bodyweight}{bodyweight in kg per capita or "standardized" for assuming standard values}

\item{bodyheight}{for mehthod FAO_WHO_UNU1985}

\item{inactivity}{the activity level and the bodyweight computed by the readHic2015 function}

\item{tmean}{mean annual temperature}

\item{method}{method for calculating intake: either FAO_WHO_UNU1985 for estimates based on height and bodyweight, schofield for just bodyweight, or HHS_USDA for recommended values for US-americans}
}
\description{
it computes the food intake pro capita through the bodyweight 
and the activity level. First it computes the basal metabolic rate (bmr) through 
the Schofield equation and then the estimated energy required (eer) depending on 
the activitiy level by FAO/WHO/UNU tables (Human Energy Requirments, Rome 2004)
}
\author{
Eleonora Martinelli
}
