% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIso2CellCountries.R
\name{toolIso2CellCountries}
\alias{toolIso2CellCountries}
\title{toolIso2CellCountries}
\usage{
toolIso2CellCountries(x)
}
\arguments{
\item{x}{magpie object on iso country level}
}
\value{
return selected input data
}
\description{
Select country names of countries which are present on cellular level
}
\author{
Kristine Karstens
}
