% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIsocode2Country.R
\name{toolIsocode2Country}
\alias{toolIsocode2Country}
\title{toolIsocode2Country}
\usage{
toolIsocode2Country(x)
}
\arguments{
\item{x}{Array of iso country codes}
}
\value{
return array of country names
}
\description{
Translate iso country code to country names
}
\author{
Kristine Karstens
}
