% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolProcessClimateTargetDatabase.R
\name{toolProcessClimateTargetDatabase}
\alias{toolProcessClimateTargetDatabase}
\title{Helper to validate data read in from UNFCCC NDC and New Climate databases and
apply some pre-processing}
\usage{
toolProcessClimateTargetDatabase(input, database, subtype)
}
\arguments{
\item{input}{data frame representing the data from climate target database}

\item{database}{database to be read in, used for logging info}

\item{subtype}{database version to be read in, used for logging info}
}
\description{
Helper to validate data read in from UNFCCC NDC and New Climate databases and
apply some pre-processing
}
\seealso{
\code{\link[=readUNFCCC_NDC]{readUNFCCC_NDC()}}, \code{\link[=readNewClimate]{readNewClimate()}}
}
