% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSolarFunctionAggregate.R
\name{toolSolarFunctionAggregate}
\alias{toolSolarFunctionAggregate}
\title{toolSolarFunctionAggregate}
\usage{
toolSolarFunctionAggregate(
  x,
  rel = NULL,
  weight = calcOutput("FE", aggregate = FALSE)[, "y2015", "FE|Electricity (EJ/yr)"]
)
}
\arguments{
\item{x}{magclass object that should be aggregated}

\item{rel}{relation matrix containing a region mapping.
A mapping object should contain 2 columns in which each element of x
is mapped to the category it should belong to after (dis-)aggregation}

\item{weight}{aggregation weight (should be FE|Electricity (EJ/yr) in 2015)}
}
\value{
return: returns region aggregated solar data
}
\description{
Aggregate Solar data into regions
}
\author{
Felix Schreyer, Renato Rodrigues, Julian Oeser
}
