% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUSGS.R
\name{readUSGS}
\alias{readUSGS}
\alias{convertUSGS}
\title{Read U.S. Geological Survey data}
\usage{
readUSGS(subtype = "cement")

convertUSGS(x, subtype = "cement")
}
\arguments{
\item{subtype}{One of
\itemize{
\item 'cement': read cement production data from
\href{https://www.usgs.gov/centers/national-minerals-information-center/cement-statistics-and-information}{U.S. Geological Survey Minerals Yearbook}
(unit: tonnes per year)
}}

\item{x}{Data returned by \code{\link[=readUSGS]{readUSGS()}}.}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read U.S. Geological Survey data
}
\author{
Michaja Pehl
}
