% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAGEB.R
\name{calcAGEB}
\alias{calcAGEB}
\title{Calculate REMIND final energy variables from historical AGEB values}
\usage{
calcAGEB()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND final energy variables from historical AGEB values
}
\author{
Falk Benke
}
