% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAirPollEmiRef.R
\name{calcAirPollEmiRef}
\alias{calcAirPollEmiRef}
\title{Calculate air pollutant emissions for a reference year, for use
in combination with GAINS data at different sectoral aggregations}
\usage{
calcAirPollEmiRef(
  subtype = "total",
  baseyear = 2015,
  outunits = "Mt/yr",
  namesformat = "GAINS2025"
)
}
\arguments{
\item{subtype}{"total" or "sectorsCEDS16"}

\item{baseyear}{year to take as a reference from CEDS, ignored for the EDGAR2005 LUC CO2 emissions}

\item{outunits}{"Mt/yr" or "kt/yr"}

\item{namesformat}{"GAINS2025" or "REMIND" or "REMINDexo", the standard to use for pollutant names}
}
\value{
magclass object
}
\description{
Calculate air pollutant emissions for a reference year, for use
in combination with GAINS data at different sectoral aggregations
}
\author{
Gabriel Abrahao
}
