% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBP.R
\name{calcBP}
\alias{calcBP}
\title{Calculate REMIND variables from historical BP values}
\usage{
calcBP(subtype)
}
\arguments{
\item{subtype}{Either "Emission", "Capacity", "Generation", "Consumption", "Trade" or "Price"}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND variables from historical BP values
}
\author{
Falk Benke
}
