% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapTarget.R
\name{calcCapTarget}
\alias{calcCapTarget}
\title{Capacity targets from two sources}
\usage{
calcCapTarget(sources)
}
\arguments{
\item{sources}{Database source}
}
\description{
The capacity targets (GW)  at regional level are produced from two different databases-
Rogelj 2017 paper (see readme in inputdata), and REN21 Global Renewables report
The Rogelj2017 capacity targets are further broken down to conditional and unconditional targets.
}
\author{
Aman Malik, Oliver Richters
}
