% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapTarget.R
\name{calcCapTarget}
\alias{calcCapTarget}
\title{Capacity targets from two sources}
\usage{
calcCapTarget(sources, condition)
}
\arguments{
\item{sources}{Database source}

\item{condition}{if applicable, conditional or unconditional}
}
\description{
The capacity targets (GW)  at regional level are produced from two different databases.
The Rogelj2017 capacity targets are further broken down to conditional and unconditional targets.
}
\author{
Aman Malik
}
