% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCCScapacity.R
\name{calcCCScapacity}
\alias{calcCCScapacity}
\title{calc CCS capacity}
\usage{
calcCCScapacity(subtype)
}
\arguments{
\item{subtype}{either `historical` for data from until 2022 or `projections`
for projections in 2025 and 2030}
}
\description{
Calculate CCS capacity from IEA CCUS data
}
\author{
Anne Merfort, Falk Benke
}
