% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCostOfCapital.R
\name{calcCostOfCapital}
\alias{calcCostOfCapital}
\title{Calculate capital costs for energy technologies}
\usage{
calcCostOfCapital()
}
\description{
Provides REMIND data for 25_WACC Module
}
\author{
Diamantis Koutsandreas
}
