% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEDGETransport.R
\name{calcEDGETransport}
\alias{calcEDGETransport}
\title{Prepare EDGETransport inputs

All subtypes should be used with the aggregate=FALSE flag.}
\usage{
calcEDGETransport(subtype = "logit_exponent")
}
\arguments{
\item{subtype}{refer to the code for a list of allowed subtypes.}
}
\value{
magpie object of EDGEtransport iterative inputs
}
\description{
Prepare EDGETransport inputs

All subtypes should be used with the aggregate=FALSE flag.
}
\examples{
\dontrun{ a <- calcOutput(type="EDGETransport", subtype="logit_exponent", aggregate=F)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Alois Dirnaichner, Marianna Rottoli
}
