% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEEAGHGProjections.R
\name{calcEEAGHGProjections}
\alias{calcEEAGHGProjections}
\title{Calculate EEA emission projections from the two projections sources provided by EEA}
\usage{
calcEEAGHGProjections()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate EEA emission projections from the two projections sources provided by EEA
}
\author{
Falk Benke
}
