% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiLULUCFCountryAcc.R
\name{calcEmiLULUCFCountryAcc}
\alias{calcEmiLULUCFCountryAcc}
\title{calcEmiLULUCFCountryAcc}
\usage{
calcEmiLULUCFCountryAcc()
}
\value{
Magpie object with historical LULUCF emissions
}
\description{
historical LULUCF CO2 emissions following country accounting
}
\author{
Felix Schreyer, Falk Benke
}
