% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiLULUCFCountryAcc.R
\name{calcEmiLULUCFCountryAcc}
\alias{calcEmiLULUCFCountryAcc}
\title{calcEmiLULUCFCountryAcc}
\usage{
calcEmiLULUCFCountryAcc(subtype)
}
\arguments{
\item{subtype}{Valid subtypes are 'UNFCCC'}
}
\value{
Magpie object with historical LULUCF emissions
}
\description{
hisorical LULUCF emissions following country accounting
}
\author{
Felix Schreyer
}
