% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionFactors.R
\name{calcEmissionFactors}
\alias{calcEmissionFactors}
\title{Calc Emission Factors}
\usage{
calcEmissionFactors(
  subtype = "emission_factors",
  sectoral_resolution = "aggregated"
)
}
\arguments{
\item{subtype}{either 'emission_factors', 'activities', or 'emissions'}

\item{sectoral_resolution}{must be 'aggregated'}
}
\description{
Calc Emission Factors
}
