% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissions.R
\name{calcEmissions}
\alias{calcEmissions}
\title{calcEmissions}
\usage{
calcEmissions(datasource = "CEDS16")
}
\arguments{
\item{datasource}{"CEDS16", "CEDS2REMIND", "CEDS2024", "EDGAR", "EDGAR6",
"EDGAR8" "LIMITS", "ECLIPSE", "GFED", "CDIAC"}
}
\value{
magpie object with historical emissions
}
\description{
calcEmissions
}
\author{
Steve Smith, Pascal Weigmann
}
