% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissions4ReportExtra.R
\name{calcEmissions4ReportExtra}
\alias{calcEmissions4ReportExtra}
\title{Calculates historical emissions needed for estimating emission factors
in extra emissions reporting}
\usage{
calcEmissions4ReportExtra()
}
\value{
list of magclass with CEDS CH4 and N2O emissionsdata for
CEDS sectors for timesteps 2000-2100, in Mt N or Mt CH4 per year
}
\description{
Calculates historical emissions needed for estimating emission factors
in extra emissions reporting
}
\author{
Gabriel Abrhahao
}
