% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissions4ReportExtra.R
\name{calcEmissions4ReportExtra}
\alias{calcEmissions4ReportExtra}
\title{Calculates historical emissions needed for estimating emission factors
in extra emissions reporting}
\usage{
calcEmissions4ReportExtra()
}
\arguments{
\item{source}{either "PBL_2007" or "PBL_2022"}
}
\value{
list of magclass with REMIND input data for
different sectors for timesteps 2000-2100, in Mt N per year
}
\description{
Calculates historical emissions needed for estimating emission factors
in extra emissions reporting
}
\author{
Gabriel Abrhahao
}
