% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTarget.R
\name{calcEmiTarget}
\alias{calcEmiTarget}
\title{Calculate NDC Emissions Targets}
\usage{
calcEmiTarget(sources, subtype, scenario)
}
\arguments{
\item{sources}{database source, either 'UNFCCC_NDC' or 'NewClimate'}

\item{subtype}{must be one of
\itemize{
\item 'Ghgfactor': target factors (target year emissions normalized by 2015 emissions)
\item 'Ghgshare': share of emissions covered under NDC target per REMIND region
}}

\item{scenario}{GDP and pop scenarios. Passed to \code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDP()}}.
turned off for inpudata generation}
}
\description{
This function calculates the emissions targets for the NDC scenarios applied in the REMIND module 45_carbonprice realization NDC.
It contains the following steps:
The parameters calculated in 3.) and 4.) are further used in the NDC realization to calculate the region-wide NDC emissions targets
in terms of total GHG emissions excl. bunkers and excl. LULUCF sectors.
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Rahel Mandaroux, Falk Benke
}
