% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTarget.R
\name{calcEmiTarget}
\alias{calcEmiTarget}
\title{Output for 2 policy cases}
\usage{
calcEmiTarget(subtype)
}
\arguments{
\item{subtype}{"share_uncond", "share_cond", "multiplier_uncond",  "multiplier_cond", "ghg_uncond" or "ghg_cond"}
}
\description{
Output for 2 policy cases
}
\author{
Aman Malik, Christoph Bertram
}
