% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTarget.R
\name{calcEmiTarget}
\alias{calcEmiTarget}
\title{Calculate Emission Targets}
\usage{
calcEmiTarget(sources, subtype, scenario)
}
\arguments{
\item{sources}{database source, either 'UNFCCC_NDC' or 'NewClimate'}

\item{subtype}{must be one of
\itemize{
\item 'Ghgfactor': for GHG factors calculated from the respective database
\item 'Ghgshare2015': ???
}}

\item{scenario}{GDP and pop scenarios. Passed to \code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDP()}}.
turned off for inpudata generation}
}
\description{
TODO: describe what this is used for in REMIND and document the meaning of the subtypes
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Rahel Mandaroux, Falk Benke
}
