% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTarget.R
\name{calcEmiTarget}
\alias{calcEmiTarget}
\title{Output for 2 policy cases}
\usage{
calcEmiTarget(sources, subtype)
}
\arguments{
\item{sources}{currently only UNFCCC_NDC}

\item{subtype}{"Ghgshare2005", "Ghgfactor", "Ghghistshare"}
}
\description{
Output for 2 policy cases
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters
}
