% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiTargetReference.R
\name{calcEmiTargetReference}
\alias{calcEmiTargetReference}
\title{Calculate Emission Targets reference to be used when calculation Emission Targets}
\usage{
calcEmiTargetReference()
}
\description{
Uses historical CEDS emissions from 1990-2023
CO2 (excl LU), CH4, N2O (so far no F-Gas historic time series)
}
\seealso{
\code{\link[=calcEmiTarget]{calcEmiTarget()}}, \code{\link[=convertUNFCCC_NDC]{convertUNFCCC_NDC()}}
}
\author{
Rahel Mandaroux, Falk Benke
}
