% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpertGuess.R
\name{calcExpertGuess}
\alias{calcExpertGuess}
\title{Calculate expert guesses}
\usage{
calcExpertGuess(subtype)
}
\arguments{
<<<<<<< HEAD
\item{subtype}{must be one of \code{biocharPrices}, \code{gridFactor}, \code{tradeConstraints}}
=======
\item{subtype}{must be one of
'biocharPrices'
'subConvergenceRollback'
'tradeConstraints'
'taxConvergence'
'taxConvergenceRollback'}
>>>>>>> taxConv
}
\description{
Calculate expert guesses
}
\author{
Falk Benke
}
