% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpertGuess.R
\name{calcExpertGuess}
\alias{calcExpertGuess}
\title{Calculate expert guesses}
\usage{
calcExpertGuess(subtype)
}
\arguments{
\item{subtype}{must be one of
'biocharPrices'
'ccsBounds'
'deltacapoffset'
'subConvergenceRollback'
'tradeConstraints'
'taxConvergence'
'taxConvergenceRollback'}
}
\description{
Calculate expert guesses
}
\author{
Falk Benke
}
