% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFE.R
\name{calcFE}
\alias{calcFE}
\title{Calculates FE historical from IEA energy balances, projections from EDGE, and historical
values from IEA WEO 2019}
\usage{
calcFE(source = "IEA", scenario_proj = "SSP2")
}
\arguments{
\item{source}{"IEA", "EDGE_projections", or "IEA_WEO"}

\item{scenario_proj}{"SSP2" by default unless overwritten}
}
\description{
Calculates FE historical from IEA energy balances, projections from EDGE, and historical
values from IEA WEO 2019
}
\author{
Lavinia Baumstark, Aman Malik
}
