% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFE.R
\name{calcFE}
\alias{calcFE}
\title{Calculates FE historical from IEA energy balances}
\usage{
calcFE(ieaVersion = "default")
}
\arguments{
\item{ieaVersion}{Release version of IEA data, either 'default' (vetted and used in REMIND)
or 'latest'.}
}
\description{
Calculates FE historical from IEA energy balances
}
\author{
Lavinia Baumstark, Aman Malik
}
