% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEdemand.R
\name{calcFEdemand}
\alias{calcFEdemand}
\title{Edge data with its original sectoral division}
\usage{
calcFEdemand(subtype, use_ODYM_RECC = FALSE)
}
\arguments{
\item{subtype}{Final energy (FE) or Useful/Final
Energy items from EDGEv3 corresponding to REMIND FE items (UE_for_Eff,
FE_for_Eff)}

\item{use_ODYM_RECC}{per-capita pathways for `SDP_xx` scenarios?  (Defaults
to `FALSE`.)}
}
\description{
Returns the Edge data at the Remind level
}
\author{
Michaja Pehl, Robin Hasse, Falk Benke
}
