% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEdemand.R
\name{calcFEdemand}
\alias{calcFEdemand}
\title{Edge data with its original sectoral division}
\usage{
calcFEdemand(subtype = "FE")
}
\arguments{
\item{subtype}{Final energy (FE) or Energy service (ES) or Useful/Final Energy items from EDGEv3 corresponding to REMIND FE items (UE_for_Eff,FE_for_Eff)}
}
\description{
Returns the Edge data at the Remind level
}
\author{
Antoine Levesque
}
