% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFEdemand.R
\name{calcFEdemand}
\alias{calcFEdemand}
\title{Calculates Final Energy Demand for Industry, Buildings and Transport}
\usage{
calcFEdemand()
}
\description{
Calculates Final Energy Demand for Industry, Buildings and Transport
}
\author{
Falk Benke
}
