% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandBuildings.R
\name{calcFeDemandBuildings}
\alias{calcFeDemandBuildings}
\title{Returns the EDGE-Buildings data as REMIND variables}
\usage{
calcFeDemandBuildings(subtype)
}
\arguments{
\item{subtype}{either "FE", "FE_buildings", or "UE_buildings"}
}
\description{
Returns the EDGE-Buildings data as REMIND variables
}
\author{
Robin Hasse
}
