% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandONONSPEC.R
\name{calcFeDemandONONSPEC}
\alias{calcFeDemandONONSPEC}
\title{Calculate historic and projected other non-specified energy demand}
\usage{
calcFeDemandONONSPEC(scenario, eoh)
}
\arguments{
\item{scenario}{character vector of remind demand scenarios}

\item{eoh}{numeric, end of history: last time step without scenario
differentiation}
}
\value{
list with MagPIE object
}
\description{
Project the IEA flow ONONSPEC into the future. As we have no idea, where this
energy demand comes from, we use a very generic methos to project it into the
future: We assume an asymptotic model. It starts from the level
\eqn{x_\text{EOD}} at the end of data (EOD) with EOD slope
\eqn{\dot{x}_\text{EOD}} and approaches the fraction \eqn{\varepsilon} of
this slope within \eqn{\Delta t}. Both \eqn{x_\text{EOD}} and
\eqn{\dot{x}_\text{EOD}} are determined through a linear regression of the
last \eqn{n} time steps with IEA data.
\deqn{x(t) = x_\text{EOD} + \dfrac{\dot{x}_\text{EOD}}{c}
      \cdot [1 - \exp (-c \cdot (t - t_\text{EOD}))]}
with the decay rate \eqn{c = -\dfrac{\ln \varepsilon}{\Delta t}} \cr
\eqn{\Delta t} is differentiated by scenarios thus approaching a \code{low},
\code{med} and \code{high} long-term level. To assure that scenarios don't differ at
the end of history (EOH), time steps between EOD and EOH are projected with
\code{med} value of \eqn{\Delta t}.
}
\details{
Each scenario \eqn{s} has a differentiated \eqn{\Delta t_s}. For
\eqn{\dot{x}_\text{EOD} > 0}, the \code{high} (\code{low}) scenario assumes a longer
(shorter) time span and for \eqn{\dot{x}_\text{EOD} < 0} vice versa to reach
a higher (lower) long-term value. We want to make sure that until the end of
history (EOH), all scenarios are still identical. So we take the \code{med}
parameterisation until EOH. Afterwards, we adjust the model such that we
start with EOH level and slope and still reach the target slope
\eqn{\varepsilon \cdot \dot{x}_\text{EOD}} until
\eqn{t_\text{EOD} + \Delta t_s}:
\deqn{x_s(t) = x_\text{EOH} + \dfrac{\dot{x}_\text{EOH}}{c_s}
      \cdot [1 - \exp (-c_s \cdot (t - t_\text{EOH}))]}
with the decay rate
\eqn{c_s = -\dfrac{\ln \varepsilon }
                  {\Delta t_s - (t_\text{EOH} - t_\text{EOD})}
     \cdot \left(1 - \dfrac{t_\text{EOH} - t_\text{EOD}}
                           {\Delta t_\text{med}} \right)}
}
\author{
Robin Hasse
}
