% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFGas.R
\name{calcFGas}
\alias{calcFGas}
\title{generate F-Gases based on IMAGE data}
\usage{
calcFGas()
}
\value{
magpie object with F-gases information
}
\description{
generate F-Gases based on IMAGE data
}
\examples{

  \dontrun{
    x <- calcOutput("FGas")
  }
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
