% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFGas.R
\name{calcFGas}
\alias{calcFGas}
\title{generate F-Gases based on IMAGE data}
\usage{
calcFGas(subtype = "interpolate2025")
}
\arguments{
\item{subtype}{"interpolate2025" will intepolate from EDGAR historical data from 2025-2050 to account for the very old IMAGE
scenarios. Any other subtype will ignore this step.}
}
\value{
magpie object with F-gases information
}
\description{
generate F-Gases based on IMAGE data
}
\examples{
\dontrun{
x <- calcOutput("FGas")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
