% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFloorspace.R
\name{calcFloorspace}
\alias{calcFloorspace}
\title{Floor space in buildings}
\usage{
calcFloorspace(scenario, onlyTotal = FALSE)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned.}

\item{onlyTotal}{boolean, only give total instead of sub-sectoral floor space}
}
\value{
MAgPIE object with buildings floor space
}
\description{
Residential, commercial and total floor space from EDGE-B. Set
}
\author{
Antoine Levesque, Robin Hasse
}
