% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGAINS2025forREMIND.R
\name{calcGAINS2025forREMIND}
\alias{calcGAINS2025forREMIND}
\title{Calculates air pollutant emissions, activities and emission factors
for all scenarios and SSPs available from GAINS, including both
legacy and 2025 variants of GAINS data. This generates the actual
files needed for REMIND in the proper format, for the scenario logic,
see \code{GAINS2025scenarios}.}
\usage{
calcGAINS2025forREMIND(subtype)
}
\arguments{
\item{subtype}{"emission_factors", "emissions","emission_factors_remindsectors",
"emissions_starting_values" (not implemented)}
}
\value{
Activity levels, emissions or emission factors
}
\description{
Calculates air pollutant emissions, activities and emission factors
for all scenarios and SSPs available from GAINS, including both
legacy and 2025 variants of GAINS data. This generates the actual
files needed for REMIND in the proper format, for the scenario logic,
see \code{GAINS2025scenarios}.
}
