% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGHGPrices.R
\name{calcGHGPrices}
\alias{calcGHGPrices}
\title{calcGHGPrices}
\usage{
calcGHGPrices(emissions = "pollutants", datasource = "REMMAG")
}
\arguments{
\item{emissions}{which type of emissions shall be returned. ghg just returns n2o, ch4 and co2, pollutants a longer list including also air pollutants}

\item{datasource}{REMIND for prices from R2M4 coupled runs, REMMAG for old coupled runs, SSPResults for prices from the SSP scenarios from the IIASA database, SSP_and_REM for a combination of REMIND and SSPResults}
}
\value{
list of magpie object with results on country level, weight on country level, unit and description.
}
\description{
reads in GHG prices from past runs
}
\examples{
  

\dontrun{ 
calcOutput("GHGPrices")
}
}
\seealso{
\code{\link{readSSPResults}}
}
\author{
David Chen, Benjamin Leon Bodirsky, David Klein
}
