% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGlobalEnergyMonitor.R
\name{calcGlobalEnergyMonitor}
\alias{calcGlobalEnergyMonitor}
\title{Calc capacities from Global Energy Monitor}
\usage{
calcGlobalEnergyMonitor()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate near-term expectations of capacities for use in fullVALIDATION.R
}
\author{
Falk Benke
}
