% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmber.R, R/calcHRE.R
\name{calcHRE}
\alias{calcHRE}
\title{Calculate Electricity Data from Ember}
\usage{
calcHRE()

calcHRE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.

A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate Electricity Data from Ember

Calculate Final Energy for the buildings sector from Heat Roadmap Europe scenarios
}
\author{
Pascal Weigmann
}
