% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_ETP.R
\name{calcIEA_ETP}
\alias{calcIEA_ETP}
\title{Calculate REMIND emission variables from IEA ETP values}
\usage{
calcIEA_ETP()
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Calculate REMIND emission variables from IEA ETP values
}
\author{
Falk Benke
}
