% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_EVOutlook.R
\name{calcIEA_EVOutlook}
\alias{calcIEA_EVOutlook}
\title{Calculate REMIND variables from IEA EV Outlook data}
\usage{
calcIEA_EVOutlook()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND variables from IEA EV Outlook data
}
\author{
Falk Benke
}
