% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_WEO_2021.R
\name{calcIEA_WEO_2021}
\alias{calcIEA_WEO_2021}
\title{Calculate REMIND variables from IEA WEO 2021 data}
\usage{
calcIEA_WEO_2021(subtype = "GLO")
}
\arguments{
\item{subtype}{Either "GLO" or "regional"}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND variables from IEA WEO 2021 data
}
\author{
Falk Benke
}
