% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIO.R
\name{calcIO}
\alias{calcIO}
\title{Calc Input Output}
\usage{
calcIO(subtype)
}
\arguments{
\item{subtype}{data subtype. Valid subtypes are 'input', 'output_EDGE', 
'trade', 'output_EDGE_buildings', 'output' and 
'output_Industry_subsectors'}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes IEA-based model data for different "subtypes" by use of raw IEA "Energy Balances" data
and a mapping that corresponds to the structure of "products" and "flows" of IEA.
}
\details{
Mapping structure example: IEA product ANTCOAL used for IEA flow TPATFUEL, contributes via REMIND technology
coaltr for generating sesofos from pecoal (REMIND names)

When using subtype \code{output_Industry_subsectors}, additional correctons 
are applied to the IEA data in 
\code{\link{fix_IEA_data_for_Industry_subsectors}}.
}
\examples{

\dontrun{ a <- calcOutput("IO", subtype = "output")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Anastasis Giannousakis
}
