% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMacroInvestments.R
\name{calcMacroInvestments}
\alias{calcMacroInvestments}
\title{Calculate 2005 macroeconomic capital investments}
\usage{
calcMacroInvestments()
}
\description{
Compute macroeconomic capital investments based on investments shares from the PWT and GDP scenarios from
\link{mrdrivers}. The final investments are the product of the two.
}
\seealso{
\itemize{
\item See the vignette \code{vignette("scenarios", "mrdrivers")} for information on the GDP scenarios.
\item \code{\link[=readPWT]{readPWT()}} for information on the PWT version used.
}
}
