% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMacroWACC.R
\name{calcMacroWACC}
\alias{calcMacroWACC}
\title{Calculate capital costs for REMIND regions}
\usage{
calcMacroWACC()
}
\description{
Provides REMIND data for 25_WACC Module
}
\author{
Diamantis Koutsandreas
}
