% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialHydro.R
\name{calcPotentialHydro}
\alias{calcPotentialHydro}
\title{Calculate hydro potential}
\usage{
calcPotentialHydro()
}
\value{
hydro potential data and corresponding weights as a list of two
MAgPIE objects
}
\description{
Provides hydro potential data
}
\examples{

\dontrun{
calcOutput("PotentialHydro")
}

}
\seealso{
\code{\link{readWGBU}}, \code{\link{convertWGBU}}
}
\author{
Lavinia Baumstark
}
