% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialWindOff.R
\name{calcPotentialWindOff}
\alias{calcPotentialWindOff}
\title{Calculate wind offshore potential}
\usage{
calcPotentialWindOff()
}
\value{
wind offshore potential data and corresonding weights as a list of
two MAgPIE objects
}
\description{
Provides wind offshore potential data
}
\examples{

\dontrun{
calcOutput("PotentialWindOff")
}
}
\seealso{
\code{\link{readNREL}}, \code{\link{convertNREL}}
}
\author{
Chen Chris Gong
}
