% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProjectPipelines.R
\name{calcProjectPipelines}
\alias{calcProjectPipelines}
\title{calc Project Pipelines}
\usage{
calcProjectPipelines(subtype)
}
\arguments{
\item{subtype}{choose technology \code{biomass}, \code{coal}, \code{geothermal}, \code{hydro},
\code{nuclear}, \code{solar}, \code{wind} or \code{CCS}}
}
\description{
Calculate the expected near-term deployment of technologies based on
projects that are currently either being built or in a planning stage.
For some technologies multiple sources are available. Output object currently
needs to contain years 2020, 2025 and 2030.
}
\details{
Discussions on sources and assumptions:
https://github.com/pik-piam/mrremind/discussions
}
\author{
Pascal Weigmann
}
