% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calcSE}
\alias{calcSE}
\title{Calculate Secondary Energy variables}
\usage{
calcSE()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate Secondary Energy variables
}
\author{
Pascal Weigmann
}
