% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSolar.R
\name{calcSolar}
\alias{calcSolar}
\title{calcSolar
calculate Area, Capacity and Energy for photovoltaics (PV) and contentrated solar power (CSP)}
\usage{
calcSolar()
}
\value{
magpie object
}
\description{
calcSolar
calculate Area, Capacity and Energy for photovoltaics (PV) and contentrated solar power (CSP)
}
\examples{

\dontrun{ a <- calcOutput(type="Solar")
}


}
\seealso{
\code{\link{calcOutput}}
}
\author{
Julian Oeser, modified by Renato Rodrigues
}
