% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTheil.R
\name{calcTheil}
\alias{calcTheil}
\title{Get regional Theil-T index projections}
\usage{
calcTheil(scenario)
}
\arguments{
\item{scenario}{A string (or vector of strings) designating the scenario(s) to be returned. Passed as is
to the scenario argument of \link[mrdrivers:calcGDP]{mrdrivers::calcGDP}.}
}
\value{
magpie objects of unweighted contribution to Theil, weights (= country shares of regional GDP)
}
\description{
To calculate the regional Theil-T index (= correction to welfare function for a lognormal income distribution) we do
the following: (1) convert country-level Gini coefficients to Theil (2) calculate contribution to Theil-T index that
includes both between-countries and within-country inequality (see e.g. https://en.wikipedia.org/wiki/Theil_index).
The latter can then be aggregated with calcOutput().
}
\details{
The projections are SSP specific and use SSP GINI projections. For non-SSP scenarios, the projections are equal
to the SSP2 projection.

The aggregation depends on the region mapping. It is implemented such
that the regionmapping specified in getConfig()$regionmapping is used.
The result of calcOutput('Theil', aggregate = FALSE), is NOT the country
Theil-T, but the unweighted contribution from a given country to the regional value.
}
\seealso{
\code{\link{readGini}}
}
