% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUBA.R
\name{calcUBA}
\alias{calcUBA}
\title{Calculate REMIND final energy variables from historical UBA values}
\usage{
calcUBA()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND final energy variables from historical UBA values
}
\author{
Falk Benke
}
