% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUNFCCC.R
\name{calcUNFCCC}
\alias{calcUNFCCC}
\title{Calculate REMIND emission variables from historical UNFCCC values}
\usage{
calcUNFCCC()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Calculate REMIND emission variables from historical UNFCCC values
}
\author{
Falk Benke, Pascal Weigmann
}
