% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYield.R
\name{calcYield}
\alias{calcYield}
\title{calcYield}
\usage{
calcYield(physical = TRUE, attributes = "dm", irrigation = FALSE,
  cellular = FALSE, cut = FALSE)
}
\arguments{
\item{physical}{physical area or havested area}

\item{attributes}{in dm, wm, ge, nr, p, k}

\item{irrigation}{distinguish irrigation or not}

\item{cellular}{if TRUE value is calculate on cellular level}

\item{cut}{FALSE (default) - do not cut off yields, number between 0 and 1 to define percentile value for cut off}
}
\value{
MAgPIE object of yields
}
\description{
calculates the yield based on FAO data
}
\author{
Debbora Leip, Jan Philipp Dietrich
}
