% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAriadneDB.R
\name{convertAriadneDB}
\alias{convertAriadneDB}
\title{Convert Ariadne DB data}
\usage{
convertAriadneDB(x)
}
\arguments{
\item{x}{A \code{\link[magclass:magclass]{magpie}} object returned from
\code{\link[=readAriadneDB]{readAriadneDB()}}.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
convert Ariadne database data
}
\author{
Felix Schreyer
}
