% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBiocharDeploymentData.R
\name{convertBiocharDeploymentData}
\alias{convertBiocharDeploymentData}
\title{Provide Biochar price path data}
\usage{
convertBiocharDeploymentData(x)
}
\arguments{
\item{x}{a magpie object}
}
\description{
Provide Biochar price path data
}
\author{
Tabea Dorndorf
}
