% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBP.R
\name{convertBP}
\alias{convertBP}
\title{Disaggregates and cleans BP data.}
\usage{
convertBP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Either "Emission", "Capacity", "Generation", "Production",
"Consumption", "Trade Oil", "Trade Gas", "Trade Coal" or "Price"}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Disaggregates historical data.
}
\author{
Aman Malik, Falk Benke
}
