% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBP.R
\name{convertBP}
\alias{convertBP}
\title{Disaggregates and cleans BP data.}
\usage{
convertBP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Either "Capacity" or "Generation"}
}
\value{
A magpie object of the BP country disaggregated data  with historical electricity capacities 
(in MW for Wind, Solar, and Geothermal for 1997-2012) AND Generation
 (in TWh for Nuclear, Hydro, Wind, Solar, Other Renewables, and Geo Biomass for 1971-2012)
}
\description{


Disagregates istorical capacity and generation data.
}
\author{
Aman Malik
}
