% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGE.R
\name{convertEDGE}
\alias{convertEDGE}
\title{Convert EDGE}
\usage{
convertEDGE(x, subtype = "FE_stationary")
}
\arguments{
\item{x}{MAgPIE object containing EDGE values at ISO country resolution}

\item{subtype}{either FE_stationary, FE_buildings, or Floorspace}
}
\value{
EDGE data as MAgPIE object aggregated to country level
}
\description{
Convert EDGE data to data on ISO country level.
}
\author{
Antoine Levesque, Robin Hasse
}
