% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGETransport.R
\name{convertEDGETransport}
\alias{convertEDGETransport}
\title{Convert IEA}
\usage{
convertEDGETransport(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing EDGE values at ISO country resolution}

\item{subtype}{EDGE entries}
}
\value{
EDGETransport data as MAgPIE object aggregated to ISO level
}
\description{
Ship EDGETransport data through, as already on ISO level
}
\author{
Marianna Rottoli
}
