% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanEnergyDatasheets.R
\name{convertEuropeanEnergyDatasheets}
\alias{convertEuropeanEnergyDatasheets}
\title{Convert European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/country
}
\usage{
convertEuropeanEnergyDatasheets(x, subtype)
}
\arguments{
\item{x}{European Energy Datasheets magpie object derived from readEuropeanEnergyDatasheets function}

\item{subtype}{data subtype. Either "EU28" (older data from Jun 2020, including GBR) or "EU27" (data from Jun 2021)}
}
\value{
converted European Energy Datasheets magpie object
}
\description{
Convert European Energy Datasheets
}
\examples{
\dontrun{
test <- readSource("EuropeanEnergyDatasheets", subtype = "EU27", convert = TRUE)
}

}
\author{
Renato Rodrigues and Atreya Shankar
}
