% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanEnergyDatasheets.R
\name{convertEuropeanEnergyDatasheets}
\alias{convertEuropeanEnergyDatasheets}
\title{Convert European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/country
}
\usage{
convertEuropeanEnergyDatasheets(x, subtype)
}
\arguments{
\item{x}{European Energy Datasheets magpie object derived from readEuropeanEnergyDatasheets function}

\item{subtype}{as string. Possibilities are "Production","Mport","Xport","NetImports","Consumption","Transformation","EnergyBranchConsumption",
"DistributionLosses","PrimaryEnergy","AvailableFe","FinalEnergy","Electricity","Indicators","Heat", "CHP","Transport","Emissions","Population" and "GDP"}
}
\value{
converted European Energy Datasheets magpie object
}
\description{
Converts European Energy Datasheets magpie object into appropriate form for the REMIND model
}
\examples{
\dontrun{a <- convertEuropeanEnergyDatasheets(x, "Production")}
}
\author{
Atreya Shankar
}
