% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExpertGuess.R
\name{convertExpertGuess}
\alias{convertExpertGuess}
\title{convertExpertGuess}
\usage{
convertExpertGuess(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{Type of data that should be read. One of
\itemize{
\item \code{biocharPrices}: Biochar price assumptions over time. Assumptions based on
collection of current bulk sale prices (see Dorndorf et al (submitted)) (Tabea Dorndorf)
\item \code{capacityFactorGlobal}: Global capacity factors for all REMIND technologies (Renato Rodrigues)
\item \code{capacityFactorRules}: Capacity factor rules for selected H12 regions and REMIND technologies
\item \code{ccsBounds}: CCS bounds indicating the if a country is expected to do CCS in the
foreseeable future (Jessica Strefler)
\item \code{co2prices}: CO2 prices (Robert Pietzcker)
\item \code{costsTradePeFinancial}: primary energy tradecosts (financial costs on import,
export and use) (Nicolas Bauer)
\item \code{deltacapoffset}: ??? (Robert Pietzcker)
\item \code{gridFactor}: Estimates distribution of electricity demands per region (Robert Pietzcker)
\item \code{ies}: intertemporal elasticity of substitution (Nicolas Bauer)
\item \code{prtp}: pure rate of time preference (Nicolas Bauer)
\item \code{subConvergenceRollback}: Subsidy convergence level in rollback scenario in US$2017 (Nicolas Bauer)
\item \code{taxConvergence}: Tax convergence level in US$2017 (Nicolas Bauer)
\item \code{taxConvergenceRollback}: Tax convergence level in rollback scenario in US$2017 (Nicolas Bauer)
\item \code{tradeConstraints}: parameter by Nicolas Bauer (2024) for the region specific trade
constraints, values different to 1 activate constraints and the value is used as
effectiveness to varying degrees such as percentage numbers (Nicolas Bauer)
}}
}
\description{
Converts data from expert guess
}
