% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGlobalEnergyMonitor.R
\name{convertGlobalEnergyMonitor}
\alias{convertGlobalEnergyMonitor}
\title{Convert Global Energy Monitor data}
\usage{
convertGlobalEnergyMonitor(x)
}
\arguments{
\item{x}{A magclass object returned from readGlobalEnergyMonitor().}
}
\description{
Convert Global Energy Monitor data
}
\author{
Rahel Mandaroux, Falk Benke
}
