% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_PVPS.R
\name{convertIEA_PVPS}
\alias{convertIEA_PVPS}
\title{Convert IEA PVPS data from REMIND regions to iso countries}
\usage{
convertIEA_PVPS(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
Magpie object with IEA PVPS investment cost per country
}
\description{
maps to iso countries
}
\author{
Felix Schreyer
}
