% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_WorldEnergyOutlook.R
\name{convertIEA_WorldEnergyOutlook}
\alias{convertIEA_WorldEnergyOutlook}
\title{Convert IEA World Energy Outlook Data from 2023}
\usage{
convertIEA_WorldEnergyOutlook(x)
}
\arguments{
\item{x}{magclass object to be converted}
}
\description{
Convert IEA World Energy Outlook Data from 2023
}
\author{
Falk Benke
}
