% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNewClimate.R
\name{convertNewClimate}
\alias{convertNewClimate}
\title{Convert policy targets for NPIs from New Climate policy database}
\usage{
convertNewClimate(x, subtype, subset)
}
\arguments{
\item{x}{a magclass object to be converted}

\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets, Emissions_YYYY_cond or
Emissions_YYYY_uncond for Emissions targets, with YYYY NPI version year}

\item{subset}{String, designating the GDP scenarios to use. Only used for emission targets.}
}
\description{
Converts conditional and unconditional capacity and production targets into total capacity (GW) in target year.
For countries and years without targets, 2020 values from IRENA and BP are used to fill the gaps.
}
\details{
Emissions targets on absolute level for total GHG emissions without bunkers and land-use change emissions are calculated
from country-specific target formulation and land-use change emissions data
}
\seealso{
\code{\link[=readIRENA]{readIRENA()}}
}
\author{
Rahel Mandaroux, Léa Hayez, Falk Benke
}
