% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNewClimate.R
\name{convertNewClimate}
\alias{convertNewClimate}
\title{Convert policy targets for NPIs from New Climate policy database}
\usage{
convertNewClimate(x, subtype, subset)
}
\arguments{
\item{x}{a magclass object to be converted}

\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets, Emissions_YYYY_cond or
Emissions_YYYY_uncond for Emissions targets, with YYYY NPI version year}

\item{subset}{String, designating the GDP scenarios to use. Only used for emission targets.}
}
\description{
Converts conditional and unconditional capacity targets into total capacity (GW) in target year
the Generation targets are similar to the capacity targets but include the capacity factors,
the Emissions targets are the total (except land CO2) emissions in the target year

Emission targets are represented by a GHG factor, which is the quotient of total GHG
emissions in the target year divided by the CEDS GHG emissions in 2005.
}
\author{
Rahel Mandaroux, Léa Hayez, Falk Benke
}
