% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertREMIND_11Regi.R
\name{convertREMIND_11Regi}
\alias{convertREMIND_11Regi}
\title{Converts REMIND regional data}
\usage{
convertREMIND_11Regi(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Name of the regional data, e.g. tradecost", "pe2se",
"deltacapoffset", capacityFactorRules", "taxConvergence", "maxFeSubsidy",
"maxPeSubsidy", "propFeSubsidy", "fossilExtractionCoeff", "uraniumExtractionCoeff"}
}
\value{
A MAgPIE object containing country disaggregated data
}
\description{
Converts REMIND regional data
}
\examples{

\dontrun{ a <- convertREMIND_11Regi(x,subtype="capacityFactorGlobal")
}

}
\author{
original: not defined - capacity factor, tax, fossil and RLDC changes: Renato Rodrigues
}
