% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertREMIND_11Regi.R
\name{convertREMIND_11Regi}
\alias{convertREMIND_11Regi}
\title{Converts REMIND 11 Regi data}
\usage{
convertREMIND_11Regi(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Name of the source, e.g. "fossilExtractionCoeff", "uraniumExtractionCoeff"}
}
\description{
Converts REMIND 11 Regi data
}
\author{
unknown
}
