% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStationary.R
\name{convertStationary}
\alias{convertStationary}
\title{Convert Stationary data to data on ISO country level.}
\usage{
convertStationary(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Convert Stationary data to data on ISO country level.
}
\author{
Antoine Levesque, Robin Hasse
}
