% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNFCCC_NDC.R
\name{convertUNFCCC_NDC}
\alias{convertUNFCCC_NDC}
\title{Policy targets for NDCs from UNFCCC_NDC}
\usage{
convertUNFCCC_NDC(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets, Emissions_YYYY_cond or
Emissions_YYYY_uncond for Emissions targets, with YYYY NDC version year}
}
\value{
Magpie object with Total Installed Capacity (GW) targets, target years differ depending upon the database.
}
\description{
Converts conditional and unconditional capacity targets into total capacity (GW) in target year
the Generation targets are similar to the capacity targets but include the capacity factors,
the Emissions targets are the total (except land CO2) emissions in the target year
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters
}
