% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSoilGrids.R
\name{downloadSoilGrids}
\alias{downloadSoilGrids}
\title{downloadSoilGrids}
\usage{
downloadSoilGrids()
}
\description{
download SoilGrids data from https://files.isric.org/soilgrids/data/recent/OCSTHA_M_30cm_250m_ll.tif
}
\examples{
\dontrun{ downloadSource("SoilGrids") }

}
\seealso{
\code{\link{readSoilGrids}}
}
\author{
Kristine Karstens
}
