% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullREMIND.R
\name{fullREMIND}
\alias{fullREMIND}
\title{fullREMIND}
\usage{
fullREMIND()
}
\description{
Function that produces the complete regional data set required for the REMIND model.
}
\examples{
\dontrun{
fullREMIND()
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}}, \code{\link[madrat]{calcOutput}}
}
\author{
Lavinia Baumstark
}
