% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTHRESHOLDS.R
\name{fullTHRESHOLDS}
\alias{fullTHRESHOLDS}
\title{export validation thresholds}
\usage{
fullTHRESHOLDS(type = "config")
}
\arguments{
\item{type}{choose either "config" to export thresholds as used in the
validationConfig or "full" to export all pipeline data}
}
\description{
assemble near-term thresholds from project pipelines and potentially other
data sources and export them to a file
}
\author{
Pascal Weigmann
}
