% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullVALIDATIONREMIND.R
\name{fullVALIDATIONREMIND}
\alias{fullVALIDATIONREMIND}
\title{Generate Validation Data for REMIND}
\usage{
fullVALIDATIONREMIND(rev = 0)
}
\arguments{
\item{rev}{Unused parameter here for the pleasure of \code{\link{madrat}}.}
}
\description{
Function that generates the historical regional dataset against which the
REMIND model results can be compared.
}
\examples{
\dontrun{
fullVALIDATIONREMIND()
}
}
\seealso{
\code{\link[=fullREMIND]{fullREMIND()}}, \code{\link[=readSource]{readSource()}}, \code{\link[=getCalculations]{getCalculations()}},
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
David Klein, Falk Benke
}
