% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAGEB.R
\name{readAGEB}
\alias{readAGEB}
\title{Read AGEB}
\usage{
readAGEB(subtype = "balances")
}
\arguments{
\item{subtype}{data subtype. Either "balances" ("Auswertungstabellen zur Energiebilanz Deutschland")
or "electricity" ("Bruttostromerzeugung in Deutschland nach Energietraegern")}
}
\description{
Read AGEB
}
\author{
Falk Benke
}
