% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAriadneDB.R
\name{readAriadneDB}
\alias{readAriadneDB}
\title{Ariadne database scenario data}
\usage{
readAriadneDB()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Scenario data from the Ariadne modeling intercomparison project for Germany.
See README in input file for more details.
}
\author{
Felix Schreyer
}
