% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBP.R
\name{readBP}
\alias{readBP}
\title{BP Capacity and Generation Data}
\usage{
readBP(subtype)
}
\arguments{
\item{subtype}{Either "Emission", Capacity", "Generation", "Production", "Consumption", "Trade Oil", "Trade Gas", "Trade Coal" or "Price"}
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
BP data. See README in input file for more details.
}
\author{
Aman Malik, Falk Benke
}
