% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBP.R
\name{readBP}
\alias{readBP}
\title{BP Capacity and Generation Data}
\usage{
readBP(subtype)
}
\arguments{
\item{subtype}{Either "Capacity" or "Generation"}
}
\value{
A magpie object
}
\description{
BP data. See README in input file for more details.
}
\details{
Data contains historical electricity renewable capacities (in MW for Wind, Solar, and Geothermal) and Generation (in TWh for Nuclear, Hydro, Wind, Solar, Other Renewables, and Geo Biomass)
}
\author{
Aman Malik
}
