% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClimateTrace.R
\name{readClimateTrace}
\alias{readClimateTrace}
\title{Read Climate Trace}
\usage{
readClimateTrace()
}
\value{
magpie object of the ClimateTrace data with historical emissions
by subsector and country
}
\description{
Read-in an Climate Trace csv files as magclass object for CO2, CH4 and N2O
}
\examples{
\dontrun{
a <- readSource(type = "ClimateTrace")
}

}
\author{
Pascal Weigmann
}
