% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGE.R
\name{readEDGE}
\alias{readEDGE}
\title{Read EDGE}
\usage{
readEDGE(
  subtype = c("FE_stationary", "FE_buildings", "Capital", "CapitalUnit", "Floorspace",
    "ES_buildings")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Load an EDGE Buildings file as magclass object.
}
\examples{
\dontrun{
a <- readSource("EDGE")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Antoine Levesque, Robin Krekeler
}
