% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGE_Industry.R
\name{readEDGE_Industry}
\alias{readEDGE_Industry}
\title{Read EDGE Industry Data}
\usage{
readEDGE_Industry(subtype)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{projections_VA_iso3c} for the projections of Value Added from the
industry sector
\item \code{steel_production_scenarios} for the projections of primary and
secondary steel production
\item \code{secondary_steel_limits} for the upper limits to secondary steel
production
\item \code{cement_chemicals_otherInd_production_scenarios} for the projections of
cement, chemicals, and other industry production
\item \code{p29_capitalQuantity_industry} for industry subsector energy efficiency
capital trajectories
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read EDGE Industry data
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Michaja Pehl
}
