% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEdgeBuildings.R
\name{readEdgeBuildings}
\alias{readEdgeBuildings}
\title{Load an EDGE Buildings file as magclass object.}
\usage{
readEdgeBuildings(subtype = c("FE", "Floorspace"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Load an EDGE Buildings file as magclass object.
}
\author{
Antoine Levesque, Robin Hasse
}
