% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGETransport.R
\name{readEDGETransport}
\alias{readEDGETransport}
\title{Read EDGETransport inputs}
\usage{
readEDGETransport(subtype = "logit_exponent")
}
\arguments{
\item{subtype}{logit_exponents, SW, pref, value_time, harmonized_intensities, price_nonmot, UCD_NEC_iso, loadFactor, esCapCost, fe_demand_tech, shares_LDV_transport, demISO, annual_mileage}
}
\value{
magpie object of EDGEtransport iterative inputs
}
\description{
Read-in EDGETransport inputs csv file as magclass object
}
\examples{
\dontrun{ a <- readSource(type="EDGETransport")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Marianna Rottoli, Alois Dirnaichner
}
