% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEU_ReferenceScenario.R
\name{readEU_ReferenceScenario}
\alias{readEU_ReferenceScenario}
\title{Read EU Reference Scenario}
\usage{
readEU_ReferenceScenario(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "techAssump.*", "2016" or "2020"}
}
\value{
magpie object of EU reference scenario data by country. Units follow
  REMIND report conventions and conversion factor is defined in
  EUReferenceScenario2REMIND.xlsx file.
}
\description{
Read EU Reference Scenario .xlsx file as magpie object
}
\examples{
\dontrun{
test <- readSource("EU_ReferenceScenario", subtype = "2020", convert = FALSE)
}
}
\author{
Renato Rodrigues, Falk Benke, Robin Hasse
}
