% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanEnergyDatasheets.R
\name{readEuropeanEnergyDatasheets}
\alias{readEuropeanEnergyDatasheets}
\title{Read European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/energy-statistical-pocketbook
}
\usage{
readEuropeanEnergyDatasheets(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "EU28" (older data from Jun 2020, including GBR) or "EU27" (data from Jun 2021)}
}
\value{
magpie object of European Energy Datasheets
}
\description{
Read European Energy Datasheets .xlsx file as magpie object.
}
\examples{
\dontrun{
test <- readSource("EuropeanEnergyDatasheet", subtype = "EU27", convert = FALSE)
}
}
\author{
Renato Rodrigues, Atreya Shankar, Falk Benke
}
