% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanEnergyDatasheets.R
\name{readEuropeanEnergyDatasheets}
\alias{readEuropeanEnergyDatasheets}
\title{Read European Energy Datasheets}
\source{
European Energy Datasheets public database https://ec.europa.eu/energy/en/data-analysis/country
}
\usage{
readEuropeanEnergyDatasheets()
}
\value{
magpie object of aggregated energy market data by country. Units in Mtoe, unless otherwise specified.
}
\description{
Read European Energy Datasheets .xlsx file as magpie object
}
\examples{
\dontrun{a <- readEuropeanEnergyDatasheets()}
}
\author{
Atreya Shankar
}
