% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanEnergyDatasheets.R
\name{readEuropeanEnergyDatasheets}
\alias{readEuropeanEnergyDatasheets}
\title{Read European Energy Datasheets}
\source{
European Energy Datasheets public database
https://energy.ec.europa.eu/data-and-analysis/eu-energy-statistical-pocketbook-and-country-datasheets_en
}
\usage{
readEuropeanEnergyDatasheets(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "EU28" (data from June 20 including GBR)
or "EU27" (latest data from August 23 without GBR)}
}
\value{
magpie object of European Energy Datasheets
}
\description{
Read European Energy Datasheets .xlsx file as magpie object.
}
\examples{
\dontrun{
test <- readSource("EuropeanEnergyDatasheet", subtype = "EU27", convert = FALSE)
}
}
\author{
Renato Rodrigues, Atreya Shankar, Falk Benke
}
