% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpertGuess.R
\name{readExpertGuess}
\alias{readExpertGuess}
\title{Read ExpertGuess}
\usage{
readExpertGuess(subtype)
}
\arguments{
\item{subtype}{Type of data that should be read.  One of
\itemize{
\item \code{Chinese_Steel_Production}: "Smooth" production estimates by Robert
Pietzcker (2022).
\item Any of the others nobody cared to document.
}}
}
\value{
magpie object of the data
}
\description{
Read-in data that are based on expert guess
}
\examples{

\dontrun{ a <- readSource(type="ExpertGuess",subtype="ies")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
