% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpertGuess.R
\name{readExpertGuess}
\alias{readExpertGuess}
\title{Read ExpertGuess}
\usage{
readExpertGuess(subtype)
}
\arguments{
\item{subtype}{Type of data that should be read.  One of
\itemize{
\item \code{biocharPrices}: Biochar price assumptions over time. Assumptions
based on collection of current bulk sale prices (see Dorndorf et al (submitted))
\item \code{capacityFactorGlobal}: Global capacity factors for all REMIND technologies
\item \code{capacityFactorRules}: Capacity factor rules for selected H12 regions and REMIND technologies
\item \code{costsTradePeFinancial}
\item \code{gridFactor}: ..
\item \code{storageFactor}: Regional storage parametrization
\item \code{CCSbounds}
\item \code{Steel_Production}: Steel production estimates
\item \code{industry_max_secondary_steel_share}: Maximum share of secondary steel
production in total steel production and years between which a linear
convergence from historic to target shares is to be applied.
\item \code{cement_production_convergence_parameters}: convergence year and level
(relative to global average) to which per-capita cement demand converges
\item \code{ies}
\item \code{prtp}
\item \code{tradeConstraints}: parameter by Nicolas Bauer (2024) for the region
specific trade constraints, values different to 1 activate constraints
and the value is used as effectiveness to varying degrees such as percentage numbers
}}
}
\value{
magpie object of the data
}
\description{
Read-in data that are based on expert guess
}
\examples{
\dontrun{
a <- readSource(type = "ExpertGuess", subtype = "ies")
}

}
\author{
Lavinia Baumstark, Falk Benke
}
