% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalEnergyMonitor.R
\name{readGlobalEnergyMonitor}
\alias{readGlobalEnergyMonitor}
\title{Read Global Energy Monitor data}
\usage{
readGlobalEnergyMonitor()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Read Global Energy Monitor data
}
\seealso{
[`readSource()`]
}
\author{
Falk Benke
}
