% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHRE.R
\name{readHRE}
\alias{readHRE}
\title{Read Heat Roadmap Europe data}
\source{
https://heatroadmap.eu/roadmaps/
}
\usage{
readHRE()
}
\value{
A [`magpie`][magclass::magclass] object.
}
\description{
Read Heat Roadmap Europe data
}
\seealso{
[`readSource()`]
}
\author{
Pascal Weigmann
}
