% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_CCUS.R
\name{readIEA_CCUS}
\alias{readIEA_CCUS}
\title{Read IEA CCUS data}
\usage{
readIEA_CCUS(subtype)
}
\arguments{
\item{subtype}{either `historical` for data from until 2022 or `projections`
for projections in 2020, 2025 and 2030}
}
\description{
Reads in capacities from projects in IEA CCUS database
}
\author{
Anne Merfort, Falk Benke
}
