% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_ETP.R
\name{readIEA_ETP}
\alias{readIEA_ETP}
\title{Read IEA ETP projections}
\usage{
readIEA_ETP(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "main" or "industry_subsectors"}
}
\description{
Read IEA ETP projections
}
\author{
Falk Benke
}
