% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_ETP.R
\name{readIEA_ETP}
\alias{readIEA_ETP}
\title{Read IEA ETP projections}
\usage{
readIEA_ETP(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "industry", "buildings", "summary", or "transport"}

\item{useCorrected}{boolean indicating if the corrected version of the input sheet should be used}
}
\description{
Read IEA ETP projections
}
\author{
Falk Benke
}
