% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WorldEnergyOutlook.R
\name{readIEA_WorldEnergyOutlook}
\alias{readIEA_WorldEnergyOutlook}
\title{Read in IEA World Energy Outlook Data from 2023}
\usage{
readIEA_WorldEnergyOutlook()
}
\description{
Read in IEA World Energy Outlook Data from 2023
}
\author{
Falk Benke
}
