% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIRENA.R
\name{readIRENA}
\alias{readIRENA}
\title{Read IRENA}
\usage{
readIRENA(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "Capacity" or "Generation"}
}
\value{
magpie object of the IRENA data with historical electricity renewable
capacities (MW) or generation levels (GWh)
}
\description{
Read-in an IRENA xlsx file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "IRENA", subtype = "Capacity")
}

}
\author{
Renato Rodrigues, Pascal Weigmann
}
