% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD.R
\name{readOECD}
\alias{readOECD}
\alias{convertOECD}
\title{Read OECD}
\usage{
readOECD()

convertOECD(x)
}
\arguments{
\item{x}{MAgPIE object returned from readOECD}
}
\description{
Read-in risk premium
}
\examples{
\dontrun{
readSource("OECD")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
