% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD.R
\name{readOECD}
\alias{readOECD}
\title{Read OECD GDP / ratio PPP2MER}
\usage{
readOECD(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "gdp" or "ratioPM"}
}
\value{
magpie object of the GDP data
}
\description{
Read-in GDP or ratio PPP2MER data xlsx file from OECD as magclass object
}
\examples{

\dontrun{ 
a <- readSource("OECD","gdp")
a <- readSource("OECD","ratioPM")

}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
