% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPWT.R
\name{readPWT}
\alias{readPWT}
\alias{convertPWT}
\title{Read PWT}
\usage{
readPWT()

convertPWT(x)
}
\arguments{
\item{x}{MAgPIE object returned by readPWT}
}
\value{
Magpie object of the PWT data
}
\description{
Read and convert PWT data
}
\section{Functions}{
\itemize{
\item \code{convertPWT()}: Convert PWT data

}}
\concept{"Past" GDPpc functions}
