% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPWT.R
\name{readPWT}
\alias{readPWT}
\title{Read PWT}
\usage{
readPWT()
}
\value{
magpie object of the PWT data
}
\description{
Read-in an PWT data .xlsx file as magclass object
}
\examples{

\dontrun{ a <- readSource(type="PWT")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Lavinia Baumstark
}
