% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND_11Regi.R
\name{readREMIND_11Regi}
\alias{readREMIND_11Regi}
\title{Read REMIND region dependent data}
\usage{
readREMIND_11Regi(subtype)
}
\arguments{
\item{subtype}{Name of the regional data, e.g. "p4", "biomass", "ch4waste",
"tradecost", "pe2se", "xpres_tax", "deltacapoffset", "capacityFactorGlobal",
"capacityFactorRules", "residuesShare", "taxConvergence", "maxFeSubsidy",
"maxPeSubsidy", "propFeSubsidy", "fossilExtractionCoeff", "uraniumExtractionCoeff",
"RLDCCoefficientsLoB", "RLDCCoefficientsPeak", "earlyRetirementAdjFactor"}
}
\value{
magpie object of region dependent data
}
\description{
Read-in an csv files that contains regional data
}
\examples{
\dontrun{
a <- readSource(type = "REMIND_11Regi", subtype = "capacityFactorGlobal")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
original: not defined, capacity factor, tax, fossil and RLDC changes: Renato Rodrigues
}
