% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND_11Regi.R
\name{readREMIND_11Regi}
\alias{readREMIND_11Regi}
\title{Read REMIND region dependent data}
\usage{
readREMIND_11Regi(subtype)
}
\arguments{
\item{subtype}{Name of the regional data, e.g.
"tradecost", "deltacapoffset", "taxConvergence", "maxFeSubsidy",
"maxPeSubsidy", "propFeSubsidy", "fossilExtractionCoeff", "uraniumExtractionCoeff"}
}
\value{
magpie object of region dependent data
}
\description{
Read-in an csv files that contains regional data
}
\examples{
\dontrun{
a <- readSource(type = "REMIND_11Regi", subtype = "tradecost")
}
}
\author{
original: not defined, tax, fossil and RLDC changes: Renato Rodrigues
}
