% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRutovitz2015.R
\name{readRutovitz2015}
\alias{readRutovitz2015}
\title{Employment factors for various power production technologies from Rutovitz et al. 2015}
\usage{
readRutovitz2015(subtype)
}
\arguments{
\item{subtype}{Either "oecd_ef","regional_ef","coal_ef","gas_ef", "regional_mult"}
}
\value{
magpie object of employment factors for different technologies and activities in Jobs/MW (all except fuel_supply) or Jobs/PJ (fuel_supply). Subtype "regional_mult" is a regional multiplier without units.
}
\description{
Rutovitz, J., Dominish, E., & Downes, J. (2015). Calculating global energy sector jobs—2015 methodology update. Institute for Sustainable Futures, University of Technology, Sydney. https://opus.lib.uts.edu.au/bitstream/10453/43718/1/Rutovitzetal2015Calculatingglobalenergysectorjobsmethodology.pdf
}
\author{
Aman Malik
}
