% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSoilGrids.R
\name{readSoilGrids}
\alias{readSoilGrids}
\title{readSoilGrids}
\usage{
readSoilGrids()
}
\value{
magpie object in cellular resolution
}
\description{
This function reads the raw SoilGrids data (available at https://files.isric.org/soilgrids/data/recent/OCSTHA_M_30cm_250m_ll.tif) 
or if available the preprocessed raster layers.
}
\examples{
\dontrun{ readSource("SoilGrids") }

}
\seealso{
\code{\link{downloadSoilGrids}}
}
\author{
Kristine Karstens
}
