% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStationary.R
\name{readStationary}
\alias{readStationary}
\title{Load Stationary File as magclass object}
\usage{
readStationary()
}
\value{
magclass object
}
\description{
Load Stationary File as magclass object
}
\author{
Antoine Levesque, Robin Hasse
}
