% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC.R
\name{readUNFCCC}
\alias{readUNFCCC}
\title{Read UNFCCC data}
\usage{
readUNFCCC()
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read UNFCCC data
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Falk Benke
}
