% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC_NDC.R
\name{readUNFCCC_NDC}
\alias{readUNFCCC_NDC}
\title{Reads NDC policy database with capacity, emission, and share targets, originally based on Rogelj et al. 2017}
\usage{
readUNFCCC_NDC(subtype, subset)
}
\arguments{
\item{subtype}{Capacity_YYYY_cond or Capacity_YYYY_uncond for Capacity Targets, Emissions_YYYY_cond or
Emissions_YYYY_uncond for Emissions targets, with YYYY NDC version year,
determines the database version to be read in}

\item{subset}{A string (or vector of strings) designating the scenario(s) to be returned.}
}
\description{
Reads excel sheet with NDC (Nationally Determined Contributions)
data on different policy targets (capacity, emission, and share targets) with different variations.
}
\author{
Aman Malik, Christoph Bertram, Oliver Richters, Sophie Fuchs, Rahel Mandaroux
}
