% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateTimeSteps.R
\name{toolAggregateTimeSteps}
\alias{toolAggregateTimeSteps}
\title{Aggregate values to n-year averages to suppress volatility}
\usage{
toolAggregateTimeSteps(x, nYears = 5)
}
\arguments{
\item{x}{a magclass object}

\item{nYears}{time steps to be used for averaging, defaults to 5}
}
\value{
magclass object with averages
}
\description{
Aggregate values to n-year averages to suppress volatility
}
\author{
Robin Hasse
}
