% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCalcEnergyProj.R
\name{toolCalcEnergyProj}
\alias{toolCalcEnergyProj}
\title{Calculate energy projections on country-level based on EDGE models outputs per country.
These energy projections are used in the input data preparation for aggregating country-specific data to REMIND regions.
They are a country-level proxy of the final energy demand trajectories on the level of REMIND regions provided by the EDGE models.}
\usage{
toolCalcEnergyProj(
  subtype,
  gdpScen,
  demScen,
  scenario,
  years = seq(2020, 2050, 5)
)
}
\arguments{
\item{subtype}{"FE" (Total final energy consumption), "SE|Electricity" (SE electricity generation)}

\item{gdpScen}{GDP scenario to use}

\item{demScen}{demand scenario to use}

\item{scenario}{set of GDP scenarios to use for calcFEDemand calculation (trigger standard cache in this function)}

\item{years}{target years for projection}
}
\description{
Calculate energy projections on country-level based on EDGE models outputs per country.
These energy projections are used in the input data preparation for aggregating country-specific data to REMIND regions.
They are a country-level proxy of the final energy demand trajectories on the level of REMIND regions provided by the EDGE models.
}
\seealso{
\code{\link[=convertNewClimate]{convertNewClimate()}}
}
\author{
Felix Schreyer
}
