% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSolarFunctionAggregate.R
\name{toolSolarFunctionAggregate}
\alias{toolSolarFunctionAggregate}
\title{toolSolarFunctionAggregate}
\usage{
toolSolarFunctionAggregate(x, rel = NULL)
}
\arguments{
\item{x}{magclass object that should be aggregated}

\item{rel}{relation matrix containing a region mapping.
A mapping object should contain 2 columns in which each element of x
is mapped to the category it should belong to after (dis-)aggregation}
}
\value{
return: returns region aggregated solar data
}
\description{
Aggregate Solar data into regions
}
\author{
Felix Schreyer, Renato Rodrigues, Julian Oeser
}
